% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_rhos_par_tab}
\alias{add_rhos_par_tab}
\title{Add measurement offset values to par_tab}
\usage{
add_rhos_par_tab(par_tab, sampled_viruses, n_obs_types = 1)
}
\arguments{
\item{par_tab}{the parameter table to add the measurement offsets to}

\item{sampled_viruses}{the vector of measured biomarker_ids to add offset terms for}

\item{n_obs_types}{number of biomarker_groups}
}
\value{
a list containing the updated parameter table and a table mapping each biomarker ID and biomarker group to a measurement-shift parameter index
}
\description{
Add measurement offset values to par_tab
}
