% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.R
\name{add_scale_pars}
\alias{add_scale_pars}
\title{Add scaling parameters to par_tab}
\usage{
add_scale_pars(
  par_tab,
  antibody_data,
  timevarying_demographics = NULL,
  scale_par_lower = -25,
  scale_par_upper = 25
)
}
\arguments{
\item{par_tab}{the parameter table, including a column called stratification which is NA if no stratification is requested, or a character value matching an entry in antibody_data or timevarying_demographics}

\item{antibody_data}{the antibody data data frame, see \code{\link{example_antibody_data}}. If NULL, then uses timevarying_demographics}

\item{timevarying_demographics}{a data frame of timevarying demographics, with columns individual, time and any stratification variables. If NULL, then uses antibody_data}

\item{scale_par_lower}{the lower bound of any used scale parameters}

\item{scale_par_upper}{the upper bound of any used scale parameters}
}
\value{
the updated `par_tab`, including rows for any added scale parameters
}
\description{
Adds entries to par_tab to stratify parameters by requested stratification levels. antibody_data and timevarying_demographics are used to find how many levels of each stratification have been requested
}
\seealso{
Other stratification:
\code{\link[=create_demographic_table]{create_demographic_table()}},
\code{\link[=get_demographic_groups]{get_demographic_groups()}},
\code{\link[=setup_stratification_table]{setup_stratification_table()}}
}
\concept{stratification}
