% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.R
\name{align_antibody_demographic_dat}
\alias{align_antibody_demographic_dat}
\title{Merge antibody data and demographics}
\usage{
align_antibody_demographic_dat(
  antibody_data,
  demographics = NULL,
  verbose = FALSE
)
}
\arguments{
\item{antibody_data}{antibody data frame}

\item{demographics}{optional fixed or time-varying demographic data}

\item{verbose}{if TRUE, prints warnings about overlapping columns}
}
\value{
the antibody data with demographic variables added
}
\description{
Internal helper that merges demographic variables into `antibody_data`, using `sample_time` to align time-varying demographics.
}
