% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenmap.R
\name{antigenmap_lengthen}
\alias{antigenmap_lengthen}
\title{Lengthen antigenic map}
\usage{
antigenmap_lengthen(antigenmap, new_epochs_per_year)
}
\arguments{
\item{antigenmap}{A dataframe of antigenic coordinates, with columns
\code{strain}, \code{x}, \code{y}, \code{epoch} and \code{epochs_per_year}.}

\item{new_epochs_per_year}{New epochs per year. Must be greater than
\code{epochs_per_year} in \code{antigenmap}}
}
\value{
The same map but longer
}
\description{
Takes an antigenic map with a certain number of epochs per year and lengthen
it so that there are more epochs per year
}
