% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{bb_mean}
\alias{bb_mean}
\title{Beta binomial mean}
\usage{
bb_mean(n, alpha, beta)
}
\arguments{
\item{n}{the number of trials}

\item{alpha}{alpha parameter}

\item{beta}{beta parameter}
}
\value{
calculates the mean of the beta binomial function
}
\description{
Beta binomial mean
}
