% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{bb_var}
\alias{bb_var}
\title{Beta binomial variance}
\usage{
bb_var(n, alpha, beta)
}
\arguments{
\item{n}{the number of trials}

\item{alpha}{alpha parameter}

\item{beta}{beta parameter}
}
\value{
calculates the variance of the beta binomial function
}
\description{
Beta binomial variance
}
\examples{
bb_var(100, 1, 1)
}
