% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{calc_a}
\alias{calc_a}
\title{Calculate the first Beta shape parameter from a mode and certainty value}
\usage{
calc_a(mode1, k)
}
\arguments{
\item{mode1}{desired mode}

\item{k}{certainty value}
}
\value{
the first Beta shape parameter
}
\description{
Calculate the first Beta shape parameter from a mode and certainty value
}
