% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{calc_lambda_probs}
\alias{calc_lambda_probs}
\title{Calculate FOI log probability}
\usage{
calc_lambda_probs(lambdas, infHist, ageMask)
}
\arguments{
\item{lambdas}{a vector of FOIs}

\item{infHist}{the matrix of infection histories}

\item{ageMask}{the age mask vector as returned by \code{\link{create_age_mask}}}
}
\value{
a single log probability
}
\description{
Given a vector of FOIs for all circulating years, a matrix of infection histories and the vector specifying if individuals were alive or not, returns the log probability of the FOIs given the infection histories.
}

