% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{calc_lambda_probs_indiv_brute}
\alias{calc_lambda_probs_indiv_brute}
\title{FOR DEBUGGING}
\usage{
calc_lambda_probs_indiv_brute(lambdas, infHist, ageMask)
}
\description{
Brute force implementation of calculating the explicit FOI
}

