% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{calc_lambda_probs_monthly}
\alias{calc_lambda_probs_monthly}
\title{Calculate FOI from spline}
\usage{
calc_lambda_probs_monthly(foi, knots, theta, infection_history, age_mask)
}
\arguments{
\item{foi}{vector of force of infections per year}

\item{knots}{vector of knots for the spline}

\item{theta}{vector of theta parameters for spline}

\item{infection_history}{the matrix of infection histories}

\item{age_mask}{the age mask vector as returned by \code{\link{create_age_mask}}}
}
\value{
a vector of log probabilities for each individual
}
\description{
Version of FOI prior that calculates a seasonal spline such that the FOI in a given time point (month) comes from this spline term
}
