% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_model_likelihoods.R
\name{calc_phi_loc_probs_indiv}
\alias{calc_phi_loc_probs_indiv}
\title{Calculate grouped FOI log probabilities by individual}
\usage{
calc_phi_loc_probs_indiv(
  phis,
  group_probs,
  infection_history,
  age_mask,
  sample_mask,
  group_indices
)
}
\arguments{
\item{phis}{vector of force-of-infection values}

\item{group_probs}{vector of relative probabilities for the population groups}

\item{infection_history}{matrix of infection histories}

\item{age_mask}{vector giving the first possible infection-time index for each individual}

\item{sample_mask}{vector giving the last possible infection-time index for each individual}

\item{group_indices}{population-group index for each individual}
}
\value{
a vector of log probabilities, one for each individual
}
\description{
Internal helper that applies group-specific infection probabilities to each individual's infection history.
}
\keyword{internal}
