% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_model_likelihoods.R
\name{calc_phi_probs}
\alias{calc_phi_probs}
\title{Calculate FOI log probability}
\usage{
calc_phi_probs(phis, infection_history, age_mask, sample_mask)
}
\arguments{
\item{phis}{a vector of FOIs}

\item{infection_history}{the matrix of infection histories}

\item{age_mask}{the age mask vector as returned by \code{\link{create_age_mask}}}

\item{sample_mask}{the sample mask vector giving the last possible infection-time index for each individual}
}
\value{
a single log probability for the infection histories
}
\description{
Given a vector of FOIs for all circulating years, a matrix of infection histories and the vector specifying if individuals were alive or not, returns the log probability of the FOIs given the infection histories.
}
\seealso{
Other priors:
\code{\link[=calc_phi_probs_indiv]{calc_phi_probs_indiv()}},
\code{\link[=calc_phi_probs_spline]{calc_phi_probs_spline()}},
\code{\link[=create_prob_shifts]{create_prob_shifts()}},
\code{\link[=dbb_prior]{dbb_prior()}},
\code{\link[=density_beta_binom]{density_beta_binom()}},
\code{\link[=find_beta_prior_mode]{find_beta_prior_mode()}},
\code{\link[=find_beta_prior_with_mean]{find_beta_prior_with_mean()}},
\code{\link[=find_beta_prior_with_mean_var]{find_beta_prior_with_mean_var()}},
\code{\link[=fit_beta_prior]{fit_beta_prior()}},
\code{\link[=fit_normal_prior]{fit_normal_prior()}},
\code{\link[=inf_mat_prior]{inf_mat_prior()}},
\code{\link[=infection_history_prior]{infection_history_prior()}},
\code{\link[=prob_shifts]{prob_shifts()}}
}
\concept{priors}
