% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{calc_phi_probs_indiv_brute}
\alias{calc_phi_probs_indiv_brute}
\title{FOR DEBUGGING}
\usage{
calc_phi_probs_indiv_brute(phis, infection_history, age_mask)
}
\description{
Brute force implementation of calculating the explicit FOI
}
