% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{calc_phi_probs_indiv_titre}
\alias{calc_phi_probs_indiv_titre}
\title{Calculate FOI log probability vector}
\usage{
calc_phi_probs_indiv_titre(phis, titres, infection_history, age_mask,
  strain_mask, alpha1, beta1)
}
\arguments{
\item{phis}{a vector of FOIs}

\item{infection_history}{the matrix of infection histories}

\item{age_mask}{the age mask vector as returned by \code{\link{create_age_mask}}}
}
\value{
a vector of log probabilities for each individual
}
\description{
Given a vector of FOIs for all circulating years, a matrix of infection histories and the vector specifying if individuals were alive or not, returns the log probability of the FOIs given the infection histories.
}
