% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_phi_probs_indiv_titre_cpp}
\alias{calc_phi_probs_indiv_titre_cpp}
\title{Calc titre probs}
\usage{
calc_phi_probs_indiv_titre_cpp(phis, titres, infection_history, age_mask,
  strain_mask, alpha1, beta1)
}
\description{
Calc titre probs
}
