% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_model_likelihoods.R
\name{calc_phi_probs_spline}
\alias{calc_phi_probs_spline}
\title{Calculate FOI from spline - INACTIVE}
\usage{
calc_phi_probs_spline(foi, knots, theta, infection_history, age_mask)
}
\arguments{
\item{foi}{vector of force of infections per year}

\item{knots}{vector of knots for the spline}

\item{theta}{vector of theta parameters for spline}

\item{infection_history}{the matrix of infection histories}

\item{age_mask}{the age mask vector as returned by \code{\link{create_age_mask}}}
}
\value{
a vector of log probabilities for each individual
}
\description{
Version of FOI prior that calculates a seasonal spline such that the FOI in a given time point (month) comes from this spline term
}
\seealso{
Other priors:
\code{\link[=calc_phi_probs]{calc_phi_probs()}},
\code{\link[=calc_phi_probs_indiv]{calc_phi_probs_indiv()}},
\code{\link[=create_prob_shifts]{create_prob_shifts()}},
\code{\link[=dbb_prior]{dbb_prior()}},
\code{\link[=density_beta_binom]{density_beta_binom()}},
\code{\link[=find_beta_prior_mode]{find_beta_prior_mode()}},
\code{\link[=find_beta_prior_with_mean]{find_beta_prior_with_mean()}},
\code{\link[=find_beta_prior_with_mean_var]{find_beta_prior_with_mean_var()}},
\code{\link[=fit_beta_prior]{fit_beta_prior()}},
\code{\link[=fit_normal_prior]{fit_normal_prior()}},
\code{\link[=inf_mat_prior]{inf_mat_prior()}},
\code{\link[=infection_history_prior]{infection_history_prior()}},
\code{\link[=prob_shifts]{prob_shifts()}}
}
\concept{priors}
\keyword{internal}
