% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_aux.R
\name{calibrate_temperatures}
\alias{calibrate_temperatures}
\title{calibrate temperatures for parallel chains}
\usage{
calibrate_temperatures(temperatures, swap_ratio)
}
\arguments{
\item{temperatures}{vector of length n: current temperatures of chains}

\item{swap_ratio}{vector of length n - 1: (proportion of accepted swaps
out of proposed swaps)/2
the factor of 2 arises because of the way the swaps are recorded, and how
we alternate between swapping 1<->2, 3<->4.... and 2<->3, 4<->5...}
}
\value{
vector of length n: new temperatures of chains
}
\description{
calibrate temperatures for parallel chains
}

