% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_attack_rates}
\alias{check_attack_rates}
\title{Checks the attack_rates supplied in simulate_data}
\usage{
check_attack_rates(attack_rates, possible_exposure_times)
}
\arguments{
\item{attack_rates}{a vector of attack_rates to be used in the simulation}

\item{possible_exposure_times}{vector of strain circulation times}
}
\value{
nothing at the moment
}
\description{
Checks the attack_rates supplied in simulate_data
}
\examples{
attack_rates <- runif(40)
possible_exposure_times <- seq_len(40)
check_attack_rates(attack_rates, possible_exposure_times)
}
\seealso{
Other check_inputs:
\code{\link[=check_data]{check_data()}},
\code{\link[=check_demographics]{check_demographics()}},
\code{\link[=check_inf_hist]{check_inf_hist()}},
\code{\link[=check_par_tab]{check_par_tab()}},
\code{\link[=check_proposals]{check_proposals()}}
}
\concept{check_inputs}
