% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_data}
\alias{check_data}
\title{Checks the entries of data used in serosolver}
\usage{
check_data(data, verbose = FALSE)
}
\arguments{
\item{data}{the data frame of data to be fitted. It must have columns: individual (integer ID of individual); sample_time (numeric time of sample taken); biomarker_id (numeric ID of the biomarker); measurement (antibody measurement); birth (birth date). `biomarker_group` and `repeat_number` are optional and are added with a value of 1 when missing. `population_group` is optional; if it is absent, downstream setup treats all individuals as belonging to group 1.}

\item{verbose}{if TRUE, prints warning messages}
}
\value{
the same data object with corrections if needed
}
\description{
Checks the entries of data used in serosolver
}
\examples{
data(example_antibody_data)
check_data(example_antibody_data)
}
\seealso{
Other check_inputs:
\code{\link[=check_attack_rates]{check_attack_rates()}},
\code{\link[=check_demographics]{check_demographics()}},
\code{\link[=check_inf_hist]{check_inf_hist()}},
\code{\link[=check_par_tab]{check_par_tab()}},
\code{\link[=check_proposals]{check_proposals()}}
}
\concept{check_inputs}
