% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_demographics}
\alias{check_demographics}
\title{Checks the entries of demographics data frame used for parameter stratification}
\usage{
check_demographics(demographics, par_tab = NULL, verbose = FALSE)
}
\arguments{
\item{demographics}{the data frame of demographics data. Must have columns: individual (integer ID of individual); birth (numeric time of birth); any other columns also in par_tab$stratification will be checked.}

\item{par_tab}{Optional: the parameter table controlling information such as bounds, initial values etc. If included, this will be used to check that the demographics data is consistent with the data.}

\item{verbose}{if TRUE, prints warning messages}
}
\value{
the same data object with corrections if needed
}
\description{
Checks the entries of demographics data frame used for parameter stratification
}
\examples{
data(example_antibody_data)
check_data(example_antibody_data)
}
\seealso{
Other check_inputs:
\code{\link[=check_attack_rates]{check_attack_rates()}},
\code{\link[=check_data]{check_data()}},
\code{\link[=check_inf_hist]{check_inf_hist()}},
\code{\link[=check_par_tab]{check_par_tab()}},
\code{\link[=check_proposals]{check_proposals()}}
}
\concept{check_inputs}
