% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inf_hist}
\alias{check_inf_hist}
\title{Check infection history matrix}
\usage{
check_inf_hist(
  antibody_data,
  possible_exposure_times,
  inf_hist,
  verbose = FALSE
)

check_inf_hist(
  antibody_data,
  possible_exposure_times,
  inf_hist,
  verbose = FALSE
)
}
\arguments{
\item{antibody_data}{the data frame of titer data}

\item{possible_exposure_times}{the vector of times corresponding to entries in DOB}

\item{inf_hist}{the starting infection history matrix}

\item{verbose}{if TRUE, prints warning messages}
}
\value{
a single boolean value, FALSE if the check passes, otherwise throws an error

nothing, prints a warning
}
\description{
Checks that the infection history matrix is allowable given the birth dates and sampling times of the data
}
\seealso{
Other check_inputs:
\code{\link[=check_attack_rates]{check_attack_rates()}},
\code{\link[=check_data]{check_data()}},
\code{\link[=check_demographics]{check_demographics()}},
\code{\link[=check_par_tab]{check_par_tab()}},
\code{\link[=check_proposals]{check_proposals()}}
}
\concept{check_inputs}
\concept{check_inputs
data(example_antibody_data)
data(example_inf_hist)
times <- 1968:2015
check_inf_hist(example_antibody_data, times, example_inf_hist)}
