% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_par_tab}
\alias{check_par_tab}
\title{Check par_tab for simulate_data}
\usage{
check_par_tab(
  par_tab,
  mcmc = FALSE,
  version = NULL,
  possible_exposure_times = NULL,
  verbose = FALSE
)
}
\arguments{
\item{par_tab}{the parameter table controlling information such as bounds, initial values etc}

\item{mcmc}{logical, if TRUE then checks are performed for the MCMC algorithm. Use FALSE when simulating data}

\item{version}{which version of the posterior function is being used? See \code{\link{create_posterior_func}}}

\item{possible_exposure_times}{optional vector of possible exposure times}

\item{verbose}{if TRUE, prints warning messages}
}
\value{
the same par_tab object with corrections if needed
}
\description{
Checks the entries of par_tab used in simulate_data
}
\examples{
data(example_par_tab)
check_par_tab(example_par_tab, FALSE, version=1)
}
\seealso{
Other check_inputs:
\code{\link[=check_attack_rates]{check_attack_rates()}},
\code{\link[=check_data]{check_data()}},
\code{\link[=check_demographics]{check_demographics()}},
\code{\link[=check_inf_hist]{check_inf_hist()}},
\code{\link[=check_proposals]{check_proposals()}}
}
\concept{check_inputs}
