% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_parTab}
\alias{check_parTab}
\title{Check parTab for simulate_data}
\usage{
check_parTab(parTab, mcmc = FALSE)
}
\arguments{
\item{parTab}{the parameter table controlling information such as bounds, initial values etc}
}
\value{
nothing at the moment
}
\description{
Checks the entries of parTab used in simulate_data
}
