% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_proposals}
\alias{check_proposals}
\title{Checks if the multivariate proposal is being used with the FOI proposal}
\usage{
check_proposals(version, mvr_pars)
}
\arguments{
\item{version}{Which version of the posterior function to use? See \code{\link{create_posterior_func}}}

\item{mvr_pars}{Leave NULL to use univariate proposals. Otherwise, a list of parameters if using a multivariate proposal. Must contain an initial covariance matrix, weighting for adapting cov matrix, and an initial scaling parameter (0-1)}
}
\value{
nothing at the moment
}
\description{
Checks if the multivariate proposal is being used with the FOI proposal
}
\seealso{
Other check_inputs:
\code{\link[=check_attack_rates]{check_attack_rates()}},
\code{\link[=check_data]{check_data()}},
\code{\link[=check_demographics]{check_demographics()}},
\code{\link[=check_inf_hist]{check_inf_hist()}},
\code{\link[=check_par_tab]{check_par_tab()}}
}
\concept{check_inputs}
\keyword{internal}
