% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_cross_reactivity_vector}
\alias{create_cross_reactivity_vector}
\title{Convert melted antigenic map to cross reactivity}
\usage{
create_cross_reactivity_vector(x, cr_gradient, exponential_waning = FALSE)
}
\arguments{
\item{x}{the melted antigenic map}

\item{cr_gradient}{the cross reactivity waning parameter}

\item{exponential_waning}{if TRUE, uses exponential waning rather than linear}
}
\value{
a vector of cross reactivity
}
\description{
Multiplies all elements of the provided vector, x such that y = 1 - cr_gradient*x. Also makes sure that no calculated value is less than 0
}
\keyword{internal}
