% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{create_model_func}
\alias{create_model_func}
\title{Model solving function}
\usage{
create_model_func(parTab, data, antigenicMap, PRIOR_FUNC, ...)
}
\description{
As in \code{\link{create_post_func}}, but returns the matrix of titres rather than a vector of likelihoods
}

