% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{create_post_func}
\alias{create_post_func}
\title{Posterior function pointer}
\usage{
create_post_func(parTab, data, antigenicMap, PRIOR_FUNC, version = 1,
  ageMask = NULL, ...)
}
\arguments{
\item{parTab}{the parameter table controlling information such as bounds, initial values etc}

\item{data}{the data frame of data to be fitted. Must have columns: group (index of group); individual (integer ID of individual); samples (numeric time of sample taken); virus (numeric time of when the virus was circulating); titre (integer of titre value against the given virus at that sampling time)}

\item{antigenicMap}{a data frame of antigenic x and y coordinates. Must have column names: x_coord; y_coord; inf_years}

\item{PRIOR_FUNC}{user function of prior for model parameters. Should take parameter values only}

\item{...}{other arguments to pass to the posterior solving function}
}
\value{
a single function pointer that takes only pars and infectionHistories as unnamed arguments
}
\description{
Essential for the MCMC algorithm. Takes all of the input data/parameters and returns a function pointer. This function finds the posterior for a given set of input parameters (theta) and infection histories without needing to passs the data set back and forth.
}
