% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{create_prior_mu}
\alias{create_prior_mu}
\title{Create strain specific bias prior}
\usage{
create_prior_mu(par_tab)
}
\arguments{
\item{par_tab}{the parameter table as in \code{\link{create_posterior_func}}}
}
\value{
a function pointer
}
\description{
Creates a function that can be used to easily solve the prior on the strain-specific boosting parameters (normal random effects)
}
