% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_sample_mask}
\alias{create_sample_mask}
\title{Create sample mask}
\usage{
create_sample_mask(antibody_data, possible_exposure_times)
}
\arguments{
\item{antibody_data}{the data frame of antibody data. See \code{\link{example_antibody_data}}}

\item{possible_exposure_times}{the vector of times that individuals can be infected}
}
\value{
a vector giving the last index of possible_exposure_times that an individual can be infected
}
\description{
Converts a data frame of individual sampling times to give the index of the last infection that individual could have had (as we can't observe anything after the last observation)
}
\examples{
data(example_antibody_data)
data(example_antigenic_map)
times <- example_antigenic_map$inf_times
sample_mask <- create_sample_mask(example_antibody_data, times)
}
\seealso{
Other create_masks:
\code{\link[=create_age_mask]{create_age_mask()}}
}
\concept{create_masks}
