% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{db}
\alias{db}
\title{Beta prior on an infection}
\usage{
db(x, a, b)
}
\arguments{
\item{x}{infection probability}

\item{a}{first shape parameter of the Beta distribution}

\item{b}{second shape parameter of the Beta distribution}
}
\value{
the density value
}
\description{
Calculates the Beta-density term used for an infection probability.
}
\keyword{internal}
