% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{dbb}
\alias{dbb}
\title{Beta binomial density}
\usage{
dbb(x, N, u, v)
}
\arguments{
\item{x}{the number of successes}

\item{N}{the number of trials}

\item{u}{equivalent to alpha in description}

\item{v}{equivalent to beta in description}
}
\value{
the probability density of the beta binomial with given input parameters
}
\description{
Beta binomial density
}
