% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{describe_priors}
\alias{describe_priors}
\title{Describe infection history priors}
\usage{
describe_priors()
}
\value{
`NULL`, invisibly
}
\description{
Prints a short description of the available infection-history prior versions.
}
