% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_longitudinal_antibody_data}
\alias{example_longitudinal_antibody_data}
\title{Example longitudinal antibody data}
\format{
A data frame with 500 rows (each corresponding to an antibody measurement) and 7 variables: `individual`, `sample_time`, `birth`, `repeat_number`, `biomarker_id`, `biomarker_group`, and `measurement`.
}
\usage{
data(example_longitudinal_antibody_data)
}
\description{
Example antibody data for a longitudinal study design, generated by \code{simulate_data}.
The data contain repeated samples from the same individuals over time and use one biomarker group.
}
\seealso{
Other example_data:
\code{\link{example_antibody_data}},
\code{\link{example_antigenic_map}},
\code{\link{example_inf_chain}},
\code{\link{example_inf_hist}},
\code{\link{example_multiple_biomarker_groups_data}},
\code{\link{example_par_tab}},
\code{\link{example_theta_chain}}
}
\concept{example_data}
\keyword{datasets}
