% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_multiple_biomarker_groups_data}
\alias{example_multiple_biomarker_groups_data}
\title{Example antibody data with multiple biomarker groups}
\format{
A data frame with 1,000 rows (each corresponding to an antibody measurement) and 7 variables: `individual`, `sample_time`, `birth`, `repeat_number`, `biomarker_id`, `biomarker_group`, and `measurement`.
}
\usage{
data(example_multiple_biomarker_groups_data)
}
\description{
Example antibody data generated by \code{simulate_data} with two biomarker groups. The two groups represent different observation types measured for the same individuals and sampling times.
}
\seealso{
Other example_data:
\code{\link{example_antibody_data}},
\code{\link{example_antigenic_map}},
\code{\link{example_inf_chain}},
\code{\link{example_inf_hist}},
\code{\link{example_longitudinal_antibody_data}},
\code{\link{example_par_tab}},
\code{\link{example_theta_chain}}
}
\concept{example_data}
\keyword{datasets}
