% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_theta_chain}
\alias{example_theta_chain}
\title{Example MCMC chain for kinetics parameters}
\format{
A data frame with 3,003 rows (each corresponding to a saved sample from one of three chains) and 22 columns containing the saved sample number, chain number, parameter draws, likelihood and prior probabilities, and total numbers of infections.
}
\usage{
data(example_theta_chain)
}
\description{
Outputs of a quick MCMC run corresponding to the example data. See \code{\link{example_inf_chain}} for corresponding infection history chain output
}
\seealso{
Other example_data:
\code{\link{example_antibody_data}},
\code{\link{example_antigenic_map}},
\code{\link{example_inf_chain}},
\code{\link{example_inf_hist}},
\code{\link{example_longitudinal_antibody_data}},
\code{\link{example_multiple_biomarker_groups_data}},
\code{\link{example_par_tab}}
}
\concept{example_data}
\keyword{datasets}
