% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_titre_dat}
\alias{example_titre_dat}
\title{Example titre data}
\format{A data frame with 4800 rows (each corresponding to a unique titre reading) and 6 variables:
\describe {
    \item{individual}{index starting from 1 for the individual ID}
    \item{samples}{time of this serum sample, matching times in \code{\link{example_antigenic_map}}}
    \item{virus}{circlation time of the strain being measured, matching times in \code{\link{example_antigenic_map}}}
    \item{titre}{the titre measurement for this strain at this sampling time}
    \item{run}{where there are repeat measurements, gives the repeat number for this serum sample and strain measured}
    \item{group}{group index of this individual}
    \item{DOB}{date of birth matching the time resolution of the model}
}}
\usage{
data(example_titre_data)
}
\description{
Example of the titre data data frame generated by \code{\link{simulate_data}}, with matching infection history in \code{\link{example_inf_hist}}
}
\seealso{
Other example_data: 
\code{\link{example_antigenic_map}},
\code{\link{example_inf_chain}},
\code{\link{example_inf_hist}},
\code{\link{example_par_tab}},
\code{\link{example_theta_chain}}
}
\concept{example_data}
\keyword{datasets}
