% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{expand_summary_inf_chain}
\alias{expand_summary_inf_chain}
\title{Expand sparse infection history matrix}
\usage{
expand_summary_inf_chain(inf_chain, j_vec = NULL)
}
\arguments{
\item{inf_chain}{the data table with the saved sparse infection history matrix}

\item{j_vec}{optional vector of infection-time indices to expand the infection history chain for}
}
\value{
a data table with zero entries added for missing infection events and one column for each infection-time index
}
\description{
Expand sparse infection history matrix
}
\seealso{
Other mcmc:
\code{\link[=generate_start_tab]{generate_start_tab()}},
\code{\link[=rm_scale]{rm_scale()}},
\code{\link[=save_infection_history_to_disk]{save_infection_history_to_disk()}},
\code{\link[=scaletuning]{scaletuning()}},
\code{\link[=scaletuning_alt]{scaletuning_alt()}},
\code{\link[=serosolver]{serosolver()}}
}
\concept{mcmc}
