% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extend_par_tab_biomarker_groups}
\alias{extend_par_tab_biomarker_groups}
\title{Extend parameter table for multiple biomarker_groups}
\usage{
extend_par_tab_biomarker_groups(par_tab, n_obs_types)
}
\arguments{
\item{par_tab}{the parameter table to extend}

\item{n_obs_types}{number of biomarker_groups}
}
\value{
the updated parameter table
}
\description{
Extend parameter table for multiple biomarker_groups
}
