% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{find_a_b}
\alias{find_a_b}
\title{Beta binomial parameters}
\usage{
find_a_b(n, mean, var)
}
\arguments{
\item{n}{the number of trials}

\item{mean}{desired mean}

\item{var}{desired variance}
}
\value{
alpha and beta for the beta binomial
}
\description{
Finds the required alpha and beta to give a desired mean and variance of the beta binomal function
}

