% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{find_bb_2}
\alias{find_bb_2}
\title{Beta binomial parameter match}
\usage{
find_bb_2(n1, a1, b1, n2)
}
\arguments{
\item{n1}{the number of trials in the first distribution}

\item{a1}{alpha in the first distribution}

\item{b1}{beta in the first distribution}

\item{n2}{the number of trials in the second distribution}
}
\value{
alpha and beta for the second distribution
}
\description{
Given a beta binomial function with known N, alpha and beta, finds the alpha and beta for a beta binomial with a different N that gives the same mean and variance
}
\seealso{
Other beta_binomial_functions: 
\code{\link{bb_mean}()},
\code{\link{dbb}()},
\code{\link{find_prior_alpha_beta}()},
\code{\link{hist_rbb}()},
\code{\link{pbb}()},
\code{\link{qbb}()},
\code{\link{rbb}()}
}
\concept{beta_binomial_functions}
