% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{find_beta_parameters}
\alias{find_beta_parameters}
\title{Find beta parameters for mean and variance}
\usage{
find_beta_parameters(mean, var, make_plot = FALSE)
}
\arguments{
\item{mean}{the mean of the beta distribution (between 0 and 1)}

\item{var}{the variance of the beta distribution (this is likely going to be somewhere less than 0.25)}

\item{make_plot}{if TRUE, plots the resulting beta distibution to the R device}
}
\value{
a list with alpha and beta
}
\description{
Finds the alpha and beta parameters for a beta distribution that gives the desired mean and variance
}
