% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{find_prior_alpha_beta}
\alias{find_prior_alpha_beta}
\title{Beta binomial parameters}
\usage{
find_prior_alpha_beta(n, mean, var)
}
\arguments{
\item{n}{the number of trials}

\item{mean}{desired mean}

\item{var}{desired variance}
}
\value{
alpha and beta for the beta binomial
}
\description{
Finds the required alpha and beta to give a desired mean and variance of the beta binomal function
}
\examples{
find_prior_alpha_beta(n=100,mean=25,var=5)
}
\seealso{
Other beta_binomial_functions: 
\code{\link{bb_mean}()},
\code{\link{dbb}()},
\code{\link{find_bb_2}()},
\code{\link{hist_rbb}()},
\code{\link{pbb}()},
\code{\link{qbb}()},
\code{\link{rbb}()}
}
\concept{beta_binomial_functions}
