% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{fit_beta_prior}
\alias{fit_beta_prior}
\title{Fit beta distribution to MCMC output}
\usage{
fit_beta_prior(
  chain_samples,
  par_name = "",
  error_tol = 999999999,
  try_attempts = 10,
  plot_fit = TRUE
)
}
\arguments{
\item{chain_samples}{the MCMC chain data frame to be fit to}

\item{par_name}{the column label to fit to}

\item{error_tol}{= 9999999999999, what's the error tolerance on the fit? Might take some tweaking}

\item{try_attempts}{= 10 how many fitting attempts to try before giving up}

\item{plot_fit}{= TRUE, if TRUE, plots the fit to the MCMC chain}
}
\value{
the model fit object as returned by optim
}
\description{
Attempts to fit a beta distribution to a data frame of MCMC output from a previous run.
}
\examples{
\dontrun{
## Output from a previous serosolver chain
chain <- read.csv("madeup_chain.csv")
results <- fit_beta_prior(chain, par_name="sigma1",plot_fit=FALSE)
}
}
\seealso{
\code{\link{fit_normal_prior}}

Other priors:
\code{\link[=calc_phi_probs]{calc_phi_probs()}},
\code{\link[=calc_phi_probs_indiv]{calc_phi_probs_indiv()}},
\code{\link[=calc_phi_probs_spline]{calc_phi_probs_spline()}},
\code{\link[=create_prob_shifts]{create_prob_shifts()}},
\code{\link[=find_beta_prior_mode]{find_beta_prior_mode()}},
\code{\link[=find_beta_prior_with_mean]{find_beta_prior_with_mean()}},
\code{\link[=find_beta_prior_with_mean_var]{find_beta_prior_with_mean_var()}},
\code{\link[=fit_normal_prior]{fit_normal_prior()}},
\code{\link[=inf_mat_prior]{inf_mat_prior()}},
\code{\link[=infection_history_prior]{infection_history_prior()}},
\code{\link[=prob_shifts]{prob_shifts()}}
}
\concept{priors}
