% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fonville_map}
\alias{fonville_map}
\title{Antigenic map}
\format{A data frame with the following variables:

  \describe{

  \item{strain}{Virus strain}

  \item{x}{X-coordinate}

  \item{y}{Y-coordinate}

  \item{year}{The year the strain circulated}

  }}
\usage{
fonville_map
}
\description{
Antigenic map
}
\keyword{datasets}
