% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fonville_map_approx}
\alias{fonville_map_approx}
\title{Antigenic map}
\format{A data frame with the following variables:

  \describe{

  \item{Strain}{Virus strain}

  \item{X}{X-coordinate}

  \item{Y}{Y-coordinate}

  }}
\usage{
fonville_map_approx
}
\description{
Antigenic map
}
\keyword{datasets}
