% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fromUnitScale}
\alias{fromUnitScale}
\title{Convert from unit scale to original scale}
\usage{
fromUnitScale(x, min, max)
}
\arguments{
\item{x}{value on the 0--1 scale}

\item{min}{lower bound of the original scale}

\item{max}{upper bound of the original scale}
}
\value{
the value on the original scale
}
\description{
Converts a value from the 0--1 scale to its original bounded scale.
}
\keyword{internal}
