% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_all_plots}
\alias{generate_all_plots}
\title{Generates all output plots}
\usage{
generate_all_plots(outputDir, adaptive_period, chainFile, infHistFile,
  titreDat, antigenicMap, parTab, ages, nIndiv = 10, nSamp = 1000,
  filename = "test", thin = 100)
}
\arguments{
\item{outputDir}{the file location to save output to}

\item{adaptive_period}{specifies how many iterations to discard as burning}

\item{chainFile}{the full file location of the MCMC chain output for theta, the parameter vector}

\item{infHistFile}{the full file location of the MCMC chain output for infection histories}

\item{titreDat}{the data frame of titre data used in the MCMC}

\item{antigenicMap}{the unmelted antigenic map specifying the relationship between strains}

\item{parTab}{the full parameter control table as used in the MCMC procedure}

\item{ages}{a data frame with columns for individual and date of birth (titled DOB), where the date of birth is in the same time resolution as the model (ie. DOB in years would be 1991, in months would be 1991*12)}

\item{nIndiv}{specifies how many random individuals to sample for the infection history plot}

\item{nSamp}{how many samples from the posterior to take when generating the plot}

\item{filename}{the actual filename is appended to this part. For example, "test" would give "test_plots.svg", "test_ar.svg" etc}

\item{thin}{Optional parameter to thin the infection history chain for speed}
}
\value{
nothing
}
\description{
Given an output file location, locations of output files from the MCMC procedure and some control parameters, generates and saves infection history plots, posterior plots (trace and density), attack rates and total number of infections.
}
