% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{generate_alpha_beta_prior_mode}
\alias{generate_alpha_beta_prior_mode}
\title{Find Beta distribution parameters with mode}
\usage{
generate_alpha_beta_prior_mode(mode1, k)
}
\arguments{
\item{mode1}{the desired mode}

\item{the}{desired certainty in the prior, must be at least 2. The higher this number, the "stronger" the prior}
}
\value{
a list with alpha and beta parameters for the Beta distribution
}
\description{
Calculates the necessary parameters for the beta distribution to give the desired mode and certainty, k
}
