% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{generate_beta_prior_mean}
\alias{generate_beta_prior_mean}
\title{Find beta parameters with maximum variance}
\usage{
generate_beta_prior_mean(desired_annual_mean, buckets)
}
\arguments{
\item{desired_annual_mean}{the desired ANNUAL mean attack rate}

\item{the}{number of buckets to split each year into}
}
\value{
a list with alpha and beta
}
\description{
Given a desired annual mean attack rate and the number of epochs to consider per year, gives the alpha and beta parameters with the desired mean but maximum variance
}
