% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infection_model_likelihoods.R
\name{generate_phis}
\alias{generate_phis}
\title{Generate FOI phis - INACTIVE}
\usage{
generate_phis(foi, knots, theta, n_years, buckets, degree = 2)
}
\arguments{
\item{foi}{vector of force of infections per year}

\item{knots}{vector of knots for the spline}

\item{theta}{vector of theta parameters for spline}

\item{n_years}{number of years to calculate}

\item{buckets}{number of buckets per year (12 for monthly, 1 for annual)}

\item{degree}{degree of the spline}
}
\value{
a vector of FOIs for each time point
}
\description{
Calculates the seasonal spline for \code{\link{calc_phi_probs_spline}}
}
\keyword{internal}
