% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_attack_rates}
\alias{get_attack_rates}
\title{Estimate attack rates}
\usage{
get_attack_rates(
  infection_histories,
  strain_isolation_times,
  titre_dat,
  n_alive = NULL,
  by_group = FALSE
)
}
\arguments{
\item{infection_histories}{the MCMC chain for infection histories}

\item{strain_isolation_times}{vector of the epochs of potential circulation}

\item{titre_dat}{the data frame of titre data}

\item{n_alive}{vector with the number of people alive in each year of circulation. Can be left as NULL, and ages will be used to infer this}

\item{by_group}{if TRUE, stratifies by group ID}
}
\value{
a tibble with the estimated attack rates for each potential epoch of circulation
}
\description{
Extracts attack rates from the MCMC output on infection histories
}
