% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_best_pars}
\alias{get_best_pars}
\title{Best pars}
\usage{
get_best_pars(chain)
}
\arguments{
\item{chain}{the MCMC chain}
}
\value{
a named vector of parameter values from the maximum-posterior row
}
\description{
Given an MCMC chain, returns the parameter values from the row with the highest posterior probability.
}
\examples{
\dontrun{
mcmc_chains <- load_theta_chains()
best_pars <- get_best_pars(mcmc_chains$chain)
}
}
\seealso{
Other mcmc_diagnostics:
\code{\link[=get_best_draw]{get_best_draw()}},
\code{\link[=get_index_pars]{get_index_pars()}}
}
\concept{mcmc_diagnostics}
