% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_index_pars}
\alias{get_index_pars}
\title{Index pars}
\usage{
get_index_pars(chain, samp_no = NULL, index = NULL, chain_no = NULL)
}
\arguments{
\item{chain}{the MCMC chain}

\item{samp_no}{the sample number}

\item{index}{if not using `samp_no`, `index returns the desired row number`}

\item{chain_no}{the chain number to subset}
}
\value{
a named vector of the best parameters
}
\description{
Given an MCMC chain, returns the parameters at the specified index
}
\examples{
\dontrun{
mcmc_chains <- load_theta_chains()
pars <- get_index_pars(mcmc_chains$chain, index=100)
}
}
\seealso{
Other mcmc_diagnostics:
\code{\link[=get_best_draw]{get_best_draw()}},
\code{\link[=get_best_pars]{get_best_pars()}}
}
\concept{mcmc_diagnostics}
