% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_n_alive_location}
\alias{get_n_alive_location}
\title{Get number alive bylocation}
\usage{
get_n_alive_location(titre_dat, times)
}
\arguments{
\item{titre_dat}{the data frame of titre data}

\item{times}{the vector of times to calculate number alive for}
}
\value{
a matrix giving the number alive in each time point in each location
}
\description{
Given the titre_dat data frame with entries for location, calculates the number that are alive (alive to be infected, that is) for each time in times by location
}
