% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_titre_predictions}
\alias{get_titre_predictions}
\title{Need to mask by age and sample time}
\usage{
get_titre_predictions(chain, infectionHistories, dat, individuals, antigenicMap,
  ages, nsamp = 100)
}
\description{
Need to mask by age and sample time
}

