% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{group_likelihood_total}
\alias{group_likelihood_total}
\title{Group likelihood}
\usage{
group_likelihood_total(theta, infectionHistories, indicesA, indicesB,
  samplingTimes, strainIsolationTimes, antigenicMapLong, antigenicMapShort,
  titres)
}
\arguments{
\item{theta}{NumericVector, the named vector of model parameters}

\item{infectionHistories}{NumericMatrix, the matrix of 1s and 0s showing presence/absence of infection for each possible time for each individual}

\item{indicesA}{IntegerVector, the range of indices of the titre vector that correspond to each individual (eg. first 5 titres are for indiv 1, indicesA = c(0,5,...)}

\item{indicesB}{IntegerVector, the range of indices of the sample vector that correspond to each individual (eg. first 2 sampling times are for indiv 1, indicesB = c(0,2,...)}

\item{samplingTimes}{NumericVector, the vector of real times that samples were taken}

\item{strainIsolationTimes}{NumericVector, the vector of times at which each virus strain circulated}

\item{antigenicMapLong}{NumericVector, the collapsed cross reactivity map for long term boosting, after multiplying by sigma1}

\item{antigenicMapShort}{NumericVector, the collapsed cross reactivity map for short term boosting, after multiplying by sigma2}

\item{titres}{NumericVector, the vector of observed titres for all individuals.}
}
\value{
a single log likelihood
}
\description{
Uses \code{\link{individual_likelihood}} for each individual and returns a summed log likelihood for all individuals
}

