% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{group_likelihood_vector}
\alias{group_likelihood_vector}
\title{Group likelihood}
\usage{
group_likelihood_vector(theta, infectionHistories, indicesSamples, indicesData,
  indicesDataOverall, samplingTimes, strainIsolationTimes, virusIndices,
  antigenicMapLong, antigenicMapShort, titres)
}
\arguments{
\item{theta}{NumericVector, the named vector of model parameters}

\item{infectionHistories}{NumericMatrix, the matrix of 1s and 0s showing presence/absence of infection for each possible time for each individual}

\item{indicesData}{IntegerVector, the range of indices of the sample vector that correspond to each individual (eg. first 2 sampling times are for indiv 1, indicesB = c(0,2,...)}

\item{samplingTimes}{NumericVector, the vector of real times that samples were taken}

\item{strainIsolationTimes}{NumericVector, the vector of times at which each virus strain circulated}

\item{antigenicMapLong}{NumericVector, the collapsed cross reactivity map for long term boosting, after multiplying by sigma1}

\item{antigenicMapShort}{NumericVector, the collapsed cross reactivity map for short term boosting, after multiplying by sigma2}

\item{titres}{NumericVector, the vector of observed titres for all individuals.}

\item{indicesSampling}{IntegerVector, the range of indices of the titre vector that correspond to each individual (eg. first 5 titres are for indiv 1, indicesA = c(0,5,...)}

\item{n_strains}{int, the maximum number of strains that could be tested against}
}
\value{
a NumericVector of log likelihoods for each individual
}
\description{
Uses \code{\link{individual_likelihood}} for each individual and returns a vector of log likelihoods for each individual
}
