% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{hist_rbb}
\alias{hist_rbb}
\title{Beta binomial histogram}
\usage{
hist_rbb(n, mean, var)
}
\arguments{
\item{n}{the number of trials}

\item{mean}{desired mean}

\item{var}{desired variance}
}
\description{
Plots a histogram of the beta binomial function with given mean and variance
}
