% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_distributions.R
\name{identify_run_lengths}
\alias{identify_run_lengths}
\title{Identify runs of infections from posterior}
\usage{
identify_run_lengths(inf_chain)
}
\arguments{
\item{inf_chain}{data table of the infection histories posterior}
}
\value{
a tibble giving the consecutive infection run length, the start and end time of each run, which index the run is (i.e., which distinct infection), and the time from the end of the previous run, for each `i` and `samp_no`
}
\description{
For each individual and MCMC iteration, uses the infection history MCMC chain and detects runs of consecutive infections.
}
\examples{
\dontrun{
identify_run_lengths(inf_chain)
}
}
