% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{individual_likelihood}
\alias{individual_likelihood}
\title{Individual likelihood}
\usage{
individual_likelihood(theta, infectionHistory, samplingTimes, indivIndices,
  strainIsolationTimes, indivVirusIndices, antigenicMapLong, antigenicMapShort,
  titres, numberStrains)
}
\arguments{
\item{theta}{NumericVector, the named vector of model parameters}

\item{infectionHistory}{NumericVector, the vector of 1s and 0s showing presence/absence of infection for each possible time.}

\item{samplingTimes}{NumericVector, the vector of real times that samples were taken}

\item{strainIsolationTimes}{NumericVector, the vector of times at which each virus strain circulated}

\item{antigenicMapLong}{NumericVector, the collapsed cross reactivity map for long term boosting, after multiplying by sigma1}

\item{antigenicMapShort}{NumericVector, the collapsed cross reactivity map for short term boosting, after multiplying by sigma2}

\item{titres}{NumericVector, the vector of observed titres.}
}
\value{
a single log likelihood
}
\description{
Uses \code{\link{infection_model_indiv}} for a single individual and then calls \code{\link{likelihood_titre}}. Works with NA
}
