% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{inf_hist_prob_lambda}
\alias{inf_hist_prob_lambda}
\title{DEPRECATED - propose inf hist from lambda}
\usage{
inf_hist_prob_lambda(new_inf, sampled_indivs, age_mask, strain_mask,
  n_infs, lambdas)
}
\description{
DEPRECATED - propose inf hist from lambda
}
