% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{inf_hist_prob_phi}
\alias{inf_hist_prob_phi}
\title{DEPRECATED - propose inf hist from phi}
\usage{
inf_hist_prob_phi(new_inf, sampled_indivs, age_mask, strain_mask, n_infs,
  phis)
}
\description{
DEPRECATED - propose inf hist from phi
}
