% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{inf_hist_swap}
\alias{inf_hist_swap}
\title{Swap infection history years}
\usage{
inf_hist_swap(
  infection_history,
  inf_hist_masks,
  proposal_inf_hist_indiv_swap_ratio,
  proposal_inf_hist_distance,
  proposal_ratios = NULL
)
}
\arguments{
\item{infection_history}{matrix of 1s and 0s to swap around representing the infection history}

\item{inf_hist_masks}{matrix of 1s and 0s corresponding to infection_history, with a 1 where the infection state is eligible to be updated and a 0 otherwise}

\item{proposal_inf_hist_indiv_swap_ratio}{what proportion of infections should be swapped?}

\item{proposal_inf_hist_distance}{How many time points away should be chosen as candidate swaps?}

\item{proposal_ratios}{optional NULL. Can set the relative sampling weights of the infection state times. Should be an integer vector of length matching nrow(antigenic_map). Otherwise, leave as NULL for uniform sampling.}
}
\value{
the same infection_history matrix, but with two columns swapped
}
\description{
Swaps the entire contents of two columns of the infection history matrix, adhering to age and sample time limitations.
}
\examples{
data(example_inf_hist)
data(example_antibody_data)
data(example_antigenic_map)
times <- example_antigenic_map$inf_times
ages <- unique(example_antibody_data[,c("birth","individual")])
age_mask <- create_age_mask(ages$birth, times)
sample_mask <- create_sample_mask(example_antibody_data, times)
new_inf_hist <- inf_hist_swap(example_inf_hist, age_mask,sample_mask, 1,3)[[1]]
}
\seealso{
Other proposals:
\code{\link[=inf_hist_swap_phi]{inf_hist_swap_phi()}},
\code{\link[=infection_history_symmetric]{infection_history_symmetric()}},
\code{\link[=mvr_proposal]{mvr_proposal()}},
\code{\link[=univ_proposal]{univ_proposal()}}
}
\concept{proposals}
