% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{inf_hist_swap_phi}
\alias{inf_hist_swap_phi}
\title{Swap infection history years with phi term}
\usage{
inf_hist_swap_phi(
  infection_history,
  phis,
  age_mask,
  sample_mask,
  proposal_inf_hist_indiv_swap_ratio,
  proposal_inf_hist_distance,
  n_alive
)
}
\arguments{
\item{infection_history}{matrix of 1s and 0s to swap around representing the infection history}

\item{phis}{vector of force of infection parameters for each column}

\item{age_mask}{the age mask, giving the first index of the infection_history matrix that each individual can be exposed to. One entry per individual}

\item{sample_mask}{the sample mask, giving the last index of the infection_history matrix that each individual can be exposed to. One entry per individual}

\item{proposal_inf_hist_indiv_swap_ratio}{what proportion of infections should be swapped?}

\item{proposal_inf_hist_distance}{How many time points away should be chosen as candidate swaps?}

\item{n_alive}{number of individuals alive in each entry of phis}
}
\value{
a list: the same infection_history matrix, but with two columns swapped; also the swapped phis
}
\description{
Swaps the entire contents of two columns of the infection history matrix, adhering to age and strain limitations. Also swaps the values of phi that correspond to these years
}
\examples{
data(example_inf_hist)
data(example_antibody_data)
data(example_antigenic_map)
times <- example_antigenic_map$inf_times
ages <- unique(example_antibody_data[,c("birth","individual")])
age_mask <- create_age_mask(ages$birth, times)
sample_mask <- create_sample_mask(example_antibody_data, times)
phis <- runif(length(times))
n_alive <- get_n_alive(example_antibody_data,times)
new_inf_hist <- inf_hist_swap_phi(example_inf_hist, phis, age_mask,sample_mask, 1,3, n_alive)
}
\seealso{
\code{\link{inf_hist_swap}}

Other proposals:
\code{\link[=inf_hist_swap]{inf_hist_swap()}},
\code{\link[=infection_history_symmetric]{infection_history_symmetric()}},
\code{\link[=mvr_proposal]{mvr_proposal()}},
\code{\link[=univ_proposal]{univ_proposal()}}
}
\concept{proposals}
