% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{infection_history_betabinom}
\alias{infection_history_betabinom}
\title{Beta binomial infection history update}
\usage{
infection_history_betabinom(newInfHist, sampledIndivs, ageMask, moveSizes,
  alpha, beta)
}
\arguments{
\item{newInfHist}{a matrix of infection histories - rows for individuals, columns for infection epochs. Contents should be 1s and 0s}

\item{sampledIndivs}{a vector of indices describing rows in the infection history matrix that should be updated}

\item{ageMask}{a vector (one value for each individual) giving the first infection epoch that an individual could have been exposed in. That is, if an individual was born in the 7th epoch, their entry in ageMask would be 7.}

\item{moveSizes}{when performing a move step, how far should two epochs be swapped?}

\item{alpha}{alpha parameter of beta binomial}

\item{beta}{beta parameter of beta binomial}
}
\value{
a matrix of infection histories matching the input newInfHist
}
\description{
Samples a new infection history from a beta binomial distribution for the specified number of individuals. Note that only one epoch is updated each time with either a flip or a swap step.
}

