% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{infection_history_betabinom_group}
\alias{infection_history_betabinom_group}
\title{DEPRECATED - implemented in Cpp for speed}
\usage{
infection_history_betabinom_group(newInfHist, sampledIndivs, ageMask,
  strainMask, moveSizes, nInfs, alpha, beta)
}
\description{
DEPRECATED - implemented in Cpp for speed
}
