% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{infection_history_proposal}
\alias{infection_history_proposal}
\title{Infection history proposal - original version}
\usage{
infection_history_proposal(new_infection_histories, sampled_indivs,
  strain_isolation_times, age_mask, strain_mask, n_infs)
}
\arguments{
\item{new_infection_histories}{an n*m matrix of 1s & 0s indicating infection histories, where n is individuals and m i strains}

\item{sampled_indivs}{the indices of sampled individuals to receive proposals}

\item{strain_isolation_times}{the vector of strain isolation times in real time}

\item{age_mask}{the vector of indices for each individual specifiying which index of strain_isolation_times is the first strain each individual could have seen}
}
\value{
a new matrix matching new_infection_histories in dimensions with proposed moves
}
\description{
Proposes new infection histories for a vector of infection histories, where rows represent individuals and columns represent years. Proposals are either removal, addition or switching of infections.
Also requires the indices of sampled individuals, the vector of strain isolation times, and a vector of age masks (ie. which index of the strain_isolation_times vector is the first year in which
an individual *could* be infected).
NOTE - MIGHT NEED TO UPDATE THIS FOR GROUPS
}
