% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infection_history_proposal_gibbs}
\alias{infection_history_proposal_gibbs}
\title{Fast infection history proposal function}
\usage{
infection_history_proposal_gibbs(pars, infHist, n_indivs_samp, n_years_samp,
  ageMask, n_alive, alpha, beta)
}
\value{
a matrix of 1s and 0s corresponding to the infection histories for all individuals
}
\description{
Fast infection history proposal function
}

