% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infection_history_proposal_gibbs_mus}
\alias{infection_history_proposal_gibbs_mus}
\title{Gibbs sampling of infection histories with random effects boosting}
\usage{
infection_history_proposal_gibbs_mus(pars, mus, mu_indices, infHist,
  indivSampPropn, n_years_samp, ageMask, n_alive, swapPropn, swapDistance,
  alpha, beta, circulationTimes, circulationMapIndices, samplingTimes,
  indicesTitreDataSample, indicesTitreDataOverall, indicesSamples,
  measuredMapIndices, antigenicMapLong, antigenicMapShort, data, temp)
}
\arguments{
\item{pars}{NumericVector, the model parameters used to solve the model, extract likelihood function parameters and alpha/beta}

\item{mus}{NumericVector, boosting parameters for random effects boosting}

\item{mu_indices}{IntegerVector, boosting parameter indices for random effects boosting}

\item{infHist}{IntegerMatrix the matrix of 1s and 0s corresponding to individual infection histories}

\item{indivSampPropn}{double, what proportion of individuals to resample in this proposal step}

\item{n_years_samp}{int, for each individual, how many time periods to resample infections for?}

\item{ageMask}{IntegerVector, length of the number of individuals, with indices specifying first time period that an individual can be infected (indexed from 1, such that a value of 1 allows an individual to be infected in any time period)}

\item{n_alive}{IntegerVector, length of the number of time periods that an individual could be infected, giving the number of individual alive in each time period}

\item{swapPropn}{double, what proportion of proposals should be swap steps (ie. swap contents of two cells in infHist rather than adding/removing infections)}

\item{swapDistance}{int, in a swap step, how many time steps either side of the chosen time period to swap with}

\item{alpha}{double, alpha parameter for beta prior on infection probability}

\item{beta}{double, beta parameter for beta prior on infection probability}

\item{circulationTimes}{NumericVector, the times that each strain circulated}

\item{circulationMapIndices}{IntegerVector, indexing vector from 1:number of strains}

\item{samplingTimes}{NumericVector, the vector of real times that samples were taken}

\item{indicesTitreDataOverall}{IntegerVector, How many rows in the titre data correspond to each individual?}

\item{indicesSamples}{IntegerVector, Split the sample times and runs for each individual}

\item{measuredMapIndices}{IntegerVector, For each titre measurement, corresponding entry in antigenic map}

\item{antigenicMapLong}{NumericVector, the collapsed cross reactivity map for long term boosting, after multiplying by sigma1}

\item{antigenicMapShort}{NumericVector, the collapsed cross reactivity map for short term boosting, after multiplying by sigma2}

\item{data}{NumericVector, the titre data for all individuals}

\item{temp}{double, tepmerature for parallel tempering}

\item{indicesTitreDataSamples}{IntegerVector, How many rows in titre data correspond to each individual, sample and repeat?}
}
\value{
a matrix of 1s and 0s corresponding to the infection histories for all individuals
}
\description{
Proposes a new matrix of infection histories by sampling from the prior on an individual's infection presence/absence in a particular time period, conditional on all other individuals in that time period. This allows us to integrate out the infection probability term for each time period. Should look at \code{\link{create_post_func}} for more details about the input parameters.
}

