% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{infection_history_proposal_gibbs_R}
\alias{infection_history_proposal_gibbs_R}
\title{R implementation of the infection history gibbs proposal, now in C++ code}
\usage{
infection_history_proposal_gibbs_R(pars, infection_history, indivSampPropn,
  n_years_samp, age_mask, n_alive, alpha, beta, strains, strainIndices,
  sampleTimes, indicesData, indicesDataOverall, indicesSamples,
  virusIndices, antigenic_map_long, antigenic_map_short, titres)
}
\description{
R implementation of the infection history gibbs proposal, now in C++ code
}
