% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{lambda_proposal}
\alias{lambda_proposal}
\title{DEPRECATED - proposal for lambda based on number of infections}
\usage{
lambda_proposal(current_pars, infection_history, years, js, alpha, beta,
  n_alive)
}
\description{
DEPRECATED - proposal for lambda based on number of infections
}
