% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood_data_individual}
\alias{likelihood_data_individual}
\title{Likelihood for one individual}
\usage{
likelihood_data_individual(theta, infectionHistory, circulationTimes,
  circulationMapIndices, samplingTimes, dataIndices, measuredMapIndices,
  antigenicMapLong, antigenicMapShort, numberStrains, data, titre_shifts,
  DOB, additional_arguments)
}
\description{
See infection_model_indiv, does the same thing but returns the log likelihood given some data. This is used by infection_history_proposal_gibbs
}
