% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood_titre_bias}
\alias{likelihood_titre_bias}
\title{Calculate likelihood}
\usage{
likelihood_titre_bias(expected, data, theta, to_add)
}
\arguments{
\item{expected}{NumericVector, as returned by \code{\link{infection_model_indiv}}}

\item{data}{NumericVector, the vector of observed titres}

\item{theta}{NumericVector, the vector of named model parameters, requiring MAX_TITRE and error}
}
\value{
a single log likelihood
}
\description{
Calculates the likelihood of a given set of observed titres given predicted titres. Based on truncated discritised normal.
}

