% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_antigenic_map_file}
\alias{load_antigenic_map_file}
\title{Read in antigenic_map}
\usage{
load_antigenic_map_file(location = getwd())
}
\value{
one of the starting parameter tables in the current working directory
}
\description{
Searches the working directory for a file with "antigenic_map" in the name and reads this in
}
