% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_titre_dat}
\alias{load_titre_dat}
\title{Read in titre_dat}
\usage{
load_titre_dat(location = getwd())
}
\value{
one of the starting parameter tables in the current working directory
}
\description{
Searches the working directory for a file with "titre_dat" in the name and reads this in
}
