% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenic_funcs.R
\name{make_antigenic_maps_vac1}
\alias{make_antigenic_maps_vac1}
\title{Function to make the antigenic maps (including vaccination)}
\usage{
make_antigenic_maps_vac1(antigenic_map_melted, theta)
}
\description{
Function to make the antigenic maps (including vaccination)
}
