% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{outputdmatrix.fromcoord.original}
\alias{outputdmatrix.fromcoord.original}
\title{Needed to solve the original implementation of the model}
\usage{
outputdmatrix.fromcoord.original(thetasigma, inf_years, anti.map.in,
  linearD = F)
}
\description{
Needed to solve the original implementation of the model
}

