% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{outputdmatrix.fromcoord}
\alias{outputdmatrix.fromcoord}
\title{Create useable antigenic map}
\usage{
outputdmatrix.fromcoord(anti.map.in)
}
\arguments{
\item{anti.map.in}{can either be a 1D antigenic line to calculate distance from, or a two dimensional matrix with x and y coordinates on an antigenic map}
}
\value{
the euclidean antigenic distance between each pair of viruses in anti.map.in
}
\description{
Creates an antigenic map from an input data frame that can be used to calculate cross reactivity. This will end up being an NxN matrix, where there are N strains circulating.
}
