% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{p_infection}
\alias{p_infection}
\title{Calculate prob of infection}
\usage{
p_infection(psi, titre, alpha1, beta1)
}
\arguments{
\item{psi}{the baseline risk of infection to an individual with no titre}

\item{titre}{the log HI titre of the individual (base 2)}

\item{alpha}{parameter 1}

\item{beta}{parameter 2}
}
\value{
a single probability of infection
}
\description{
Calculates the probability of infection given a baseline risk and the individual's titre
}
