% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pad_inf_chain}
\alias{pad_inf_chain}
\title{Pad infection history chain}
\usage{
pad_inf_chain(inf_chain, pad_by_group = FALSE, times = NULL, indivs = NULL)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{serosolver}}}

\item{pad_by_group}{if TRUE, accounts for population group when expanding}

\item{times}{if not NULL, uses this as a vector of times to replace j when expanding to all combinations}

\item{indivs}{if not NULL, uses this as a vector of individuals to replace i when expanding to all combinations}
}
\value{
the same `inf_chain` with 0s added for missing `i`/`j`/`samp_no` combinations, and `population_group` combinations when `pad_by_group = TRUE`
}
\description{
Given that the MCMC sampler only stores present infections (ie. there are no entries for 0s from the infection history matrix), for some summaries we need to add these 0s back in to avoid bias.
}
