% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{par_tab_base}
\alias{par_tab_base}
\title{Base parameter table}
\format{A data frame with the following variables:

  \describe{

  \item{names}{Parameter name}

  \item{values}{Parameter value}

  \item{fixed}{1 - fixed, 0 - needs to be estimated}

  \item{steps}{Who knows}

  \item{lower_bound}{Lower bound for value}

  \item{upper_bound}{Upper bound for value}

  \item{lower_start}{No idea}

  \item{upper_start}{Haven't a clue}

  \item{type}{Don't know}

  }}
\usage{
par_tab_base
}
\description{
Base parameter table
}
\keyword{datasets}
