% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_PT_aux_fast.R
\name{parallel_tempering}
\alias{parallel_tempering}
\title{performs parallel tempering}
\usage{
parallel_tempering(mcmc_list, temperatures, offset)
}
\arguments{
\item{mcmc_list}{a list of lists: values, log likelihood etc. of parallel MCMC chains}

\item{temperatures}{numeric vector: temperatures of chains}

\item{offset}{integer: 0 or 1. 0 = swap chains 1 <-> 2, 3 <-> 4...
1 = swap chains 2<->3, 4<->5...}
}
\value{
a list of lists: values, log likelihood etc. of paralle chains after parallel tempering
}
\description{
performs parallel tempering
}
