% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_pt.R
\name{parallel_tempering_find_min}
\alias{parallel_tempering_find_min}
\title{Find chain with minimum posterior values}
\usage{
parallel_tempering_find_min(mcmc_list)
}
\description{
Find chain with minimum posterior values
}
