% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{pbb}
\alias{pbb}
\title{Beta binomial distribution function}
\usage{
pbb(q, N, u, v)
}
\arguments{
\item{q}{vector of quantiles}

\item{N}{the number of trials}

\item{u}{equivalent to alpha in description}

\item{v}{equivalent to beta in description}
}
\value{
the beta binomial distribution function
}
\description{
Beta binomial distribution function
}
\examples{
pbb(10,50,1,1)
}
\seealso{
Other beta_binomial_functions: \code{\link{bb_mean}},
  \code{\link{dbb}}, \code{\link{find_bb_2}},
  \code{\link{find_prior_alpha_beta}},
  \code{\link{hist_rbb}}, \code{\link{qbb}},
  \code{\link{rbb}}
}
\concept{beta_binomial_functions}
