% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{phi_proposal}
\alias{phi_proposal}
\title{DEPRECATED - proposal for phi based on number of infections}
\usage{
phi_proposal(current_pars, infection_history, years, js, alpha, beta,
  n_alive)
}
\description{
DEPRECATED - proposal for phi based on number of infections
}
