% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_other.R
\name{plot_2d_density}
\alias{plot_2d_density}
\title{Plot joint posterior densities for two parameters}
\usage{
plot_2d_density(chain, par1, par2)
}
\arguments{
\item{chain}{the MCMC parameter chain}

\item{par1}{the name of the first parameter column in `chain`}

\item{par2}{the name of the second parameter column in `chain`}
}
\value{
A ggplot object showing the joint posterior density, coloured by MCMC chain.
}
\description{
Plot joint posterior densities for two parameters
}
\seealso{
Other theta_plots:
\code{\link[=plot_mcmc_diagnostics]{plot_mcmc_diagnostics()}},
\code{\link[=plot_posteriors_theta]{plot_posteriors_theta()}},
\code{\link[=plot_samples_distances]{plot_samples_distances()}}
}
\concept{theta_plots}
