% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_antibody_model.R
\name{plot_antibody_dependent_boosting}
\alias{plot_antibody_dependent_boosting}
\title{Antibody dependent boosting relationship}
\usage{
plot_antibody_dependent_boosting(chain, n, titres = seq(0, 8, by = 0.1))
}
\arguments{
\item{chain}{the MCMC chain containing the antibody kinetics parameters}

\item{n}{number of posterior samples to use}

\item{titres}{the antibody levels at which to calculate the boosting relationship}
}
\value{
A matrix of posterior 2.5%, 50%, and 97.5% quantiles for the inferred boost at each value of `titres`.
}
\description{
Calculates the inferred antibody dependent boosting relationship from the MCMC chain
}
