% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_attack_rate_residuals}
\alias{plot_attack_rate_residuals}
\title{Plot attack rate residuals}
\usage{
plot_attack_rate_residuals(infection_histories, titre_dat,
  strain_isolation_times, true_ar, n_alive = NULL, pointsize = 1,
  fatten = 1, pad_chain = TRUE, prior_pars = NULL)
}
\arguments{
\item{titre_dat}{the data frame of titre data}

\item{strain_isolation_times}{vector of the epochs of potential circulation}

\item{true_ar}{data frame of true attack rates, with first column `year` equal to `strain_isolation_times`, and second column `AR` giving the attack rate}

\item{n_alive}{vector with the number of people alive in each year of circulation. Can be left as NULL, and ages will be used to infer this}

\item{pointsize}{Numeric - how big should each point be?}

\item{fatten}{Numeric - fatten parameter for ggplot pointrange}

\item{pad_chain}{if TRUE, fills the infection history data table with entries for non-infection events (ie. 0s). Can be switched to FALSE for speed to get a rough idea of what the attack rates look like.}

\item{prior_pars}{if not NULL, a list of parameters for the attack rate prior, giving the assumed prior_version along with alpha and beta}
}
\value{
a ggplot2 object with the inferred attack rates for each potential epoch of circulation
}
\description{
Plots inferred historical attack rates from the MCMC output on infection histories compared against known attack rates.
}
