% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_attack_rates_monthly}
\alias{plot_attack_rates_monthly}
\title{Plot historical attack rates monthly}
\usage{
plot_attack_rates_monthly(infectionHistories, dat, ages, yearRange,
  ymax = 0.1, buckets = 1)
}
\arguments{
\item{infectionHistories}{the MCMC chain for infection histories}

\item{dat}{the data frame of titre data}

\item{ages}{the data frame of ages for each individual, with columns for individual and DOB (date of birth)}

\item{yearRange}{vector of the first and last epoch of potential circulation}

\item{ymax}{Numeric. the maximum y value to put on the axis}

\item{buckets}{Integer. How many buckets of time is each year split into? ie. 12 for monthly data, 4 for quarterly etc.}
}
\value{
a ggplot2 object with the inferred attack rates for each potential epoch of circulation
}
\description{
Plots inferred historical attack rates from the MCMC output on infection histories for monthly. The main difference compared to the normal attack rate plot is that pointrange plots don't make as much sense at a very fine time resolution.
}
