% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infection_histories.R
\name{plot_cumulative_infection_histories}
\alias{plot_cumulative_infection_histories}
\title{Plot cumulative and per time posterior infection probability densities}
\usage{
plot_cumulative_infection_histories(
  inf_chain,
  burnin = 0,
  indivs,
  real_inf_hist = NULL,
  start_inf = NULL,
  possible_exposure_times,
  nsamp = 100,
  ages = NULL,
  number_col = 1,
  pad_chain = TRUE,
  subset_times = NULL,
  return_data = FALSE
)
}
\arguments{
\item{inf_chain}{the infection history chain}

\item{burnin}{only plot samples where samp_no > burnin}

\item{indivs}{vector of individual ids to plot}

\item{real_inf_hist}{if not NULL, adds lines to the plots showing the known true infection times}

\item{start_inf}{if not NULL, adds lines to show where the MCMC chain started}

\item{possible_exposure_times}{vector of times at which individuals could have been infected}

\item{nsamp}{how many samples from the MCMC chain to take?}

\item{ages}{if not NULL, adds lines to show when an individual was born}

\item{number_col}{how many columns to use for the cumulative infection history plot}

\item{pad_chain}{if TRUE, pads the infection history MCMC chain to have entries for non-infection events}

\item{subset_times}{if not NULL, pass a vector of indices to only take a subset of indices from possible_exposure_times}

\item{return_data}{if TRUE, returns the infection history posterior densities used to generate the plots}
}
\value{
two ggplot objects
}
\description{
For each individual requested, plots the median and 95% quantiles on a) the cumulative number of infections over a lifetime and b) the posterior probability that an infection occured in a given time point
}
\examples{
\dontrun{
data(example_inf_chain)
data(example_antigenic_map)
data(example_inf_hist)
data(example_antibody_data)

ages <- unique(example_antibody_data[,c("individual","birth")])
times <- example_antigenic_map$inf_times
indivs <- 1:10
plot_cumulative_infection_histories(example_inf_chain, 0, indivs, example_inf_hist, NULL, times,
                              ages=ages, number_col=2,pad_chain=FALSE, return_data=TRUE)
}
}
\seealso{
Other infection_history_plots:
\code{\link[=calculate_infection_history_statistics]{calculate_infection_history_statistics()}},
\code{\link[=plot_antibody_data]{plot_antibody_data()}},
\code{\link[=plot_antibody_predictions]{plot_antibody_predictions()}},
\code{\link[=plot_estimated_antibody_model]{plot_estimated_antibody_model()}},
\code{\link[=plot_individual_number_infections]{plot_individual_number_infections()}},
\code{\link[=plot_infection_history_chains_indiv]{plot_infection_history_chains_indiv()}},
\code{\link[=plot_infection_history_chains_time]{plot_infection_history_chains_time()}},
\code{\link[=plot_infection_history_posteriors]{plot_infection_history_posteriors()}},
\code{\link[=plot_model_fits]{plot_model_fits()}},
\code{\link[=plot_total_number_infections]{plot_total_number_infections()}}
}
\concept{infection_history_plots}
