% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infection_histories.R
\name{plot_infection_history_chains_time}
\alias{plot_infection_history_chains_time}
\title{Plot MCMC trace for infections per year}
\usage{
plot_infection_history_chains_time(
  inf_chain,
  burnin = 0,
  times = NULL,
  n_alive = NULL,
  pad_chain = TRUE
)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{serosolver}}}

\item{burnin}{optionally remove all samp_no < burnin from the chain}

\item{times}{vector of integers, if not NULL, only plots a subset of years (where 1 is the first year eg. 1968)}

\item{n_alive}{if not NULL, then divides number of infections per year by number alive to give attack rates rather than total infections}

\item{pad_chain}{if TRUE, pads the infection history MCMC chain to have entries for non-infection events}
}
\value{
a list of two ggplot objects - the MCMC trace and MCMC densities
}
\description{
Plot MCMC trace for infections per year
}
\examples{
\dontrun{
data(example_inf_chain)
data(example_antibody_data)
data(example_antigenic_map)
times <- example_antigenic_map$inf_times
n_alive_group <- get_n_alive_group(example_antibody_data, possible_exposure_times,melt_dat = TRUE)
n_alive_group$j <- possible_exposure_times[n_alive_group$j]
plot_infection_history_chains_time(example_inf_chain, 0, sample(1:length(times),10),n_alive,FALSE)
}
}
\seealso{
\code{\link{plot_infection_history_chains_indiv}}

Other infection_history_plots:
\code{\link[=calculate_infection_history_statistics]{calculate_infection_history_statistics()}},
\code{\link[=plot_antibody_data]{plot_antibody_data()}},
\code{\link[=plot_antibody_predictions]{plot_antibody_predictions()}},
\code{\link[=plot_cumulative_infection_histories]{plot_cumulative_infection_histories()}},
\code{\link[=plot_estimated_antibody_model]{plot_estimated_antibody_model()}},
\code{\link[=plot_individual_number_infections]{plot_individual_number_infections()}},
\code{\link[=plot_infection_history_chains_indiv]{plot_infection_history_chains_indiv()}},
\code{\link[=plot_infection_history_posteriors]{plot_infection_history_posteriors()}},
\code{\link[=plot_model_fits]{plot_model_fits()}},
\code{\link[=plot_total_number_infections]{plot_total_number_infections()}}
}
\concept{infection_history_plots}
