% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_number_infections}
\alias{plot_number_infections}
\title{Plot point range number infections per individual}
\usage{
plot_number_infections(inf_chain)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{run_MCMC}}}
}
\value{
a ggplot object
}
\description{
Plot point range number infections per individual
}
