% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_other.R
\name{plot_samples_distances}
\alias{plot_samples_distances}
\title{Plot time between serum samples}
\usage{
plot_samples_distances(antibody_data)
}
\arguments{
\item{antibody_data}{the data frame of antibody data, including labels for individuals and time sample was taken}
}
\value{
A ggplot2 object showing the distribution of time intervals between samples from the same individual.
}
\description{
Plot time between serum samples
}
\examples{
\dontrun{
data(example_antibody_data)
plot_samples_distances(example_antibody_data)
}
}
\seealso{
Other theta_plots:
\code{\link[=plot_2d_density]{plot_2d_density()}},
\code{\link[=plot_mcmc_diagnostics]{plot_mcmc_diagnostics()}},
\code{\link[=plot_posteriors_theta]{plot_posteriors_theta()}}
}
\concept{theta_plots}
