% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-antigenic-map.R
\name{predict_antigenic_map}
\alias{predict_antigenic_map}
\title{Predict an antigenic map}
\usage{
predict_antigenic_map(antigenic_map, epochs = NULL)
}
\arguments{
\item{antigenic_map}{A dataframe of antigenic coordinates, with columns
\code{strain}, \code{x}, \code{y}, \code{epoch} and \code{epochs_per_year}.
See \code{\link{fonville_map}} for an example.}

\item{epochs}{Epochs to predict for. If \code{NULL}, then will predict for
the observed epochs}
}
\value{
a predicted antigenic map
}
\description{
Fits a linear model to a set of antigenic coordinates in order
to predict antigenic coordinates for infection epochs.
}
