% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{prob_mus}
\alias{prob_mus}
\title{Prior probability of strain specific boosting}
\usage{
prob_mus(mus, pars)
}
\arguments{
\item{mus}{the vector of boosting parameters}

\item{pars}{the vector of other model parameters including mu_mean and mu_sd}
}
\value{
a single log prior probability
}
\description{
Function find the prior probability of a set of boosting parameters assuming that boosting is drawn from a normal distribution (random effects)
}
