% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{protect}
\alias{protect}
\title{Protect function}
\usage{
protect(f)
}
\arguments{
\item{f}{the function to be protected}
}
\value{
the protected function
}
\description{
Wrapper function to protect calls to a function. If the function does not compute correctly, returns -10000000.
}
\keyword{internal}
