% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{qbb}
\alias{qbb}
\title{Beta binomial quantile function}
\usage{
qbb(p, N, u, v)
}
\arguments{
\item{p}{vector of probabilities}

\item{N}{the number of trials}

\item{u}{equivalent to alpha in description}

\item{v}{equivalent to beta in description}
}
\value{
the beta binomial quantile function
}
\description{
Beta binomial quantile function
}
\examples{
qbb(10,50,1,1)
}
\seealso{
Other beta_binomial_functions: \code{\link{bb_mean}},
  \code{\link{dbb}}, \code{\link{find_bb_2}},
  \code{\link{find_prior_alpha_beta}},
  \code{\link{hist_rbb}}, \code{\link{pbb}},
  \code{\link{rbb}}
}
\concept{beta_binomial_functions}
