% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_model_likelihoods.R
\name{r_likelihood_continuous}
\alias{r_likelihood_continuous}
\title{Likelihood function given continuous data (normal)}
\usage{
r_likelihood_continuous(
  expected,
  data,
  theta,
  expected_indices = NULL,
  measurement_shifts = NULL
)
}
\arguments{
\item{expected}{vector of expected antibody levels}

\item{data}{vector of observed continuous measurements}

\item{theta}{named parameter vector containing `min_measurement`, `max_measurement`, and `obs_sd`}

\item{expected_indices}{the indices of the measurement_shifts vector that each predicted antibody levels needs adding to it}

\item{measurement_shifts}{the vector of measurement shifts for each cluster to add to the predicted antibody levels}
}
\value{
a vector of log-likelihood contributions for each observation
}
\description{
Calculates the likelihood of observing a set of continuous and bounded measurements given a corresponding set of predicted antibody levels
}
