% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run MCMC}
\usage{
run_mcmc(parameters, titres, antigenmap, mcmc_settings)
}
\arguments{
\item{parameters}{Parameter table. See \code{\link{par_tab_base}} for an
example.}

\item{titres}{Titre data. See \code{\link{titre_data_example}} for an
example.}

\item{antigenmap}{Antigenic map. See \code{\link{fonville_map}} for an
example.}

\item{mcmc_settings}{MCMC settings. See \code{\link{mcmc_settings_default}}
for an example.}
}
\description{
Run MCMC
}
