% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_PT_fast.R
\name{run_MCMC_pt_fast}
\alias{run_MCMC_pt_fast}
\title{Parallel tempering MCMC}
\usage{
run_MCMC_pt_fast(par_tab, titre_dat, antigenic_map, mcmc_pars = c(),
  start_inf_hist = NULL, filename = "test",
  CREATE_POSTERIOR_FUNC = create_posterior_func_fast,
  CREATE_PRIOR_FUNC = NULL, seed, mu_indices = NULL,
  measurement_indices = NULL, measurement_random_effects = FALSE,
  OPT_TUNING = 0.1, solve_likelihood = TRUE, n_alive = NULL, ...)
}
\arguments{
\item{par_tab}{as in \code{\link{run_MCMC}}, but a list of such data structures, one for each temperature, rather than a single data frame}

\item{titre_dat}{The data frame of titre data to be fitted. Must have columns: group (index of group); individual (integer ID of individual); samples (numeric time of sample taken); virus (numeric time of when the virus was circulating); titre (integer of titre value against the given virus at that sampling time); run (integer giving the repeated number of this titre); DOB (integer giving date of birth matching time units used in model)}

\item{antigenic_map}{A data frame of antigenic x and y coordinates. Must have column names: x_coord; y_coord; inf_years}

\item{mcmc_pars}{Named vector named vector with parameters for the MCMC procedure. See details}

\item{start_inf_hist}{Infection history matrix to start MCMC at. Can be left NULL}

\item{filename}{The full filepath at which the MCMC chain should be saved. "_chain.csv" will be appended to the end of this, so filename should have no file extensions}

\item{CREATE_POSTERIOR_FUNC}{Pointer to posterior function used to calculate a likelihood. This will probably be \code{\link{create_posterior_func}}}

\item{seed}{random seed to set for the MCMC, if desired}

\item{mu_indices}{For random effects on boosting parameter, mu. Vector of indices of length equal to number of circulation times. If random mus are included in the parameter table, this vector specifies which mu to use for each circulation year. For example, if years 1970-1976 have unique boosting, then mu_indices should be c(1,2,3,4,5,6). If every 3 year block shares has a unique boosting parameter, then this should be c(1,1,1,2,2,2)}

\item{measurement_indices}{For measurement bias function. Vector of indices of length equal to number of circulation times. For each year, gives the index of parameters named "rho" that correspond to each time period}

\item{measurement_random_effects}{Boolean indicating if measurement bias is a random effects term. If TRUE adds a component to the posterior calculation that calculates the probability of a set of measurement shifts "rho", given a mean and standard deviation}

\item{OPT_TUNING}{Constant describing the amount of leeway when adapting the proposals steps to reach a desired acceptance rate (ie. does not change step size if within OPT_TUNING of the specified acceptance rate)}

\item{solve_likelihood}{if FALSE, returns only the prior and does not solve the likelihood}

\item{n_alive}{if not NULL, uses this as the number alive for the infection history prior, rather than calculating the number alive based on titre_dat}

\item{...}{Other arguments to pass to CREATE_POSTERIOR_FUNC}
}
\description{
Runs the MCMC chain for the serosolver model, but uses parallel tempering to run a number of chains at different temperatures, 
 swapping their states to improve mixing. NOTE that this can be quite slow!
}
\details{
The `mcmc_pars` argument has the following options in addition to those in \code{\link{run_MCMC}}:
 * temperature (a vector of temperatures to run the MCMC chains at)
 * parallel_tempering_iter how many iterations to run before swapping MCMC chains
}
