% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{save_infHist_to_disk}
\alias{save_infHist_to_disk}
\title{Write given infection history to disk}
\usage{
save_infHist_to_disk(infHist, file, sampno, append = TRUE, colNames = FALSE)
}
\description{
Write given infection history to disk
}

