% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{scaletuning_alt}
\alias{scaletuning_alt}
\title{Scale step sizes}
\usage{
scaletuning_alt(step, target_acceptance_rate_theta, pcur)
}
\arguments{
\item{step}{the current step size}

\item{target_acceptance_rate_theta}{the desired acceptance rate}

\item{pcur}{the current acceptance rate}
}
\value{
the scaled step size
}
\description{
Scales the given step size (between 0 and 1) based on the current acceptance rate to get closed to the desired acceptance rate
}
\seealso{
Other mcmc:
\code{\link[=generate_start_tab]{generate_start_tab()}},
\code{\link[=rm_scale]{rm_scale()}},
\code{\link[=save_infection_history_to_disk]{save_infection_history_to_disk()}},
\code{\link[=scaletuning]{scaletuning()}},
\code{\link[=serosolver]{serosolver()}}
}
\concept{mcmc}
