% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setup_antibody_data_for_posterior_func}
\alias{setup_antibody_data_for_posterior_func}
\title{Setup antibody data indices}
\usage{
setup_antibody_data_for_posterior_func(
  par_tab,
  antibody_data,
  antigenic_map = NULL,
  possible_exposure_times = NULL,
  age_mask = NULL,
  n_alive = NULL,
  verbose = FALSE,
  use_demographic_groups = NULL,
  timevarying_demographics = NULL
)
}
\arguments{
\item{par_tab}{the parameter table controlling information such as bounds, initial values etc. See \code{\link{example_par_tab}}}

\item{antibody_data}{the data frame of data to be fitted. Must have columns: individual (integer ID of individual); sample_time (numeric time of sample taken); biomarker_id (numeric ID of the biomarker); measurement (antibody measurement); and birth (birth time). `biomarker_group`, `repeat_number`, and `population_group` are optional. `population_group` identifies groups with different infection rates or attack rates; if it is absent, all individuals are assigned to group 1. See \code{\link{example_antibody_data}}}

\item{antigenic_map}{(optional) a data frame of antigenic x and y coordinates. Must have column names: x_coord; y_coord; inf_times. See \code{\link{example_antigenic_map}}}

\item{possible_exposure_times}{(optional) if no antigenic map is specified, this argument gives the vector of times at which individuals can be infected}

\item{age_mask}{see \code{\link{create_age_mask}} - a vector with one entry for each individual specifying the first epoch of circulation in which an individual could have been exposed}

\item{n_alive}{if not NULL, uses this as the number alive in a given year rather than calculating from the ages. This is needed if the number of alive individuals is known, but individual birth dates are not}

\item{verbose}{if TRUE, brings warning messages}

\item{use_demographic_groups}{vector of variable names in `antibody_data` which should form the stratification for the antibody kinetics model}

\item{timevarying_demographics}{if not NULL, then calculates an individual's demographic group over the entire time period of the simulation rather than assuming fixed demographics}
}
\value{
a list of pre-processed data vectors, indices, masks, demographic groups, and numbers alive used by the posterior function.
}
\description{
Sets up a large list of pre-indexing and pre-processing to speed up the model solving during MCMC fitting.
Note that this should be `antibody_data` after subsetting to only `repeat_number == 1`, as we will figure out elsewhere which solves to use as repeats.
}
\seealso{
\code{\link{create_posterior_func}}
}
\keyword{internal}
