% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{setup_infection_histories_immunity}
\alias{setup_infection_histories_immunity}
\title{Initial infection history immunity}
\usage{
setup_infection_histories_immunity(titre_dat, strain_isolation_times,
  alpha = 1, beta = 1, titre_prob_default = 0.8)
}
\arguments{
\item{titre_dat}{the matrix of titres data with columns for individual, sample, and titre}

\item{strain_isolation_times}{vector of real times for all strains}

\item{alpha}{alpha parameter for beta distribution to sample from}

\item{beta}{beta parameter for beta distribution to sample from}
}
\value{
an infection history matrix and exposure history matrix
}
\description{
Sets up an initial random infection history, where start is sampled from a prior on the total number of infections across all individuals and years
}
\seealso{
Other setup_infection_histories: \code{\link{setup_infection_histories_new_2}},
  \code{\link{setup_infection_histories_old}},
  \code{\link{setup_infection_histories_total}}
}
\concept{setup_infection_histories}
