% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{setup_infection_histories_new}
\alias{setup_infection_histories_new}
\title{Propose initial infection histories}
\usage{
setup_infection_histories_new(data, ages, strainIsolationTimes, space = 5,
  titre_cutoff = 2)
}
\arguments{
\item{data}{the matrix of titres data with columns for individual, sample, and titre}

\item{ages}{a matrix of ages for each individual. Columns should be 1) the individual's ID; 2) the individual's date of birth (called DOB) in the same time resolution of the model (ie. 1991 for years, 1991*12 for months etc)}

\item{strainIsolationTimes}{vector of real times for all strains}

\item{space}{how many epochs must separate proposed infections}

\item{titre_cutoff}{specifies how high the titre must be to imply an infection}
}
\value{
an nxm matrix of infection histories containing 1s and 0s, where n is the number of individuals and m is the number of potential infecting strains
}
\description{
Given a matrix of titre data, proposes plausible initial infection histories from which to begin MCMC sampling.
The idea is to move along time in the context of antigenic drift and look at an individual's titre against each strain. Where titres are raised, we suggest an infection. However, to avoid suggesting multiple infections for regions of high antigenic similarity, we place a necessary gap (defined by `space`) between proposed infection times.
NOTE - MIGHT NEED TO UPDATE THIS FOR GROUPS
}

