% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{setup_infection_histories_total}
\alias{setup_infection_histories_total}
\title{Initial infection history prior on total}
\usage{
setup_infection_histories_total(
  titre_dat,
  strain_isolation_times,
  alpha = 1,
  beta = 1
)
}
\arguments{
\item{titre_dat}{the matrix of titres data with columns for individual, sample, and titre}

\item{strain_isolation_times}{vector of real times for all strains}

\item{alpha}{alpha parameter for beta distribution to sample from}

\item{beta}{beta parameter for beta distribution to sample from}
}
\value{
an infection history matrix
}
\description{
Sets up an initial random infection history, where start is sampled from a prior on the total number of infections across all individuals and years
}
\examples{
data(example_titre_dat)
data(example_antigenic_map)
start_inf <- setup_infection_histories_total(example_titre_dat, example_antigenic_map$inf_times, 2,10)
}
\seealso{
Other setup_infection_histories: 
\code{\link{setup_infection_histories_old}()},
\code{\link{setup_infection_histories_titre}()},
\code{\link{setup_infection_histories}()}
}
\concept{setup_infection_histories}
