% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.R
\name{setup_stratification_table}
\alias{setup_stratification_table}
\title{Create indexing table for parameter stratifications}
\usage{
setup_stratification_table(par_tab, unique_demographic_combinations)
}
\arguments{
\item{par_tab}{the parameter table, including a column called stratification which is NA if no stratification is requested, or a character value matching an entry in antibody_data or timevarying_demographics}

\item{unique_demographic_combinations}{table of unique stratification combinations. Columns give stratification variable, rows give each unique combinations}
}
\value{
a list with two entries: 1) a list of matrices with entries for each stratification level, with number of columns equal to the number of parameters, and number of rows equal to the number of stratification levels. Each entry is the index of the scale parameter to use for that stratification level and parameter; 2) a vector of scale parameters
}
\description{
Creates an indexing table to tell serosolver which scale parameter to add to which model parameter.
}
\seealso{
Other stratification:
\code{\link[=add_scale_pars]{add_scale_pars()}},
\code{\link[=create_demographic_table]{create_demographic_table()}},
\code{\link[=get_demographic_groups]{get_demographic_groups()}}
}
\concept{stratification}
\keyword{internal}
