% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setup_titredat_for_posterior_func}
\alias{setup_titredat_for_posterior_func}
\title{Setup titre data indices}
\usage{
setup_titredat_for_posterior_func(titre_dat, antigenic_map,
  age_mask = NULL, n_alive = NULL)
}
\value{
a very long list. See source code directly.
}
\description{
Sets up a large list of pre-indexing and pre-processing to speed up the model solving during MCMC fitting.
}
\seealso{
\code{\link{create_posterior_func}}
}
