% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_attack_rates}
\alias{simulate_attack_rates}
\title{Simulate attack rates}
\usage{
simulate_attack_rates(
  infection_years,
  mean_par = 0.15,
  sd_par = 0.5,
  large_first_year = FALSE,
  big_year_mean = 0.5,
  n_groups = 1
)
}
\arguments{
\item{infection_years}{the number of infection years}

\item{mean_par}{the mean of the log normal}

\item{sd_par}{the sd of the log normal}

\item{large_first_year}{simulate an extra large attach rate in the first year?}

\item{big_year_mean}{if large first year, what mean to use?}

\item{n_groups}{defaults to 1, otherwise gives an attack rate vector for each group}
}
\value{
a matrix of attack rates for each group
}
\description{
Given a number of possible infection years, simulates attack rates from a log normal distribution with specified mean and standard deviation.
}
\seealso{
Other simulation_functions:
\code{\link[=simulate_data]{simulate_data()}},
\code{\link[=simulate_infection_histories]{simulate_infection_histories()}}
}
\concept{simulation_functions}
