% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cross_sectional.R
\name{simulate_cross_sectional}
\alias{simulate_cross_sectional}
\title{Simulates multiple cohorts in a cross-sectional framework.}
\usage{
simulate_cross_sectional(
  parTab,
  n_indiv,
  buckets = 12,
  strainIsolationTimes,
  samplingTimes,
  antigenicMap,
  ageMin = 5,
  ageMax = 80,
  attackRates,
  group = 1,
  sampleSensoring = 0,
  titreSensoring = 0
)
}
\arguments{
\item{parTab}{the full parameter table controlling parameter ranges and values}

\item{n_indiv}{number of individuals to simulate}

\item{strainIsolationTimes}{vector of strain circulation times}

\item{samplingTimes}{possible sampling times for the individuals}

\item{antigenicMap}{the raw antigenic map with colnames x_coord, y_coord and inf_times}

\item{ageMin}{minimum age to simulate}

\item{ageMax}{maximum age to simulate}

\item{attackRates}{a vector of attackRates to be used in the simulation}

\item{group}{which group index to give this simulated data}

\item{titreSensoring}{what proportion of titres are randomly missing?}
}
\value{
a list with: 1) the data frame of titre data as returned by \code{\link{simulate_group}}; 2) a matrix of infection histories as returned by \code{\link{simulate_infection_histories}}; 3) a vector of ages
}
\description{
Simulates multiple cohorts in a cross-sectional framework.
}
