% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_individual_faster}
\alias{simulate_individual_faster}
\title{Simulate individual data quickly}
\usage{
simulate_individual_faster(
  theta,
  unique_theta_indices = NULL,
  infection_history,
  antigenic_map_long,
  antigenic_map_short,
  antigenic_distances = NULL,
  unique_obs_types = seq_len(ncol(antigenic_map_long)),
  sampling_times,
  possible_exposure_times,
  measured_biomarker_ids,
  measurement_bias = NULL,
  measurement_indices = NULL,
  obs_dist = NULL,
  repeats = 1,
  DOB = NULL
)
}
\arguments{
\item{theta}{the named parameter vector}

\item{infection_history}{the vector of 1s and 0s giving presence/absence of infections}

\item{antigenic_map_long}{the long term antigenic cross reactivity map generated from \code{\link{create_cross_reactivity_vector}}}

\item{antigenic_map_short}{the short term antigenic cross reactivity map generated from \code{\link{create_cross_reactivity_vector}}}

\item{antigenic_distances}{(optional) same dimensions as antigenic_map_long and antigenic_map_short, but gives the raw euclidean antigenic distances}

\item{sampling_times}{vector of times at which blood samples were taken}

\item{possible_exposure_times}{(optional) If no antigenic map is specified, this argument gives the vector of times at which individuals can be infected}

\item{measured_biomarker_ids}{vector of which strains had titres measured in `possible_exposure_times`}

\item{measurement_indices}{default NULL, optional vector giving the index of `measurement_bias` that each strain uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{obs_dist}{if not NULL, then a vector of data types to use for each obs_type}

\item{repeats}{number of repeat observations for each year}
}
\value{
a data frame with columns samples, virus and titre of simulated data
}
\description{
FOR USERS: USE \code{\link{simulate_individual}}. This function does the same thing, but with a few short cuts for speed. Simulates a full set of titre data for an individual with known theta and infection_history.
}
\seealso{
Other simulation_functions: 
\code{\link{simulate_attack_rates}()},
\code{\link{simulate_data}()},
\code{\link{simulate_group}()},
\code{\link{simulate_individual}()},
\code{\link{simulate_infection_histories}()}
}
\concept{simulation_functions}
