% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_infection_histories_titre}
\alias{simulate_infection_histories_titre}
\title{Simulate infection histories}
\usage{
simulate_infection_histories_titre(pars, p_inf, strain_isolation_times,
  ages, antigenic_map, mus = NULL, mu_indices = NULL,
  measurement_bias = NULL, measurement_indices = NULL,
  separate_exosures = FALSE)
}
\arguments{
\item{pars}{vector of model kinetics parameters}

\item{p_inf}{a vector of attack rates (infection probabilities) for each year}

\item{strain_isolation_times}{the vector of possible infection times}

\item{ages}{a vector of ages for each individual}

\item{mus}{vector of boosting terms for strain dependent boosting}

\item{mu_indices}{default NULL, optional vector giving the index of `mus` that each strain uses the boosting parameter from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{measurement_bias}{vector of strain-specific measurement shifts}

\item{measurement_indices}{default NULL, optional vector giving the index of `measurement_bias` that each strain uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{separate_exposures}{if TRUE, returns the exposure and infection histories as separate matrices}
}
\value{
a list with a matrix of infection histories for each individual in ages and the true attack rate for each epoch
}
\description{
Given a vector of infection probabilities and potential infection times, simulates infections for each element of ages (ie. each element is an individual age. Only adds infections for alive individuals)
}
