% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth-antigenic-map.R
\name{smooth_antigenic_map}
\alias{smooth_antigenic_map}
\title{Smooth antigenic map}
\usage{
smooth_antigenic_map(antigenic_map, epochs_per_year = 1, spar = 0.3)
}
\arguments{
\item{antigenic_map}{A dataframe of antigenic coordinates, with columns
\code{strain}, \code{x}, \code{y} and \code{year}. See
\code{\link{fonville_map}} for an example.}

\item{epochs_per_year}{the number of epochs per year. 1 means that each year
has 1 strain; 12 means that each year has 12 strains (monthly resolution).}

\item{spar}{to be passed to \code{\link[stats]{smooth.spline}}}
}
\value{
a fitted antigenic map
}
\description{
Fits a smoothing spline through a set of antigenic coordinates, and uses this
to predict antigenic coordinates for all potential infection time points
}
