% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{solve_model_individual_original}
\alias{solve_model_individual_original}
\title{Solves the original model as in the original implementation}
\usage{
solve_model_individual_original(theta_star, sample, antigenic.map.in, years,
  infHist)
}
\description{
Solves the original model as in the original implementation
}

