% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{strain_years}
\alias{strain_years}
\title{Strain years}
\format{A data frame with the following variables:

  \describe{

  \item{strain}{Virus strain}

  \item{year}{The year it circulated}

  }}
\usage{
strain_years
}
\description{
Strain years
}
\keyword{datasets}
