% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_buckets}
\alias{sum_buckets}
\title{Sums a vector based on bucket sizes}
\usage{
sum_buckets(a, buckets)
}
\arguments{
\item{a}{the vector to be bucketed}

\item{buckets}{the vector of bucket sizes to sum a over}
}
\value{
the vector of summed a
}
\description{
Given a vector (a) and another vector of bucket sizes, returns the summed vector (a)
}
\keyword{internal}
