% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_likelihoods}
\alias{sum_likelihoods}
\title{Sum likelihoods into buckets}
\usage{
sum_likelihoods(liks, indices, n_indivs)
}
\arguments{
\item{liks}{NumericVector of likelihoods}

\item{indices}{IntegerVector of indices of same length as liks, where the max value of this should be the same as n_indivs - 1}

\item{n_indivs}{int, number of individuals to generate bucketed likelihoods for}
}
\description{
Given a large vector of likelihood values and a vector, indices, of length n_indivs, sums the likelihoods to give one value per individual, where indices indicates which individual each index of liks corresponds to.
}
\keyword{internal}
