% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_distributions.R
\name{summarize_run_lengths}
\alias{summarize_run_lengths}
\title{Summarize runs of infections from posterior}
\usage{
summarize_run_lengths(inf_chain)
}
\arguments{
\item{inf_chain}{data table of the infection histories posterior or tibble from \code{\link{identify_run_lengths}}}
}
\value{
a tibble summarizing the infection run lengths for each individual and distinct infection event
}
\description{
Either takes the output of \code{\link{identify_run_lengths}} and produces summary statistics giving the posterior median and 95% quantiles for the length of each infection run
}
\examples{
\dontrun{
summarize_run_lengths(inf_chain)
summarize_run_lengths(identify_run_lengths(inf_chain))
}
}
