% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titre_data_example}
\alias{titre_data_example}
\title{Example titre data}
\format{A data frame with the following variables:

  \describe{

  \item{individual}{Subject index}

  \item{samples}{Sample time index}

  \item{virus}{Virus circulation time index}

  \item{titre}{Log titre}

  \item{run}{Measurement index}

  \item{DOB}{Date of birth time index}

  \item{group}{Group index}

  }}
\usage{
titre_data_example
}
\description{
\code{titre_dat} from the quick start vignette
}
\keyword{datasets}
