% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{titre_data_fast}
\alias{titre_data_fast}
\title{Overall model function, fast implementation}
\usage{
titre_data_fast(theta, infection_history_mat, circulation_times,
  circulation_times_indices, sample_times, rows_per_indiv_in_samples,
  cum_nrows_per_individual_in_data, nrows_per_blood_sample,
  measurement_strain_indices, antigenic_map_long, antigenic_map_short, mus,
  boosting_vec_indices)
}
\arguments{
\item{mus}{NumericVector, if length is greater than one, assumes that strain-specific boosting is used rather than a single boosting parameter}

\item{boosting_vec_indices}{IntegerVector, same length as circulation_times, giving the index in the vector \code{mus} that each entry should use as its boosting parameter.}
}
\value{
NumericVector of predicted titres for each entry in measurement_strain_indices
}
\description{
See documentation for \code{\link{titre_data_group}}, as the interface is almost identical
}
\concept{titre_model}
