% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{titre_data_fast_individual_strain_dependent}
\alias{titre_data_fast_individual_strain_dependent}
\title{Base boosting fast}
\usage{
titre_data_fast_individual_strain_dependent(
  predicted_titres,
  theta,
  infection_info,
  vaccination_info,
  setup_data,
  indexing,
  antigenic_maps,
  other_pars
)
}
\description{
A fast implementation of the basic boosting function, giving predicted titres for a number of samples for one individual. Note that this version attempts to minimise memory allocations.
}
\seealso{
\code{\link{titre_data_fast}}

Other boosting_functions: 
\code{\link{titre_data_fast_individual_base}()},
\code{\link{titre_data_fast_individual_titredep}()},
\code{\link{titre_data_fast_individual_wane2}()}
}
\concept{boosting_functions}
