% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{titre_data_fast_individual_wane2}
\alias{titre_data_fast_individual_wane2}
\title{Alternative waning fast}
\usage{
titre_data_fast_individual_wane2(
  predicted_titres,
  theta,
  infection_info,
  vaccination_info,
  setup_data,
  indexing,
  antigenic_maps,
  other_pars
)
}
\description{
A fast implementation of the alternative waning function, giving predicted titres for a number of samples for one individual. Note that this version attempts to minimise memory allocations.
}
\seealso{
\code{\link{titre_data_fast}}

Other boosting_functions: 
\code{\link{titre_data_fast_individual_base}()},
\code{\link{titre_data_fast_individual_strain_dependent}()},
\code{\link{titre_data_fast_individual_titredep}()}
}
\concept{boosting_functions}
