% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{titre_data_group}
\alias{titre_data_group}
\title{Model function overall}
\usage{
titre_data_group(theta, infectionHistories, circulationTimes,
  circulationMapIndices, samplingTimes, indicesTitreDataSample,
  indicesTitreDataOverall, indicesSamples, measuredMapIndices,
  antigenicMapLong, antigenicMapShort, DOBs, additional_arguments = NULL)
}
\arguments{
\item{theta}{NumericVector, the named vector of model parameters}

\item{infectionHistories}{IntegerMatrix, the matrix of 1s and 0s showing presence/absence of infection for each possible time for each individual.}

\item{circulationTimes}{NumericVector, the actual times of circulation that the infection history vector corresponds to}

\item{circulationMapIndices}{IntegerVector, which entry in the melted antigenic map that these infection times correspond to}

\item{samplingTimes}{NumericVector, the times that each blood sample was taken}

\item{indicesTitreDataSample}{IntegerVector, How many rows in titre data correspond to each individual, sample and repeat?}

\item{indicesTitreDataOverall}{IntegerVector, How many rows in the titre data correspond to each individual?}

\item{indicesSamples}{IntegerVector, Split the sample times and runs for each individual}

\item{measuredMapIndices}{IntegerVector, the indices of all measured strains in the melted antigenic map}

\item{antigenicMapLong}{NumericVector, the collapsed cross reactivity map for long term boosting, after multiplying by sigma1}

\item{antigenicMapShort}{NumericVector, the collapsed cross reactivity map for short term boosting, after multiplying by sigma2}

\item{additional_arguments, }{Nullable<List> currently not used, but the idea is to use thsi object to pass more flexible additional arguments to the bottom of the call stack}

\item{DOB}{NumericVector, the date of birth of all individuals. Currently not used.}
}
\value{
NumericVector of predicted titres for each entry in measuredMapIndices
}
\description{
The main model solving function for a single individual for a vector of sampling timesthe entire titreDat structure
}
