% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{titre_dependent_boosting_plot}
\alias{titre_dependent_boosting_plot}
\title{Titre dependent boosting relationship}
\usage{
titre_dependent_boosting_plot(chain, n, titres = seq(0, 8, by = 0.1))
}
\arguments{
\item{chain}{the MCMC chain}

\item{n}{number of samples to take}

\item{titres}{the vector of titres to calculate boosting values at}
}
\value{
a data frame of quantiles for the inferred boost from different titre levels
}
\description{
Calculates the inferred titre dependent boosting relationship from the MCMC chain
}
