% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{titre_protection}
\alias{titre_protection}
\title{Calculate titre protection}
\usage{
titre_protection(titre, alpha, beta1)
}
\arguments{
\item{titre}{the log HI titre of the individual (base 2)}

\item{alpha}{parameter 1}
}
\value{
a single probability of infection
}
\description{
Calculates the reduced risk of infection for a given HI titre at the time of exposure as per Coudeville et al. 2010
}
