% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{toUnitScale}
\alias{toUnitScale}
\title{Convert to unit scale}
\usage{
toUnitScale(x, min, max)
}
\arguments{
\item{x}{value to convert}

\item{min}{lower bound of the original scale}

\item{max}{upper bound of the original scale}
}
\value{
the value on the 0--1 scale
}
\description{
Converts a value from a bounded scale to the 0--1 scale.
}
\keyword{internal}
