% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_theta.R
\name{univ_proposal}
\alias{univ_proposal}
\title{MCMC proposal function}
\usage{
univ_proposal(
  values,
  lower_bounds,
  upper_bounds,
  steps,
  index,
  gaussian_proposal = TRUE
)
}
\arguments{
\item{values}{a vector of the parameters to be explored}

\item{lower_bounds}{a vector of the low allowable bounds for the proposal}

\item{upper_bounds}{a vector of the upper allowable bounds for the proposal}

\item{steps}{a vector of step sizes for the proposal}

\item{index}{numeric value for the index of the parameter to be moved from the param table and vector}

\item{gaussian_proposal}{retained for compatibility with the current interface}
}
\value{
the parameter vector after step
}
\description{
Proposal function for MCMC random walk, taking random steps of a given size.
}
\seealso{
Other proposals:
\code{\link[=inf_hist_swap]{inf_hist_swap()}},
\code{\link[=inf_hist_swap_phi]{inf_hist_swap_phi()}},
\code{\link[=infection_history_symmetric]{infection_history_symmetric()}},
\code{\link[=mvr_proposal]{mvr_proposal()}}
}
\concept{proposals}
\keyword{internal}
