% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unregister_dopar}
\alias{unregister_dopar}
\title{Unregister parallel backgrounds forcefully}
\usage{
unregister_dopar()
}
\description{
When the serosolver function uses a parallel backend but is not closed correctly, this can confuse `dopar` if the function is called again. This function tidies the parallel backend and corrects the error.
}
\keyword{internal}
