% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wane_function}
\alias{wane_function}
\title{Function to calculate non-linear waning
 All additional parameters for the function are declared here}
\usage{
wane_function(theta, time_infected, wane)
}
\arguments{
\item{theta}{NumericVector, the named vector of model parameters}

\item{time_infected}{double the time infected (samplingTime - circulation_time)}
}
\value{
value of waning parameter based on time since infected
}
\description{
Function to calculate non-linear waning
 All additional parameters for the function are declared here
}
