% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_measurement_shifts}
\alias{add_measurement_shifts}
\title{Add measurement shifts to predictions}
\usage{
add_measurement_shifts(
  predicted_antibody_levels,
  to_add,
  start_index_in_data,
  end_index_in_data
)
}
\arguments{
\item{predicted_antibody_levels}{NumericVector, the predicted antibody levels. Note that this vector will be changed!}

\item{to_add}{NumericVector the vector of all measurement shifts to apply}

\item{start_index_in_data}{int the first index of to_add and predicted_antibody_levels to combine}

\item{end_index_in_data}{int the end index of to_add and predicted_antibody_levels to combine}
}
\value{
nothing
}
\description{
Adds observation error shifts to predicted antibody levels.
}
