% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_measurement_shifts}
\alias{add_measurement_shifts}
\title{Add measurement shifts to predictions}
\usage{
add_measurement_shifts(predicted_titres, to_add, start_index_in_data,
  end_index_in_data)
}
\arguments{
\item{to_add}{NumericVector the vector of all measurement shifts to apply}

\item{start_index_in_data}{int the first index of to_add and predicted_titres to combine}

\item{end_index_in_data}{int the end index of to_add and predicted_titres to combine}

\item{predicted_titre}{NumericVector, the predicted titres. Note that this vector will be changed!}
}
\value{
nothing
}
\description{
Adds observation error shifts to predicted antibody titres.
}
