% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{add_noise}
\alias{add_noise}
\title{Add noise}
\usage{
add_noise(y, theta, measurement_bias = NULL, indices = NULL, data_type = 1)
}
\arguments{
\item{y}{the underlying antibody level}

\item{theta}{a named parameter vector containing `min_measurement`, `max_measurement`, and the relevant error parameters}

\item{measurement_bias}{optional vector of measurement shifts}

\item{indices}{optional integer vector selecting the shift to apply to each value of `y`}

\item{data_type}{integer, currently accepting 1, 2, or 3. Set to 1 for discrete, bounded data; 2 for continuous, bounded data; or 3 for continuous data in which true negatives can produce false-positive measurements.}
}
\value{
A vector of noisy, bounded antibody measurements.
}
\description{
Adds observation noise to antibody data and truncates the result to the specified measurement limits.
}
\examples{
\dontrun{
## ... example in simulate_individual
pars <- c("obs_sd"=1)
y <- runif(100)
noisy_y <- add_noise(y, pars)
}
}
