% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{add_noise}
\alias{add_noise}
\title{Add noise}
\usage{
add_noise(y, theta, measurement_bias = NULL, indices = NULL)
}
\arguments{
\item{y}{the titre}

\item{theta}{a vector with MAX_TITRE and error parameters}
}
\value{
a noisy titre
}
\description{
Adds truncated noise to titre data
}
