% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.R
\name{add_stratifying_variables}
\alias{add_stratifying_variables}
\title{Align stratification levels for serosolver infection history and antibody kinetics models}
\usage{
add_stratifying_variables(
  antibody_data,
  timevarying_demographics = NULL,
  par_tab,
  use_demographic_groups = NULL
)
}
\description{
Align stratification levels for serosolver infection history and antibody kinetics models
}
