% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.R
\name{align_antibody_demographic_dat}
\alias{align_antibody_demographic_dat}
\title{Internal function -- merges antibody_data and demographics such that antibody_data has correct variables}
\usage{
align_antibody_demographic_dat(
  antibody_data,
  demographics = NULL,
  verbose = FALSE
)
}
\description{
Internal function -- merges antibody_data and demographics such that antibody_data has correct variables
}
