% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{bb_mean}
\alias{bb_mean}
\title{Beta binomial mean}
\usage{
bb_mean(n, alpha, beta)
}
\arguments{
\item{n}{the number of trials}

\item{alpha}{alpha parameter}

\item{beta}{beta parameter}
}
\value{
calculates the mean of the beta binomial function
}
\description{
Beta binomial mean
}
\examples{
bb_mean(100, 1, 1)
}
\seealso{
Other beta_binomial_functions: \code{\link{dbb}},
  \code{\link{find_bb_2}},
  \code{\link{find_prior_alpha_beta}},
  \code{\link{hist_rbb}}, \code{\link{pbb}},
  \code{\link{qbb}}, \code{\link{rbb}}
}
\concept{beta_binomial_functions}
