% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{calc_b}
\alias{calc_b}
\title{Calculate the second Beta shape parameter from a mode and certainty value}
\usage{
calc_b(mode1, k)
}
\arguments{
\item{mode1}{desired mode}

\item{k}{certainty value}
}
\value{
the second Beta shape parameter
}
\description{
Calculate the second Beta shape parameter from a mode and certainty value
}
