% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{calc_phi_probs_indiv}
\alias{calc_phi_probs_indiv}
\title{Calculate FOI log probability vector}
\usage{
calc_phi_probs_indiv(phis, infection_history, age_mask, strain_mask)
}
\arguments{
\item{phis}{a vector of FOIs}

\item{infection_history}{the matrix of infection histories}

\item{age_mask}{the age mask vector as returned by \code{\link{create_age_mask}}}
}
\value{
a vector of log probabilities for each individual
}
\description{
Given a vector of FOIs for all circulating years, a matrix of infection histories and the vector specifying if individuals were alive or not, returns the log probability of the FOIs given the infection histories.
}
\seealso{
Other priors: \code{\link{calc_phi_probs_spline}},
  \code{\link{calc_phi_probs}},
  \code{\link{create_prior_mu}},
  \code{\link{create_prob_shifts}},
  \code{\link{find_beta_prior_mode}},
  \code{\link{find_beta_prior_with_mean_var}},
  \code{\link{find_beta_prior_with_mean}},
  \code{\link{fit_beta_prior}},
  \code{\link{fit_normal_prior}},
  \code{\link{inf_mat_prior}},
  \code{\link{infection_history_prior}},
  \code{\link{prob_mus}}, \code{\link{prob_shifts}}
}
\concept{priors}
