% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{calculate_infection_history_statistics}
\alias{calculate_infection_history_statistics}
\title{Get posterior information infection histories}
\usage{
calculate_infection_history_statistics(inf_chain, burnin = 0,
  years = NULL, n_alive = NULL, known_ar = NULL,
  known_infection_history = NULL, pad_chain = TRUE)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{run_MCMC}}}

\item{burnin}{if not already discarded, discard burn in from chain (takes rows where sampno > burnin)}

\item{years}{vector of the epochs of potential circulation}

\item{n_alive}{vector with the number of people alive in each year of circulation.}

\item{known_ar}{data frame of known attack rates, if known.}

\item{known_infection_history}{data frame of known infection histories.}

\item{pad_chain}{if TRUE, pads the infection history MCMC chain with non-infection events}
}
\value{
a list of data frames with summary statistics
}
\description{
Finds the median, mean and 95% credible intervals for the attack rates and total number of infections per individual
}
