% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_attack_rates}
\alias{check_attack_rates}
\title{Checks the attack_rates supplied in simulate_data}
\usage{
check_attack_rates(attack_rates, strain_isolation_times)
}
\arguments{
\item{attack_rates}{a vector of attack_rates to be used in the simulation}

\item{strain_isolation_times}{vector of strain ciruclation times}
}
\value{
nothing at the moment
}
\description{
Checks the attack_rates supplied in simulate_data
}
