% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_attackRates}
\alias{check_attackRates}
\title{Checks the attackRates supplied in simulate_data}
\usage{
check_attackRates(attackRates, strainIsolationTimes)
}
\arguments{
\item{attackRates}{a vector of attackRates to be used in the simulation}

\item{strainIsolationTimes}{vector of strain ciruclation times}
}
\value{
nothing at the moment
}
\description{
Checks the attackRates supplied in simulate_data
}
