% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_data}
\alias{check_data}
\title{Checks the entries of data used in run_MCMC}
\usage{
check_data(data)
}
\arguments{
\item{data}{the data frame of data to be fitted. Must have columns: group (index of group); individual (integer ID of individual); samples (numeric time of sample taken); virus (numeric time of when the virus was circulating); titre (integer of titre value against the given virus at that sampling time)}
}
\value{
nothing at the moment
}
\description{
Checks the entries of data used in run_MCMC
}
\examples{
data(example_titre_dat)
check_data(example_titre_dat)
}
\seealso{
Other check_inputs: \code{\link{check_attack_rates}},
  \code{\link{check_par_tab}},
  \code{\link{check_proposals}}
}
\concept{check_inputs}
