% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inf_hist}
\alias{check_inf_hist}
\title{Check infection history matrix}
\usage{
check_inf_hist(titre_dat, strain_isolation_times, inf_hist)
}
\arguments{
\item{titre_dat}{the data frame of titre data}

\item{strain_isolation_times}{the circulation times of possible infecting strains eg. 1968:2015}

\item{inf_hist}{the infection history matrix, with nrows = number indivs and ncols = length(strain_isolation_times)}
}
\value{
a single boolean value, FALSE if the check passes, otherwise throws an error
}
\description{
Checks that the infection history matrix is allowable given the birth dates and sampling times of the data
}
\concept{check_inputs
data(example_titre_dat)
data(example_inf_hist)
times <- 1968:2015
check_inf_hist(example_titre_dat, times, example_inf_hist)}
