% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_par_tab}
\alias{check_par_tab}
\title{Check par_tab for simulate_data}
\usage{
check_par_tab(par_tab, mcmc = FALSE, version = NULL)
}
\arguments{
\item{par_tab}{the parameter table controlling information such as bounds, initial values etc}

\item{mcmc}{logical, if TRUE then checks are performed for the MCMC algorithm. Use FALSE when simulating data}

\item{version}{which version of the posterior function is being used? See \code{\link{create_posterior_func}}}
}
\value{
nothing at the moment
}
\description{
Checks the entries of par_tab used in simulate_data
}
