% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{create_age_mask}
\alias{create_age_mask}
\title{Create age mask}
\usage{
create_age_mask(DOBs, strainIsolationTimes)
}
\description{
Converts a data frame of individual ages to give the index of the first infection that individual could have had
}
