% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.R
\name{create_demographic_table}
\alias{create_demographic_table}
\title{Create a demographics data frame of unique stratification levels}
\usage{
create_demographic_table(antibody_data, par_tab)
}
\arguments{
\item{antibody_data}{the antibody data data frame, see \code{\link{example_antibody_data}}.}

\item{par_tab}{the parameter table, including a column called stratification which is NA if no stratification is requested, for with a character value matching an entry in antibody_data}
}
\value{
a data frame of unique stratification level combinations. Each column is a stratification variable, and each row is a unique combination of levels
}
\description{
Takes all stratifications requested in par_tab and checks antibody_data for all unique levels of that stratification. If no stratifications are requested, then returns a data frame with a single row and a column called "all" with value 0.
}
