% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{create_posterior_func}
\alias{create_posterior_func}
\title{Posterior function pointer - FAST VERSION}
\usage{
create_posterior_func(par_tab, titre_dat, antigenic_map, version = 1,
  solve_likelihood = TRUE, age_mask = NULL,
  measurement_indices_by_time = NULL, mu_indices = NULL,
  n_alive = NULL, function_type = 1, ...)
}
\value{
a single function pointer that takes only pars and infection_histories as unnamed arguments. This function goes on to return a vector of posterior values for each individual
}
\description{
Takes all of the input data/parameters and returns a function pointer. This function finds the posterior for a given set of input parameters (theta) and infection histories without needing to pass the data set back and forth.
}
