% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{create_posterior_func}
\alias{create_posterior_func}
\title{Posterior function pointer}
\usage{
create_posterior_func(parTab, data, antigenicMap, version = 1,
  ageMask = NULL, measurement_indices_by_time = NULL, ...)
}
\arguments{
\item{parTab}{the parameter table controlling information such as bounds, initial values etc}

\item{data}{the data frame of data to be fitted. Must have columns: group (index of group); individual (integer ID of individual); samples (numeric time of sample taken); virus (numeric time of when the virus was circulating); titre (integer of titre value against the given virus at that sampling time)}

\item{antigenicMap}{a data frame of antigenic x and y coordinates. Must have column names: x_coord; y_coord; inf_years}

\item{version}{integer specifying which version of the posterior function to use. This is mainly used for turning off or on the likelihood for debugging the implicit prior. The versions are as follows: 1) Likelihood turned on and no explicit prior. Any prior is implicit from the proposal distribution; 2) No likelihood and no explicit prior; 3) Explicit prior term (as specified by PRIOR_FUNC) and no likelihood. Any implicit prior from the proposal will also be included; 4) Version of the posterior distribution where we infer an explicit force of infection for each epoch; 5) Explicit likelihood and prior - any implicit prior from the proposal will also be included, so proposals should be symmetric; 6) just the negative log likelihood, but function only expects vector of pars (not infectionHistories), which must be specified with ...; 7+) Returns the predicted titres (model solving function); 99) Returns a gibbs-sampled infection history matrix}

\item{ageMask}{see \code{\link{create_age_mask}} - a vector with one entry for each individual specifying the first epoch of circulation in which an individual could have been exposed}

\item{...}{other arguments to pass to the posterior solving function}

\item{PRIOR_FUNC}{user function of prior for model parameters. Should take parameter values only}

\item{temp}{temperature for parallel tempering}
}
\value{
a single function pointer that takes only pars and infectionHistories as unnamed arguments. This function goes on to return a vector of posterior values for each individual
}
\description{
Essential for the MCMC algorithm. Takes all of the input data/parameters and returns a function pointer. This function finds the posterior for a given set of input parameters (theta) and infection histories without needing to pass the data set back and forth.
}
