% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_prior_lookup}
\alias{create_prior_lookup}
\title{Create infection-history prior lookup table}
\usage{
create_prior_lookup(
  antibody_data,
  possible_exposure_times,
  infection_model_prior_shape1,
  beta1,
  n_alive = NULL
)
}
\arguments{
\item{antibody_data}{the antibody data used to calculate the number alive at each exposure time}

\item{possible_exposure_times}{the vector of possible exposure times}

\item{infection_model_prior_shape1}{first shape parameter of the Beta distribution}

\item{beta1}{second shape parameter of the Beta distribution}

\item{n_alive}{optional number alive at each exposure time. If NULL, this is calculated from `antibody_data`.}
}
\value{
a matrix of beta-binomial log-prior values
}
\description{
Pre-calculates beta-binomial log-prior values for each possible number of infections and exposure time.
}
\keyword{internal}
