% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_prior_lookup_groups}
\alias{create_prior_lookup_groups}
\title{Create grouped infection-history prior lookup table}
\usage{
create_prior_lookup_groups(
  antibody_data,
  demographics = NULL,
  possible_exposure_times,
  infection_model_prior_shape1,
  beta1,
  n_alive = NULL
)
}
\arguments{
\item{antibody_data}{the antibody data used to calculate the number alive in each population group}

\item{demographics}{optional time-varying demographic data used to calculate group-specific numbers alive}

\item{possible_exposure_times}{the vector of possible exposure times}

\item{infection_model_prior_shape1}{first shape parameter of the Beta distribution}

\item{beta1}{second shape parameter of the Beta distribution}

\item{n_alive}{optional matrix of numbers alive by population group and exposure time}
}
\value{
an array of beta-binomial log-prior values
}
\description{
Pre-calculates beta-binomial log-prior values for each population group, possible number of infections, and exposure time.
}
\keyword{internal}
