% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{create_prior_mu}
\alias{create_prior_mu}
\title{Create strain specific bias prior}
\usage{
create_prior_mu(par_tab)
}
\arguments{
\item{par_tab}{the parameter table as in \code{\link{create_posterior_func}}}
}
\value{
a function pointer
}
\description{
Creates a function that can be used to easily solve the prior on the strain-specific boosting parameters (normal random effects)
}
\seealso{
Other priors: 
\code{\link{calc_phi_probs_indiv}()},
\code{\link{calc_phi_probs_spline}()},
\code{\link{calc_phi_probs}()},
\code{\link{create_prob_shifts}()},
\code{\link{find_beta_prior_mode}()},
\code{\link{find_beta_prior_with_mean_var}()},
\code{\link{find_beta_prior_with_mean}()},
\code{\link{fit_beta_prior}()},
\code{\link{fit_normal_prior}()},
\code{\link{inf_mat_prior}()},
\code{\link{infection_history_prior}()},
\code{\link{prob_mus}()},
\code{\link{prob_shifts}()}
}
\concept{priors}
