% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_model_likelihoods.R
\name{create_prob_shifts}
\alias{create_prob_shifts}
\title{Measurement shift creation}
\usage{
create_prob_shifts(par_tab)
}
\arguments{
\item{par_tab}{the parameter table as in \code{\link{create_posterior_func}}}
}
\value{
a function to calculate the log prior probability of the measurement shifts
}
\description{
Creates a function to calculate the prior probability of a set of measurement shifts. Assumes normally distributed random effects
}
\seealso{
Other priors:
\code{\link[=calc_phi_probs]{calc_phi_probs()}},
\code{\link[=calc_phi_probs_indiv]{calc_phi_probs_indiv()}},
\code{\link[=calc_phi_probs_spline]{calc_phi_probs_spline()}},
\code{\link[=dbb_prior]{dbb_prior()}},
\code{\link[=density_beta_binom]{density_beta_binom()}},
\code{\link[=find_beta_prior_mode]{find_beta_prior_mode()}},
\code{\link[=find_beta_prior_with_mean]{find_beta_prior_with_mean()}},
\code{\link[=find_beta_prior_with_mean_var]{find_beta_prior_with_mean_var()}},
\code{\link[=fit_beta_prior]{fit_beta_prior()}},
\code{\link[=fit_normal_prior]{fit_normal_prior()}},
\code{\link[=inf_mat_prior]{inf_mat_prior()}},
\code{\link[=infection_history_prior]{infection_history_prior()}},
\code{\link[=prob_shifts]{prob_shifts()}}
}
\concept{priors}
