% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{create_prob_shifts}
\alias{create_prob_shifts}
\title{Measurement shift creation}
\usage{
create_prob_shifts(par_tab)
}
\arguments{
\item{par_tab}{the parameter table as in \code{\link{create_posterior_func}}}
}
\value{
a function pointer to solve the measurement shifts prior
}
\description{
Creates a function to calculate the prior probability of a set of measurement shifts. Assumes normally distributed random effects
}
