% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{create_prob_shifts}
\alias{create_prob_shifts}
\title{Measurement shift creation}
\usage{
create_prob_shifts(par_tab)
}
\arguments{
\item{par_tab}{the parameter table as in \code{\link{create_posterior_func}}}
}
\value{
a function pointer to solve the measurement shifts prior
}
\description{
Creates a function to calculate the prior probability of a set of measurement shifts. Assumes normally distributed random effects
}
\seealso{
Other priors: \code{\link{calc_phi_probs_indiv}},
  \code{\link{calc_phi_probs_spline}},
  \code{\link{calc_phi_probs}},
  \code{\link{create_prior_mu}},
  \code{\link{find_beta_prior_mode}},
  \code{\link{find_beta_prior_with_mean_var}},
  \code{\link{find_beta_prior_with_mean}},
  \code{\link{fit_beta_prior}},
  \code{\link{fit_normal_prior}},
  \code{\link{inf_mat_prior}},
  \code{\link{infection_history_prior}},
  \code{\link{prob_mus}}, \code{\link{prob_shifts}}
}
\concept{priors}
