% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs1_chains_sim}
\alias{cs1_chains_sim}
\title{MCMC chains case study 1, sim}
\format{A list of length 4, as returned by \code{\link{load_mcmc_chains}}.}
\usage{
data(cs1_chains_sim)
}
\description{
MCMC outputs from case study 1 analysis on simulated data. This example data object contains 3 MCMC chains thinned down to 100 samples from the posterior for each chain. The structure of the data is as returned from \code{\link{load_mcmc_chains}}.
}
\seealso{
Other case_study_data: \code{\link{cs1_chains_real_b}},
  \code{\link{cs1_chains_real}},
  \code{\link{cs2_chains_real_b}},
  \code{\link{cs2_chains_real}},
  \code{\link{cs2_chains_sim}}
}
\concept{case_study_data}
\keyword{datasets}
