% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs2_chains_real}
\alias{cs2_chains_real}
\title{MCMC chains case study 2, real}
\format{A list of length 4, as returned by \code{\link{load_mcmc_chains}}.}
\usage{
data(cs2_chains_real)
}
\description{
MCMC outputs from case study 2 analysis on real data. This example data object contains 3 MCMC chains thinned down to 100 samples from the posterior for each chain. The structure of the data is as returned from \code{\link{load_mcmc_chains}}.
}
\seealso{
Other case_study_data: \code{\link{cs1_chains_real_b}},
  \code{\link{cs1_chains_real}},
  \code{\link{cs1_chains_sim}},
  \code{\link{cs2_chains_real_b}},
  \code{\link{cs2_chains_sim}}
}
\concept{case_study_data}
\keyword{datasets}
