% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{estimate_mode}
\alias{estimate_mode}
\title{Estimate vector mode}
\usage{
estimate_mode(x)
}
\arguments{
\item{x}{the vector to be estimated}
}
\value{
the estimated mode of the given vector of values
}
\description{
Estimate vector mode
}
