% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antigenic_map.R
\name{euc_distance}
\alias{euc_distance}
\title{Calculate Euclidean antigenic distance}
\usage{
euc_distance(i1, i2, fit_data)
}
\arguments{
\item{i1}{index of the first map entry}

\item{i2}{index of the second map entry}

\item{fit_data}{data frame containing `x_coord` and `y_coord`}
}
\value{
the Euclidean distance between the two map entries
}
\description{
Calculates the Euclidean distance between two rows of an antigenic map.
}
\seealso{
Other antigenic_maps:
\code{\link[=generate_antigenic_map_flexible]{generate_antigenic_map_flexible()}},
\code{\link[=melt_antigenic_coords]{melt_antigenic_coords()}},
\code{\link[=setup_antigenic_map]{setup_antigenic_map()}}
}
\concept{antigenic_maps}
\keyword{internal}
