% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_antibody_data}
\alias{example_antibody_data}
\title{Example antibody data}
\format{
A data frame with 4800 rows (each corresponding to a unique antibody measurement) and 6 variables:
\describe{
    \item{individual}{index starting from 1 for the individual ID. This should be continuous from 1:n.}
    \item{sample_time}{time the serum sample was taken, matching the time resolution used in the overall model.}
    \item{biomarker_id}{ID of the biomarker being measured. Usually this would correspond to the time period in which an antigen was assumed to be circulating, e.g., an entry of 2015 would correspond to the variant assumed to be circulating throughout 2015.}
    \item{biomarker_group}{Group ID of the biomarker being measured. This will usually just be 1, but can be used where there are multiple biomarker types (not just antigens/targets) per sample. For example, this might represent antibody levels and avidity.}
    \item{measurement}{the antibody measurement for this biomarker ID at this sampling time.}
    \item{repeat_number}{where there are repeat measurements, gives the repeat number for this serum sample and strain measured.}
    \item{population_group}{group index of this individual, for example, corresponding to a study location.}
    \item{birth}{date of birth matching the time resolution of the model.}
}
}
\usage{
data(example_antibody_data)
}
\description{
Example of the antibody data frame generated by \code{\link{simulate_data}}, with matching infection history in \code{\link{example_inf_hist}}
}
\seealso{
Other example_data:
\code{\link{example_antigenic_map}},
\code{\link{example_inf_chain}},
\code{\link{example_inf_hist}},
\code{\link{example_par_tab}},
\code{\link{example_theta_chain}}
}
\concept{example_data}
\keyword{datasets}
