% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_antigenic_map}
\alias{example_antigenic_map}
\title{Example antigenic map}
\format{A data frame with 48 rows and 3 variables:
\describe{
    \item{x_coord}{x-coordinate for this entry in the 2D map}
    \item{y_coord}{y-coordinate for this entry in the 2D map}
    \item{inf_years}{circulation time of this entry. By default, this can be considered as the circulation year, though the logic extends to quarters, months etc}
}}
\usage{
data(example_antigenic_map)

data(example_inf_hist)

data(example_titre_data)
}
\description{
Example of the 2-dimensional antigenic map used through \code{serosolver}. This data frame specifies the antigenic coordinates for each strain, with the assumption that only one representative strain circulates in each time period

Example of the infection history matrix for a cohort of 50 individuals, matching \code{\link{example_titre_dat}} and \code{\link{example_antigenic_map}}

Example of the titre data data frame generated by \code{\link{simulate_data}}, with matching infection history in \code{\link{example_inf_hist}}
}
\seealso{
Other example_data: 
\code{\link{example_inf_chain}},
\code{\link{example_par_tab}},
\code{\link{example_theta_chain}}

Other example_data: 
\code{\link{example_inf_chain}},
\code{\link{example_par_tab}},
\code{\link{example_theta_chain}}

Other example_data: 
\code{\link{example_inf_chain}},
\code{\link{example_par_tab}},
\code{\link{example_theta_chain}}
}
\concept{example_data}
\keyword{datasets}
