% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_inf_chain}
\alias{example_inf_chain}
\title{Example MCMC chain for infection histories}
\format{
A data table with 17236 rows and 4 variables: `i` (individual), `j` (time period), `x` (inferred infection), and `sampno` (saved MCMC sample).
}
\usage{
data(example_inf_chain)
}
\description{
Outputs of a quick MCMC run corresponding to the example data. See \code{\link{example_theta_chain}} for corresponding theta chain output. Note that this was run with prior version 2
}
\seealso{
Other example_data:
\code{\link{example_antibody_data}},
\code{\link{example_antigenic_map}},
\code{\link{example_inf_hist}},
\code{\link{example_par_tab}},
\code{\link{example_theta_chain}}
}
\concept{example_data}
\keyword{datasets}
