% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_inf_hist}
\alias{example_inf_hist}
\title{Example infection history matrix}
\format{
A matrix with 50 rows (individuals) and 48 columns. Each entry can only be 1 (presence of infection) or a 0 (absence of infection)
}
\usage{
data(example_inf_hist)
}
\description{
Example of the infection history matrix for a cohort of 50 individuals, matching \code{\link{example_antibody_data}} and \code{\link{example_antigenic_map}}
}
\seealso{
Other example_data: 
\code{\link{example_antibody_data}},
\code{\link{example_antigenic_map}},
\code{\link{example_inf_chain}},
\code{\link{example_par_tab}},
\code{\link{example_theta_chain}}
}
\concept{example_data}
\keyword{datasets}
