% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_par_tab}
\alias{example_par_tab}
\title{Example parameter table}
\format{A data frame with 21 rows and 9 variables:
\describe{
    \item{names}{string names of the model parameters}
    \item{values}{numeric values of the parameters}
    \item{fixed}{binary values indicating if the parameter should be fixed (1) or estimated (0) during the MCMC procedure}
    \item{steps}{value between 0 and 1, giving the initial step size in the MCMC proposals. Note that these are adapted automatically}
    \item{lower_bound}{lower numeric bound for the parameter during fitting (lower uniform prior bound)}
    \item{upper_bound}{upper numeric bound for the parameter during fitting (upper uniform prior bound)}
    \item{lower_start}{can be used to set the lower allowable random starting value for the MCMC}
    \item{upper_start}{can be used to set the upper allowable random starting value for the MCMC}
    \item{type}{used for admin during MCMC, set to 1 for normal model parameters, 0 for model options (eg. see titre_dependent), or 2 for attack rate terms, phi}
}}
\usage{
data(example_par_tab)
}
\description{
Example of the parameter table used through \code{serosolver}. This data frame is used to control everything related to the model parameters, including values, names, fixed/estimates, and uniform prior bounds.
}
\seealso{
Other example_data: 
\code{\link{example_antigenic_map}},
\code{\link{example_inf_chain}},
\code{\link{example_theta_chain}}
}
\concept{example_data}
\keyword{datasets}
