% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/mcmc_help.R
\name{expand_summary_inf_chain}
\alias{expand_summary_inf_chain}
\title{Expands default MCMC saved inf_chain}
\usage{
expand_summary_inf_chain(inf_chain, j_vec = NULL)

expand_summary_inf_chain(inf_chain, j_vec = NULL)
}
\arguments{
\item{inf_chain}{the data table with the saved sparse infection history matrix}

\item{j}{optional vector of js to expand the infection history chain for}
}
\value{
the MCMC saved infection history expanded with infection times as columns

long format, full infection history matrix chain
}
\description{
The MCMC function saves sparse matrix summaries of the infection history chain to
save space and run time. This function returns the expanded infection history chain,
as in the original version of the code. Returned value is a data table with leftmost columns
giving sample number and individual, with columns expanded to the right for each infection
period.
}
