% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expand_summary_inf_chain}
\alias{expand_summary_inf_chain}
\title{Expands default MCMC saved infChain}
\usage{
expand_summary_inf_chain(inf_chain)
}
\arguments{
\item{inf_chain}{a data table with the MCMC saved infection history chain}
}
\value{
the MCMC saved infection history expanded with infection times as columns
}
\description{
The MCMC function saves sparse matrix summaries of the infection history chain to
save space and run time. This function returns the expanded infection history chain,
as in the original version of the code. Returned value is a data table with leftmost columns
giving sample number and individual, with columns expanded to the right for each infection
period.
}
