% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{find_beta_prior_mode}
\alias{find_beta_prior_mode}
\title{Find Beta distribution parameters with mode}
\usage{
find_beta_prior_mode(mode1, k)
}
\arguments{
\item{mode1}{the desired mode}

\item{k}{the desired certainty in the prior, must be at least 2. The higher this number, the "stronger" the prior}
}
\value{
a list with shape1 and shape2 parameters for the Beta distribution
}
\description{
Calculates the necessary parameters for the Beta distribution to give the desired mode and certainty, k
}
\examples{
find_beta_prior_mode(0.15, 10)
}
\seealso{
Other priors:
\code{\link[=calc_phi_probs]{calc_phi_probs()}},
\code{\link[=calc_phi_probs_indiv]{calc_phi_probs_indiv()}},
\code{\link[=calc_phi_probs_spline]{calc_phi_probs_spline()}},
\code{\link[=create_prob_shifts]{create_prob_shifts()}},
\code{\link[=find_beta_prior_with_mean]{find_beta_prior_with_mean()}},
\code{\link[=find_beta_prior_with_mean_var]{find_beta_prior_with_mean_var()}},
\code{\link[=fit_beta_prior]{fit_beta_prior()}},
\code{\link[=fit_normal_prior]{fit_normal_prior()}},
\code{\link[=inf_mat_prior]{inf_mat_prior()}},
\code{\link[=infection_history_prior]{infection_history_prior()}},
\code{\link[=prob_shifts]{prob_shifts()}}
}
\concept{priors}
