% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{find_beta_prior_with_mean}
\alias{find_beta_prior_with_mean}
\title{Find beta parameters with maximum variance}
\usage{
find_beta_prior_with_mean(desired_annual_mean, buckets)
}
\arguments{
\item{desired_annual_mean}{the desired ANNUAL mean attack rate}

\item{the}{number of buckets to split each year into}
}
\value{
a list with shape1 and shape2
}
\description{
Given a desired annual mean attack rate and the number of epochs to consider per year, gives the shape1 and shape2 parameters with the desired mean but maximum variance
}
\examples{
find_beta_prior_with_mean(0.15, 1)
}
\seealso{
Other priors: 
\code{\link{calc_phi_probs}()},
\code{\link{calc_phi_probs_indiv}()},
\code{\link{calc_phi_probs_spline}()},
\code{\link{create_prob_shifts}()},
\code{\link{find_beta_prior_mode}()},
\code{\link{find_beta_prior_with_mean_var}()},
\code{\link{fit_beta_prior}()},
\code{\link{fit_normal_prior}()},
\code{\link{inf_mat_prior}()},
\code{\link{infection_history_prior}()},
\code{\link{prob_shifts}()}
}
\concept{priors}
