% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{find_beta_prior_with_mean_var}
\alias{find_beta_prior_with_mean_var}
\title{Find beta parameters for mean and variance}
\usage{
find_beta_prior_with_mean_var(mean, var, make_plot = FALSE)
}
\arguments{
\item{mean}{the mean of the beta distribution (between 0 and 1)}

\item{var}{the variance of the beta distribution (between 0 and 0.25)}

\item{make_plot}{if TRUE, plots the resulting Beta distribution to the R device}
}
\value{
a list with shape1 and shape2
}
\description{
Finds the shape1 and shape2 parameters for a Beta distribution that gives the desired mean and variance.
}
\examples{
find_beta_prior_with_mean_var(0.15,0.1,FALSE)
}
\seealso{
Other priors:
\code{\link[=calc_phi_probs]{calc_phi_probs()}},
\code{\link[=calc_phi_probs_indiv]{calc_phi_probs_indiv()}},
\code{\link[=calc_phi_probs_spline]{calc_phi_probs_spline()}},
\code{\link[=create_prob_shifts]{create_prob_shifts()}},
\code{\link[=dbb_prior]{dbb_prior()}},
\code{\link[=density_beta_binom]{density_beta_binom()}},
\code{\link[=find_beta_prior_mode]{find_beta_prior_mode()}},
\code{\link[=find_beta_prior_with_mean]{find_beta_prior_with_mean()}},
\code{\link[=fit_beta_prior]{fit_beta_prior()}},
\code{\link[=fit_normal_prior]{fit_normal_prior()}},
\code{\link[=inf_mat_prior]{inf_mat_prior()}},
\code{\link[=infection_history_prior]{infection_history_prior()}},
\code{\link[=prob_shifts]{prob_shifts()}}
}
\concept{priors}
