% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{fit_normal_prior}
\alias{fit_normal_prior}
\title{Fit normal distribution to MCMC output}
\usage{
fit_normal_prior(chain_samples, par_name = "", error_tol = 999999999,
  try_attempts = 10, plot_fit = TRUE)
}
\arguments{
\item{chain_samples}{the MCMC chain data frame to be fit to}

\item{par_name}{the column label to fit to}

\item{error_tol}{= 9999999999999, what's the error tolerance on the fit? Might take some tweaking}

\item{try_attempts}{= 10 how many fitting attempts to try before giving up}

\item{plot_fit}{= TRUE, if TRUE, plots the fit to the MCMC chain}
}
\value{
the model fit object as returned by optim
}
\description{
Attempts to fit a normal distribution to a data frame of MCMC output from a previous run.
}
\seealso{
\code{\link{fit_normal_prior}}
}
