% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fonville_map}
\alias{fonville_map}
\title{Antigenic map}
\format{A data frame with the following variables:

  \describe{

  \item{strain}{Virus strain}

  \item{x}{X-coordinate}

  \item{y}{Y-coordinate}

  \item{epoch}{The epoch the strain circulated. Since there is 1 epoch per
  year, this is the same as year.}

  \item{epochs_per_year}{Number of epochs per year.}

  }}
\usage{
fonville_map
}
\description{
Antigenic map
}
\keyword{datasets}
