% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{gen_spline_y}
\alias{gen_spline_y}
\title{Generates a spline for \code{\link{generate_lambdas}}}
\usage{
gen_spline_y(x, knots, degree, theta, intercept = TRUE)
}
\description{
Generates a spline for \code{\link{generate_lambdas}}
}
