% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_antigenic_map_flexible}
\alias{generate_antigenic_map_flexible}
\title{Generate antigenic map, flexible}
\usage{
generate_antigenic_map_flexible(antigenic_distances, buckets = 1,
  clusters = NULL, use_clusters = FALSE, spar = 0.3,
  year_min = 1968, year_max = 2016)
}
\arguments{
\item{antigenic_distances}{a data frame of antigenic coordinates, with columns labelled X, Y and Strain for x coord, y coord and Strain label respectively. "Strain" should be a single number giving the year of circulation of that strain.}

\item{buckets}{= 1 the number of epochs per year. 1 means that each year has 1 strain; 12 means that each year has 12 strains (monthly resolution)}

\item{clusters}{= NULL a data frame of cluster labels, indicating which cluster each circulation year belongs to. Note that each row (year) gets repeated by the number of buckets. Column names "year" and "cluster_used"}

\item{use_clusters}{= FALSE if TRUE, uses the clusters data frame above, otherwise just returns as normal}

\item{spar}{= 0.3 to be passed to smooth.spline}

\item{year_min}{= 1968 first year in the antigenic map (usually 1968)}

\item{year_max}{= 2016 last year in the antigenic map}
}
\value{
a fitted antigenic map
}
\description{
Firts a smoothing spline through a set of antigenic coordinates, and uses this to predict antigenic coordinates for all potential infection time points. This version is more flexible than \code{\link{generate_antigenic_map}}, and allows the user to specify "clusters" to assume that strains circulating in a given period are all identical, rather than on a continuous path through space as a function of time.
}
\seealso{
\code{\link{generate_antigenic_map}}
}
