% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{generate_quantiles}
\alias{generate_quantiles}
\title{Formatted quantiles}
\usage{
generate_quantiles(x, sigF = 3, qs = c(0.025, 0.5, 0.975),
  asText = TRUE)
}
\arguments{
\item{x}{the vector to summarise}

\item{sigF}{how many significant figures to print}

\item{qs}{the vector of quantiles}

\item{asText}{if TRUE, formats nicely as text rather than a vector of numbers}
}
\value{
the formatted quantiles
}
\description{
Given a vector of MCMC samples, generates and formats the desired quantile estimates
}
