% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{generate_start_tab}
\alias{generate_start_tab}
\title{Generate starting parameter table}
\usage{
generate_start_tab(par_tab)
}
\arguments{
\item{par_tab}{See \code{\link{example_par_tab}}}
}
\value{
a data frame matching par_tab
}
\description{
Generates a version of \code{par_tab} with random values between \code{lower_start} and \code{upper_start}
}
\examples{
data(example_par_tab)
start_tab <- generate_start_tab(example_par_tab)
}
\seealso{
Other mcmc:
\code{\link[=expand_summary_inf_chain]{expand_summary_inf_chain()}},
\code{\link[=rm_scale]{rm_scale()}},
\code{\link[=save_infection_history_to_disk]{save_infection_history_to_disk()}},
\code{\link[=scaletuning]{scaletuning()}},
\code{\link[=scaletuning_alt]{scaletuning_alt()}},
\code{\link[=serosolver]{serosolver()}}
}
\concept{mcmc}
