% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_antibody_level_predictions}
\alias{get_antibody_level_predictions}
\title{Generate antibody level credible intervals}
\usage{
get_antibody_level_predictions(
  chain,
  infection_histories,
  antibody_data,
  individuals,
  antigenic_map = NULL,
  possible_exposure_times = NULL,
  par_tab,
  nsamp = 1000,
  add_residuals = FALSE,
  measurement_indices_by_time = NULL,
  for_res_plot = FALSE,
  expand_antibody_data = FALSE,
  expand_to_all_times = FALSE,
  antibody_level_before_infection = FALSE,
  for_regression = FALSE,
  data_type = 1,
  start_level = "none"
)
}
\arguments{
\item{chain}{the full MCMC chain to generate antibody level trajectories from}

\item{infection_histories}{the MCMC chain for infection histories}

\item{antibody_data}{the data frame of antibody level data}

\item{individuals}{the subset of individuals to generate credible intervals for}

\item{antigenic_map}{(optional) a data frame of antigenic x and y coordinates. Must have column names: x_coord; y_coord; inf_times. See \code{\link{example_antigenic_map}}}

\item{possible_exposure_times}{(optional) if no antigenic map is specified, this argument gives the vector of times at which individuals can be infected}

\item{par_tab}{the table controlling the parameters in the MCMC chain}

\item{nsamp}{number of samples to take from posterior}

\item{add_residuals}{if true, returns an extra output summarising residuals between the model prediction and data}

\item{measurement_indices_by_time}{default NULL, optional vector giving the index of `measurement_bias` that each strain uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters}

\item{for_res_plot}{TRUE/FALSE value. If using the output of this for plotting of residuals, returns the actual data points rather than summary statistics}

\item{expand_antibody_data}{TRUE/FALSE value. If TRUE, solves antibody level predictions for the entire study period (i.e., between the range of antibody_data$sample_time). If left FALSE, then only solves for the infections times at which a antibody level against the circulating biomarker_id was measured in antibody_data.}

\item{expand_to_all_times}{TRUE/FALSE value. If TRUE, solves antibody level predictions for all possible infection times (i.e., for the range in possible_exposure_times). If left FALSE, then only solves for the infections times at which a antibody level against the circulating biomarker_id was measured in antibody_data.}

\item{antibody_level_before_infection}{TRUE/FALSE value. If TRUE, solves antibody level predictions, but gives the predicted antibody level at a given time point BEFORE any infection during that time occurs.}

\item{for_regression}{if TRUE, returns posterior draws rather than posterior summaries}

\item{data_type}{integer, currently accepting 1 or 2. Set to 1 for discretized, bounded data, or 2 for continuous, bounded data.}
}
\value{
a list with the antibody level predictions (95% credible intervals, median and multivariate posterior mode) and the probabilities of infection for each individual in each epoch
}
\description{
Generates credible intervals on antibody levels and infection histories from an MCMC chain output.
}
\examples{
\dontrun{
data(example_theta_chain)
data(example_inf_chain)
data(example_antibody_data)
data(example_antigenic_map)
data(example_par_tab)

y <- get_antibody_level_predictions(example_theta_chain, example_inf_chain, example_antibody_data,
                          unique(example_antibody_data$individual), example_antigenic_map,
                          example_par_tab,expand_antibody_data = FALSE)
}
}
