% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_best_draw}
\alias{get_best_draw}
\title{Best infection history and theta parameters}
\usage{
get_best_draw(chain, inf_chain, max_indivs = NULL, max_times = NULL)
}
\arguments{
\item{chain}{the MCMC chain for theta parameters}

\item{inf_chain}{the MCMC chain for infection histories}

\item{max_indivs}{the maximum number of individuals in the infection history matrix (default is max individual in inf_chain)}

\item{max_times}{the maximum number of time points in the infection history matrix (default is max time in inf_chain)}
}
\value{
a list with a vector of the best parameters and a matrix of the best infection history
}
\description{
Given an MCMC chain of theta parameters and infection histories, returns the MAP infection history
}
\examples{
\dontrun{
mcmc_chains <- load_theta_chains()
inf_chains <- load_infection_chains()
best_pars <- get_best_draw(mcmc_chains$chain)
}
}
\seealso{
Other mcmc_diagnostics: 
\code{\link{get_best_pars}()},
\code{\link{get_index_pars}()}
}
\concept{mcmc_diagnostics}
