% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification.R
\name{get_demographic_groups}
\alias{get_demographic_groups}
\title{Get stratification groups used to group infection history model}
\usage{
get_demographic_groups(
  par_tab,
  antibody_data,
  timevarying_demographics = NULL,
  demographic_groups = NULL
)
}
\arguments{
\item{par_tab}{the parameter table. See \code{\link{example_par_tab}} for an example.}

\item{antibody_data}{the antibody data data frame. See \code{\link{example_antibody_data}} for an example.}

\item{timevarying_demographics}{optional demographic data. Include `individual`, any stratification variables, and `time` for time-varying demographics. Without `time`, the table is treated as fixed for each individual. If NULL, then uses antibody_data.}

\item{demographic_groups}{optional. a data frame of demographic groups, with columns for each demographic group and rows for each unique combination of demographic groups. If NULL, then create this from antibody_data or timevarying_demographics}
}
\value{
a list with three entries: 1) the names of the demographic groups used, 2) a data frame of demographic groups, with each row a unique combination of demographic groups, and 3) a boolean indicating whether timevarying demographics were used
}
\description{
Takes antibody_data and demographics and returns a list with the stratification groups used to group the antibody kinetics model. If timevarying_demographics are provided, then these are used to create the demographic groups. If not, then antibody_data is used.
}
\seealso{
Other stratification:
\code{\link[=add_scale_pars]{add_scale_pars()}},
\code{\link[=create_demographic_table]{create_demographic_table()}},
\code{\link[=setup_stratification_table]{setup_stratification_table()}}
}
\concept{stratification}
