% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_DOBs}
\alias{get_DOBs}
\title{Get DOBs}
\usage{
get_DOBs(antibody_data)
}
\arguments{
\item{antibody_data}{the data frame of antibody data. See \code{\link{example_antibody_data}}}
}
\value{
a data frame with individual and birth
}
\description{
Gets the dates of birth for each individual in antibody_data
}
\examples{
\dontrun{
data(example_antibody_data)
get_DOBs(antibody_data)
}
}
\seealso{
Other get_summary:
\code{\link[=get_n_alive]{get_n_alive()}},
\code{\link[=get_n_alive_group]{get_n_alive_group()}}
}
\concept{get_summary}
\keyword{internal}
