% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_DOBs}
\alias{get_DOBs}
\title{Get DOBs}
\usage{
get_DOBs(titre_dat)
}
\arguments{
\item{titre_dat}{the data frame of titre data. See \code{\link{example_titre_dat}}}
}
\value{
a data frame with individual ID and DOB
}
\description{
Gets the dates of birth for each individual in titre_dat
}
\examples{
data(example_titre_dat)
get_DOBs(titre_dat)
}
\seealso{
Other get_summary: \code{\link{get_n_alive_group}},
  \code{\link{get_n_alive}}
}
\concept{get_summary}
