% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_index_pars}
\alias{get_index_pars}
\title{Index pars}
\usage{
get_index_pars(chain, index)
}
\arguments{
\item{chain}{the MCMC chain}

\item{index}{the index}
}
\value{
a named vector of the best parameters
}
\description{
Given an MCMC chain, returns the parameters at the specified index
}
\examples{
\dontrun{
mcmc_chains <- load_theta_chains()
pars <- get_index_pars(mcmc_chains$chain, 1000)
}
}
\seealso{
Other mcmc_diagnostics: 
\code{\link{get_best_pars}()}
}
\concept{mcmc_diagnostics}
