% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_n_alive}
\alias{get_n_alive}
\title{Get number alive}
\usage{
get_n_alive(antibody_data, times)
}
\arguments{
\item{antibody_data}{the data frame of titre data. See \code{\link{example_antibody_data}}}

\item{times}{the vector of times to calculate number alive for}
}
\value{
a vector giving the number alive in each time point
}
\description{
Given the antibody_data data frame, calculates the number that are alive (alive to be infected, that is) for each time in times
}
\examples{
data(example_antibody_data)
data(example_antigenic_map)
times <- unique(example_antigenic_map$inf_times)
get_n_alive(example_antibody_data, times)
}
\seealso{
Other get_summary:
\code{\link[=get_DOBs]{get_DOBs()}},
\code{\link[=get_n_alive_group]{get_n_alive_group()}}
}
\concept{get_summary}
