% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_n_alive}
\alias{get_n_alive}
\title{Get number alive}
\usage{
get_n_alive(titre_dat, times)
}
\arguments{
\item{titre_dat}{the data frame of titre data}

\item{times}{the vector of times to calculate number alive for}
}
\value{
a vector giving the number alive in each time point
}
\description{
Given the titre_dat data frame, calculates the number that are alive (alive to be infected, that is) for each time in times
}
