% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_n_alive_group}
\alias{get_n_alive_group}
\title{Get number alive by location}
\usage{
get_n_alive_group(antibody_data, times, demographics = NULL, melt_data = FALSE)
}
\arguments{
\item{antibody_data}{the data frame of antibody data. See \code{\link{example_antibody_data}}}

\item{times}{the vector of times to calculate number alive for}

\item{melt_data}{if TRUE, returns a melted data frame. Returns a wide matrix otherwise.}
}
\value{
a matrix giving the number alive in each time point in each location
}
\description{
Given the antibody_data data frame with entries for location, calculates the number that are alive (alive to be infected, that is) for each time in times by location
}
\examples{
data(example_antibody_data)
data(example_antigenic_map)
times <- unique(example_antigenic_map$inf_times)
get_n_alive_group(example_antibody_data, times)
}
\seealso{
Other get_summary:
\code{\link[=get_DOBs]{get_DOBs()}},
\code{\link[=get_n_alive]{get_n_alive()}}
}
\concept{get_summary}
