% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_total_number_infections}
\alias{get_total_number_infections}
\title{Get total number of infections}
\usage{
get_total_number_infections(inf_chain, pad_chain = TRUE)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{run_MCMC}}}

\item{pad_chain}{if TRUE, pads the infection history MCMC chain to have entries for non-infection events}
}
\value{
a data table
}
\description{
Finds the total number of infections for each iteration of an MCMC chain
}
