% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{hist_rbb}
\alias{hist_rbb}
\title{Beta binomial histogram}
\usage{
hist_rbb(n, mean, var)
}
\arguments{
\item{n}{the number of trials}

\item{mean}{desired mean}

\item{var}{desired variance}
}
\description{
Plots a histogram of the beta binomial function with given mean and variance
}
\examples{
hist_rbb(100, 0.15, 0.1)
}
\seealso{
Other beta_binomial_functions: \code{\link{bb_mean}},
  \code{\link{dbb}}, \code{\link{find_bb_2}},
  \code{\link{find_prior_alpha_beta}}, \code{\link{pbb}},
  \code{\link{qbb}}, \code{\link{rbb}}
}
\concept{beta_binomial_functions}
