% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inf_hist_prop_prior_v2_and_v4}
\alias{inf_hist_prop_prior_v2_and_v4}
\title{Infection history gibbs proposal, fast
 Generates a new infection history matrix and corresponding individual likelihoods, using a gibbs sampler from the infection history prior. See \code{\link{infection_history_proposal_gibbs}}, as inputs are very similar.}
\usage{
inf_hist_prop_prior_v2_and_v4(theta, infection_history_mat, old_probs_1,
  sampled_indivs, n_years_samp_vec, age_mask, strain_mask, n_alive,
  n_infections, n_infected_group, swap_propn, swap_distance, alpha, beta,
  circulation_times, circulation_times_indices, sample_times,
  rows_per_indiv_in_samples, cum_nrows_per_individual_in_data,
  cum_nrows_per_individual_in_repeat_data, nrows_per_blood_sample,
  group_id_vec, measurement_strain_indices, antigenic_map_long,
  antigenic_map_short, data, repeat_data, repeat_indices, titre_shifts,
  mus, boosting_vec_indices, total_alive, temp = 1,
  solve_likelihood = TRUE)
}
\arguments{
\item{theta}{NumericVector, the model parameters used to solve the model}

\item{infection_history_mat}{IntegerMatrix the matrix of 1s and 0s corresponding to individual infection histories}

\item{old_probs_1}{NumericVector, the current likelihoods for each individual}

\item{sampled_indivs}{IntegerVector, indices of sampled individuals}

\item{age_mask}{IntegerVector, length of the number of individuals, with indices specifying first time period that an individual can be infected (indexed from 1, such that a value of 1 allows an individual to be infected in any time period)}

\item{strain_mask}{IntegerVector, length of the number of individuals, with indices specifying last time period that an individual can be infected (ie. last time a sample was taken)}

\item{n_alive}{IntegerMatrix, number of columns is the number of time periods that an individual could be infected, giving the number of individual alive in each time period. Number of rows is the number of distinct groups.}

\item{swap_propn}{double, what proportion of proposals should be swap steps (ie. swap contents of two cells in infection_history rather than adding/removing infections)}

\item{swap_distance}{int, in a swap step, how many time steps either side of the chosen time period to swap with}

\item{alpha}{double, alpha parameter for beta prior on infection probability}

\item{beta}{double, beta parameter for beta prior on infection probability}

\item{circulation_times}{NumericVector, the times that each strain circulated}

\item{circulation_times_indices}{IntegerVector, indexing vector from 0:(number of strains-1)}

\item{sample_times}{NumericVector, the vector of real times that samples were taken}

\item{rows_per_indiv_in_samples}{IntegerVector, How many rows in titre data correspond to each individual, sample and repeat?}

\item{cum_nrows_per_individual_in_data}{IntegerVector, How many rows in the titre data correspond to each individual?}

\item{cum_nrows_per_individual_in_repeat_data}{IntegerVector, For the repeat data (ie. already calculated these titres), how many rows in the titre data correspond to each individual?}

\item{nrows_per_blood_sample}{IntegerVector, Split the sample times and runs for each individual}

\item{group_id_vec}{IntegerVector, vector with 1 entry per individual, giving the group ID of that individual}

\item{measurement_strain_indices}{IntegerVector, For each titre measurement, corresponding entry in antigenic map}

\item{antigenic_map_long}{NumericVector, the collapsed cross reactivity map for long term boosting, after multiplying by sigma1, see \code{\link{create_cross_reactivity_vector}}}

\item{antigenic_map_short}{NumericVector, the collapsed cross reactivity map for short term boosting, after multiplying by sigma2, see \code{\link{create_cross_reactivity_vector}}}

\item{data}{NumericVector, data for all individuals for the first instance of each calculated titre}

\item{repeat_data}{NumericVector, the repeat titre data for all individuals (ie. do not solve the same titres twice)}

\item{repeat_indices}{IntegerVector, which index in the main data vector does each entry in repeat_data correspond to ie. which calculated titre in predicted_titres should be used for each observation?}

\item{titre_shifts}{NumericVector, if length matches the length of \code{data}, adds these as measurement shifts to the predicted titres. If lengths do not match, is not used.}

\item{mus}{NumericVector, if length is greater than one, assumes that strain-specific boosting is used rather than a single boosting parameter}

\item{boosting_vec_indices}{IntegerVector, same length as circulation_times, giving the index in the vector \code{mus} that each entry should use as its boosting parameter.}

\item{temp}{double, temperature for parallel tempering MCMC}

\item{solve_likelihood}{bool, if FALSE does not solve likelihood when calculating acceptance probability}

\item{n_years_samp}{int, for each individual, how many time periods to resample infections for?}
}
\value{
an R list, one entry with the matrix of 1s and 0s corresponding to the infection histories for all individuals and the other with the new corresponding likelihoods per individual
}
\description{
Infection history gibbs proposal, fast
 Generates a new infection history matrix and corresponding individual likelihoods, using a gibbs sampler from the infection history prior. See \code{\link{infection_history_proposal_gibbs}}, as inputs are very similar.
}
\seealso{
Other infection_history_proposal: \code{\link{inf_hist_prop_prior_v3}}
}
\concept{infection_history_proposal}
