% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inf_hist_prop_prior_v3}
\alias{inf_hist_prop_prior_v3}
\title{Infection history proposal function}
\usage{
inf_hist_prop_prior_v3(
  infection_history_mat,
  sampled_indivs,
  age_mask,
  sample_mask,
  proposal_inf_hist_distances,
  n_infs,
  shape1,
  shape2,
  rand_ns,
  proposal_inf_hist_indiv_swap_ratio
)
}
\arguments{
\item{infection_history_mat}{and RcppArmadillo matrix of infection histories, where rows represent individuals and columns represent potential infection times. The contents should be a set of 1s (presence of infection) and 0s (absence of infection)}

\item{sampled_indivs}{IntegerVector, indices of which individuals to resample. Note that this is indexed from 1 (ie. as if passing straight from R)}

\item{age_mask}{IntegerVector, for each individual gives the first column in the infection history matrix that an individual could have been exposed to indexed from 1. ie. if alive for the whole period, entry would be 1. If alive for the 11th epoch, entry would be 11.}

\item{proposal_inf_hist_distances}{IntegerVector, how far can a swap step sample from specified for each individual}

\item{n_infs}{IntegerVector, how many infections to add/remove/swap with each proposal step for each individual}

\item{shape1}{double, shape1 (alpha) parameter of the beta binomial}

\item{shape2}{double, shape2 (beta) parameter of the beta binomial}

\item{rand_ns}{NumericVector, a vector of random numbers for each sampled individual. The idea is to pre-specify whether an individual experiences an add/remove step or a swap step to avoid random number sampling in C++}

\item{strain_mask}{IntegerVector, for each individual gives the last column in the infection history matrix that an individual could have been exposed to indexed from 1. ie. if their last serum sample was in the 40th epoch, entry would be 40}
}
\value{
a matrix of 1s and 0s corresponding to the infection histories for all individuals
}
\description{
Proposes a new matrix of infection histories using a beta binomial proposal distribution. This particular implementation allows for n_infs epoch times to be changed with each function call. Furthermore, the size of the swap step is specified for each individual by proposal_inf_hist_distances.
}
\seealso{
Other infection_history_proposal:
\code{\link[=inf_hist_prop_prior_v2_and_v4]{inf_hist_prop_prior_v2_and_v4()}}
}
\concept{infection_history_proposal}
