% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{inf_hist_swap}
\alias{inf_hist_swap}
\title{Swap infection history years}
\usage{
inf_hist_swap(infection_history, exposure_history = NULL, age_mask,
  strain_mask, swap_propn, move_size)
}
\arguments{
\item{infection_history}{matrix of 1s and 0s to swap around representing the infection history}

\item{exposure_history}{optional NULL, matrix of 1s and 0s to swap around representing the expousre history}

\item{age_mask}{the first index in infection_history that each individual (row) could be infected in}

\item{strain_mask}{the last index in infection_history that each individual (row) could be infected in ie. the time of the latest blood sample}

\item{swap_propn}{what proportion of infections should be swapped?}

\item{move_size}{How many time points away should be chosen as candidate swaps?}
}
\value{
the same infection_history matrix, but with two columns swapped
}
\description{
Swaps the entire contents of two columns of the infection history matrix, adhering to age and strain limitations.
}
