% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{inf_hist_swap}
\alias{inf_hist_swap}
\title{Swap infection history years}
\usage{
inf_hist_swap(
  infection_history,
  age_mask,
  strain_mask,
  swap_propn,
  move_size,
  proposal_ratios = NULL
)
}
\arguments{
\item{infection_history}{matrix of 1s and 0s to swap around representing the infection history}

\item{age_mask}{the first index in infection_history that each individual (row) could be infected in}

\item{strain_mask}{the last index in infection_history that each individual (row) could be infected in ie. the time of the latest blood sample}

\item{swap_propn}{what proportion of infections should be swapped?}

\item{move_size}{How many time points away should be chosen as candidate swaps?}

\item{proposal_ratios}{optional NULL. Can set the relative sampling weights of the infection state times. Should be an integer vector of length matching nrow(antigenic_map). Otherwise, leave as NULL for uniform sampling.}
}
\value{
the same infection_history matrix, but with two columns swapped
}
\description{
Swaps the entire contents of two columns of the infection history matrix, adhering to age and strain limitations.
}
\examples{
data(example_inf_hist)
data(example_titre_dat)
data(example_antigenic_map)
times <- example_antigenic_map$inf_years
ages <- unique(example_titre_dat[,c("DOB","individual")])
age_mask <- create_age_mask(ages$DOB, times)
strain_mask <- create_strain_mask(example_titre_dat, times)
new_inf_hist <- inf_hist_swap(example_inf_hist, age_mask,strain_mask, 1,3)[[1]]
}
\seealso{
Other proposals: 
\code{\link{inf_hist_swap_phi}()},
\code{\link{infection_history_symmetric}()},
\code{\link{mvr_proposal}()},
\code{\link{univ_proposal}()}
}
\concept{proposals}
