% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{inf_hist_swap_phi}
\alias{inf_hist_swap_phi}
\title{Swap infection history years with phi term}
\usage{
inf_hist_swap_phi(infection_history, phis, age_mask, strain_mask,
  swap_propn, move_size, n_alive)
}
\arguments{
\item{infection_history}{matrix of 1s and 0s to swap around representing the infection history}

\item{phis}{vector of force of infection parameters for each column}

\item{age_mask}{the first index in infection_history that each individual (row) could be infected in}

\item{strain_mask}{the last index in infection_history that each individual (row) could be infected in ie. the time of the latest blood sample}

\item{swap_propn}{what proportion of infections should be swapped?}

\item{move_size}{How many time points away should be chosen as candidate swaps?}

\item{n_alive}{number of individuals alive in each entry of phis}
}
\value{
a list: the same infection_history matrix, but with two columns swapped; also the swapped phis
}
\description{
Swaps the entire contents of two columns of the infection history matrix, adhering to age and strain limitations. Also swaps the values of phi that correspond to these years
}
\seealso{
\code{\link{inf_hist_swap}}
}
