% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{inf_mat_prior}
\alias{inf_mat_prior}
\title{Infection history prior R}
\usage{
inf_mat_prior(
  infection_history,
  age_mask,
  infection_model_prior_shape1,
  infection_model_prior_shape2
)
}
\arguments{
\item{infection_history}{the infection history matrix}

\item{age_mask}{the age mask, giving the first index of the infection_history matrix that each individual can be exposed to. One entry per individual}

\item{infection_model_prior_shape1}{shape1 parameter for the Beta distribution}

\item{infection_model_prior_shape2}{shape2 parameter for the Beta distribution}
}
\value{
the total log prior for the infection history
}
\description{
R implementation of the infection history prior assuming common infection risk across individuals in a given year.
}
\seealso{
Other priors:
\code{\link[=calc_phi_probs]{calc_phi_probs()}},
\code{\link[=calc_phi_probs_indiv]{calc_phi_probs_indiv()}},
\code{\link[=calc_phi_probs_spline]{calc_phi_probs_spline()}},
\code{\link[=create_prob_shifts]{create_prob_shifts()}},
\code{\link[=dbb_prior]{dbb_prior()}},
\code{\link[=density_beta_binom]{density_beta_binom()}},
\code{\link[=find_beta_prior_mode]{find_beta_prior_mode()}},
\code{\link[=find_beta_prior_with_mean]{find_beta_prior_with_mean()}},
\code{\link[=find_beta_prior_with_mean_var]{find_beta_prior_with_mean_var()}},
\code{\link[=fit_beta_prior]{fit_beta_prior()}},
\code{\link[=fit_normal_prior]{fit_normal_prior()}},
\code{\link[=infection_history_prior]{infection_history_prior()}},
\code{\link[=prob_shifts]{prob_shifts()}}
}
\concept{priors}
\keyword{internal}
