% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{inf_mat_prior}
\alias{inf_mat_prior}
\title{Infection history prior R}
\usage{
inf_mat_prior(infection_history, age_mask, alpha, beta1)
}
\arguments{
\item{infection_history}{the infection history matrix}

\item{age_mask}{the age mask, giving the first index of the infection_history matrix that each individual can be exposed to. One entry per individual}

\item{alpha}{alpha parameter for beta distribution}

\item{beta}{beta parameter for beta distribution}
}
\value{
a prior probability for each individual
}
\description{
R implementation of the infection history prior assuming common infection risk across individuals in a given year
}
\seealso{
Other priors: \code{\link{infection_history_prior}}
}
\concept{priors}
