% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{inf_mat_prior}
\alias{inf_mat_prior}
\title{Infection history prior R}
\usage{
inf_mat_prior(infection_history, age_mask, alpha1, beta1)
}
\arguments{
\item{infection_history}{the infection history matrix}

\item{age_mask}{the age mask, giving the first index of the infection_history matrix that each individual can be exposed to. One entry per individual}

\item{alpha1}{alpha parameter for beta distribution}

\item{beta1}{beta parameter for beta distribution}
}
\value{
a prior probability for each individual
}
\description{
R implementation of the infection history prior assuming common infection risk across individuals in a given year
}
\seealso{
Other priors: 
\code{\link{calc_phi_probs_indiv}()},
\code{\link{calc_phi_probs}()},
\code{\link{create_prob_shifts}()},
\code{\link{find_beta_prior_mode}()},
\code{\link{find_beta_prior_with_mean_var}()},
\code{\link{find_beta_prior_with_mean}()},
\code{\link{fit_beta_prior}()},
\code{\link{fit_normal_prior}()},
\code{\link{infection_history_prior}()},
\code{\link{prob_shifts}()}
}
\concept{priors}
