% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inf_mat_prior_cpp}
\alias{inf_mat_prior_cpp}
\title{Marginal prior probability (p(Z)) of a particular infection history matrix}
\usage{
inf_mat_prior_cpp(infHist, n_alive, alpha, beta)
}
\arguments{
\item{infHist}{IntegerMatrix, the infection history matrix}

\item{n_alive}{IntegerVector, vector giving the number of individuals alive in each time unit}

\item{alpha}{NumericVector, alpha parameter for beta distribution prior, one for each time unit}

\item{beta}{NumericVector, beta parameter for beta distribution prior, one for each time unit}
}
\description{
Marginal prior probability (p(Z)) of a particular infection history matrix
}
