% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inf_mat_prior_cpp_vector}
\alias{inf_mat_prior_cpp_vector}
\title{Marginal prior probability (p(Z)) of a particular infection history matrix vector prior
 Prior is independent contribution from each year, but each year has its own alpha and beta}
\usage{
inf_mat_prior_cpp_vector(infection_history, n_alive, alphas, betas)
}
\arguments{
\item{infection_history}{IntegerMatrix, the infection history matrix}

\item{n_alive}{IntegerVector, vector giving the number of individuals alive in each time unit}

\item{alphas}{NumericVector, alpha parameters for beta distribution prior, one for each time unit}

\item{betas}{NumericVector, beta parameters for beta distribution prior, one for each time unit}
}
\value{
a single prior probability
}
\description{
Marginal prior probability (p(Z)) of a particular infection history matrix vector prior
 Prior is independent contribution from each year, but each year has its own alpha and beta
}
\seealso{
Other inf_mat_prior: \code{\link{inf_mat_prior_cpp}},
  \code{\link{inf_mat_prior_group_cpp_vector}},
  \code{\link{inf_mat_prior_group_cpp}},
  \code{\link{inf_mat_prior_total_group_cpp}}
}
\concept{inf_mat_prior}
