% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inf_mat_prior_group_cpp}
\alias{inf_mat_prior_group_cpp}
\title{Marginal prior probability (p(Z)) of infection history matrix with groups
 Prior is independent contribution from each year and group}
\usage{
inf_mat_prior_group_cpp(n_infections, n_alive, alpha, beta)
}
\arguments{
\item{n_infections}{IntegerMatrix, the total number of infections in each time point/group}

\item{n_alive}{IntegerMatrix, matrix giving the number of individuals alive in each time unit in each group}

\item{alpha}{double, alpha parameter for beta distribution prior}

\item{beta}{double, beta parameter for beta distribution prior}
}
\value{
a single prior probability
}
\description{
Marginal prior probability (p(Z)) of infection history matrix with groups
 Prior is independent contribution from each year and group
}
\seealso{
Other inf_mat_prior: 
\code{\link{inf_mat_prior_cpp_vector}()},
\code{\link{inf_mat_prior_cpp}()},
\code{\link{inf_mat_prior_group_cpp_vector}()},
\code{\link{inf_mat_prior_total_group_cpp}()}
}
\concept{inf_mat_prior}
