% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inf_mat_prior_total_group_cpp}
\alias{inf_mat_prior_total_group_cpp}
\title{Marginal prior probability (p(Z)) of a particular infection history matrix total prior
 Prior here is on the total number of infections across all individuals and times}
\usage{
inf_mat_prior_total_group_cpp(n_infections_group, n_alive_group, alpha,
  beta)
}
\arguments{
\item{n_infections_group}{IntegerVector, the total number of infections in each group}

\item{n_alive_group}{IntegerVector, vector giving total number of potential infection events per group}

\item{alpha}{double, alpha parameter for beta distribution prior}

\item{beta}{double, beta parameter for beta distribution prior}
}
\value{
a single prior probability
}
\description{
Marginal prior probability (p(Z)) of a particular infection history matrix total prior
 Prior here is on the total number of infections across all individuals and times
}
\seealso{
Other inf_mat_prior: \code{\link{inf_mat_prior_cpp_vector}},
  \code{\link{inf_mat_prior_cpp}},
  \code{\link{inf_mat_prior_group_cpp_vector}},
  \code{\link{inf_mat_prior_group_cpp}}
}
\concept{inf_mat_prior}
