% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_infection_history.R
\name{infection_history_betabinom_group}
\alias{infection_history_betabinom_group}
\title{DEPRECATED - implemented in Cpp for speed}
\usage{
infection_history_betabinom_group(new_inf_hist, sampled_indivs, age_mask,
  strain_mask, move_sizes, n_infs, alpha, beta)
}
\description{
DEPRECATED - implemented in Cpp for speed
}
