% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{infection_history_prior}
\alias{infection_history_prior}
\title{Beta binomial infection history prior}
\usage{
infection_history_prior(pars, infection_history, age_mask)
}
\arguments{
\item{pars}{the named parameter vector with infection_model_prior_shape1 and infection_model_prior_shape2 corresponding to the beta binomial distribution}

\item{infection_history}{the infection history matrix}

\item{age_mask}{the age mask, giving the first index of the infection_history matrix that each individual can be exposed to. One entry per individual}
}
\value{
the log prior contribution for each individual
}
\description{
Calculates the beta binomial infection history prior for each individual, taking into account the number of years they could be infected.
}
\seealso{
Other priors:
\code{\link[=calc_phi_probs]{calc_phi_probs()}},
\code{\link[=calc_phi_probs_indiv]{calc_phi_probs_indiv()}},
\code{\link[=calc_phi_probs_spline]{calc_phi_probs_spline()}},
\code{\link[=create_prob_shifts]{create_prob_shifts()}},
\code{\link[=dbb_prior]{dbb_prior()}},
\code{\link[=density_beta_binom]{density_beta_binom()}},
\code{\link[=find_beta_prior_mode]{find_beta_prior_mode()}},
\code{\link[=find_beta_prior_with_mean]{find_beta_prior_with_mean()}},
\code{\link[=find_beta_prior_with_mean_var]{find_beta_prior_with_mean_var()}},
\code{\link[=fit_beta_prior]{fit_beta_prior()}},
\code{\link[=fit_normal_prior]{fit_normal_prior()}},
\code{\link[=inf_mat_prior]{inf_mat_prior()}},
\code{\link[=prob_shifts]{prob_shifts()}}
}
\concept{priors}
