% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{infection_history_prior}
\alias{infection_history_prior}
\title{Beta binomial infection history prior}
\usage{
infection_history_prior(pars, infection_history, age_mask)
}
\arguments{
\item{pars}{the named parameter vector with alpha and beta}

\item{infection_history}{the infection hsitory matrix}

\item{age_mask}{the age mask, giving the first index of the infection_history matrix that each individual can be exposed to. One entry per individual}
}
\value{
a prior probability for each individual
}
\description{
Calculates the beta binomial infection history prior for each individual, taking into account the number of years they could be infected.
}
\seealso{
Other priors: \code{\link{calc_phi_probs_indiv}},
  \code{\link{calc_phi_probs_spline}},
  \code{\link{calc_phi_probs}},
  \code{\link{create_prior_mu}},
  \code{\link{create_prob_shifts}},
  \code{\link{find_beta_prior_mode}},
  \code{\link{find_beta_prior_with_mean_var}},
  \code{\link{find_beta_prior_with_mean}},
  \code{\link{fit_beta_prior}},
  \code{\link{fit_normal_prior}},
  \code{\link{inf_mat_prior}}, \code{\link{prob_mus}},
  \code{\link{prob_shifts}}
}
\concept{priors}
