% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{infection_history_proposal}
\alias{infection_history_proposal}
\title{Infection history proposal}
\usage{
infection_history_proposal(newInfectionHistories, sampledIndivs,
  strainIsolationTimes, ageMask)
}
\arguments{
\item{newInfectionHistories}{an n*m matrix of 1s & 0s indicating infection histories, where n is individuals and m i strains}

\item{sampledIndivs}{the indices of sampled individuals to receive proposals}

\item{strainIsolationTimes}{the vector of strain isolation times in real time}

\item{ageMask}{the vector of indices for each individual specifiying which index of strainIsolationTimes is the first strain each individual coul dhave seen}
}
\value{
a new matrix matching newInfectionHistories in dimensions with proposed moves
}
\description{
Proposes new infection histories for a vector of infection histories, where rows represent individuals and columns represent years. Proposals are either removal, addition or switching of infections.
Also requires the indices of sampled individuals, the vector of strain isolation times, and a vector of age masks (ie. which index of the strainIsolationTimes vector is the first year in which
an individual *could* be infected).
NOTE - MIGHT NEED TO UPDATE THIS FOR GROUPS
}

