% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{infection_history_proposal}
\alias{infection_history_proposal}
\title{Age mask is a value indicating from which strain index an individual may be infected}
\usage{
infection_history_proposal(newInfectionHistories, sampledIndivs,
  strainIsolationTimes, ageMask)
}
\description{
Age mask is a value indicating from which strain index an individual may be infected
}

