% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infection_history_proposal_gibbs}
\alias{infection_history_proposal_gibbs}
\title{Gibbs sampling of infection histories}
\usage{
infection_history_proposal_gibbs(pars, infection_history_mat,
  indiv_samp_propn, n_years_samp_vec, age_mask, strain_mask, n_alive,
  swap_propn, swap_distance, alpha, beta, circulation_times,
  circulation_times_indices, sample_times, rows_per_indiv_in_samples,
  cum_nrows_per_individual_in_data, nrows_per_blood_sample,
  measurement_strain_indices, antigenic_map_long, antigenic_map_short,
  data, to_add, additional_arguments, DOBs, solve_likelihood = TRUE,
  total_alive = -1L, temp = 1)
}
\arguments{
\item{pars}{NumericVector, the model parameters used to solve the model, extract likelihood function parameters and alpha/beta}

\item{infection_history_mat}{IntegerMatrix the matrix of 1s and 0s corresponding to individual infection histories}

\item{indiv_samp_propn}{double, what proportion of individuals to resample in this proposal step}

\item{age_mask}{IntegerVector, length of the number of individuals, with indices specifying first time period that an individual can be infected (indexed from 1, such that a value of 1 allows an individual to be infected in any time period)}

\item{strain_mask}{IntegerVector, length of the number of individuals, with indices specifying last time period that an individual can be infected (ie. last time a sample was taken)}

\item{n_alive}{IntegerVector, length of the number of time periods that an individual could be infected, giving the number of individual alive in each time period}

\item{swap_propn}{double, what proportion of proposals should be swap steps (ie. swap contents of two cells in infection_history rather than adding/removing infections)}

\item{swap_distance}{int, in a swap step, how many time steps either side of the chosen time period to swap with}

\item{alpha}{double, alpha parameter for beta prior on infection probability}

\item{beta}{double, beta parameter for beta prior on infection probability}

\item{circulation_times}{NumericVector, the times that each strain circulated}

\item{circulation_times_indices}{IntegerVector, indexing vector from 0:(number of strains-1)}

\item{sample_times}{NumericVector, the vector of real times that samples were taken}

\item{rows_per_indiv_in_samples}{IntegerVector, How many rows in titre data correspond to each individual, sample and repeat?}

\item{cum_nrows_per_individual_in_data}{IntegerVector, How many rows in the titre data correspond to each individual?}

\item{nrows_per_blood_sample}{IntegerVector, Split the sample times and runs for each individual}

\item{measurement_strain_indices}{IntegerVector, For each titre measurement, corresponding entry in antigenic map}

\item{antigenic_map_long}{NumericVector, the collapsed cross reactivity map for long term boosting, after multiplying by sigma1}

\item{antigenic_map_short}{NumericVector, the collapsed cross reactivity map for short term boosting, after multiplying by sigma2}

\item{data}{NumericVector, the titre data for all individuals}

\item{to_add}{Nullable<NumericVector>, optional vector of measurement shifts to apply to all titres}

\item{additional_arguments}{Nullable<List>, optional list to pass down to titre solving function}

\item{DOBs}{NumericVector, vector of ages for each individual}

\item{solve_likelihood}{bool, if FALSE does not solve likelihood when calculating acceptance probability}

\item{total_alive}{int, if a positive number, uses this rather than the vector of alive individuals for the infection history prior}

\item{temp}{double, temperature for parallel tempering MCMC}

\item{n_years_samp}{int, for each individual, how many time periods to resample infections for?}
}
\value{
a matrix of 1s and 0s corresponding to the infection histories for all individuals
}
\description{
Proposes a new matrix of infection histories by sampling from the prior on an individual's infection presence/absence in a particular time period, conditional on all other individuals in that time period. This allows us to integrate out the infection probability term for each time period. Should look at \code{\link{create_posterior_func}} for more details about the input parameters.
}
\seealso{
Other infection_history_proposal: \code{\link{inf_hist_prop_cpp}},
  \code{\link{infection_history_proposal_gibbs_fast}}
}
\concept{infection_history_proposal}
