% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{infection_history_symmetric}
\alias{infection_history_symmetric}
\title{Brute force infection history proposal}
\usage{
infection_history_symmetric(newInfHist, sampledIndivs, ageMask, moveSizes,
  nInfs, randNs)
}
\arguments{
\item{newInfHist}{a matrix of infection histories - rows for individuals, columns for infection epochs. Contents should be 1s and 0s}

\item{sampledIndivs}{a vector of indices describing rows in the infection history matrix that should be updated}

\item{ageMask}{a vector (one value for each individual) giving the first infection epoch that an individual could have been exposed in. That is, if an individual was born in the 7th epoch, their entry in ageMask would be 7.}

\item{moveSizes}{when performing a move step, how far should two epochs be swapped?}

\item{nInfs}{number of infection epochs to flip}

\item{randNs}{pre-computed random numbers (0-1) for each individual, deciding whether to do a flip or swap}
}
\value{
a matrix of infection histories matching the input newInfHist
}
\description{
Performs a flipping/swapping infection history update step for a matrix of infection histories. 50/50 chance of performing a flip or a swap
}

