% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{infection_model_indiv}
\alias{infection_model_indiv}
\title{Model function}
\usage{
infection_model_indiv(theta, infectionHistory, infectionTimes,
  infectionMapIndices, samplingTime, measurementMapIndices, antigenicMapLong,
  antigenicMapShort, numberStrains)
}
\arguments{
\item{theta}{NumericVector, the named vector of model parameters}

\item{infectionHistory}{NumericVector, the vector of 1s and 0s showing presence/absence of infection for each possible time.}

\item{samplingTime}{double, the real time that the sample was taken}

\item{antigenicMapLong}{NumericVector, the collapsed cross reactivity map for long term boosting, after multiplying by sigma1}

\item{antigenicMapShort}{NumericVector, the collapsed cross reactivity map for short term boosting, after multiplying by sigma2}

\item{strainIsolationTimes}{NumericVector, the vector of times at which each virus strain circulated}
}
\value{
NumericVector of predicted titres for each strainIsolationTime
}
\description{
The main model solving function for a single individual.
NOTES:
- Do we want infection history to be a vector of infection times?
- Treat the contents of infectionHistory as a parameter (ie. exposure type)
}
