% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood_func_fast}
\alias{likelihood_func_fast}
\title{Fast observation error function
 Calculate the probability of a set of observed antibody levels given a corresponding set of predicted antibody levels.}
\usage{
likelihood_func_fast(theta, obs, predicted_antibody_levels)
}
\arguments{
\item{theta}{NumericVector, a named parameter vector giving the normal distribution standard deviation and the max observable antibody level}

\item{obs}{NumericVector, the vector of observed log antibody levels}

\item{predicted_antibody_levels}{NumericVector, the vector of predicted log antibody levels}
}
\value{
a likelihood for each observed antibody level
}
\description{
Fast observation error function
 Calculate the probability of a set of observed antibody levels given a corresponding set of predicted antibody levels.
}
\seealso{
Other likelihood_functions:
\code{\link{Fast observation error function continuous}}
}
\concept{likelihood_functions}
\keyword{internal}
