% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood_func_fast}
\alias{likelihood_func_fast}
\title{Fast observation error function
 Calculate the probability of a set of observed titres given a corresponding set of predicted titres. FAST IMPLEMENTATION}
\usage{
likelihood_func_fast(theta, obs, predicted_titres)
}
\arguments{
\item{theta}{NumericVector, a named parameter vector giving the normal distribution standard deviation and the max observable titre}

\item{obs}{NumericVector, the vector of observed log titres}

\item{predicted_titres}{NumericVector, the vector of predicted log titres}

\item{a}{vector of same length as the input data giving the probability of observing each observation given the predictions}
}
\value{
a likelihood for each observed titre
}
\description{
Fast observation error function
 Calculate the probability of a set of observed titres given a corresponding set of predicted titres. FAST IMPLEMENTATION
}
\seealso{
Other likelihood_functions: 
\code{\link{Fast observation error function continuous}}
}
\concept{likelihood_functions}
