% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{load_antibody_data}
\alias{load_antibody_data}
\title{Load serological antibody data from CSV file}
\usage{
load_antibody_data(location = getwd())
}
\arguments{
\item{location}{A character string specifying the path to the directory where the function should search for the file. Defaults to the current working directory.}
}
\value{
A data frame with antibody data, or `NULL` if no matching file is found.
}
\description{
Searches the specified working directory for a csv file with "antibody_data.csv" or "titer_dat.csv" in the name and reads in the first file found as a data frame.
}
\examples{
\dontrun{
  antibody_data <- load_antibody_data()
  antibody_data <- load_antibody_data("data/")
}

}
\seealso{
[load_start_tab()], [load_mcmc_chains()]

Other load_data_functions: 
\code{\link{load_antigenic_map}()},
\code{\link{load_infection_chains}()},
\code{\link{load_mcmc_chains}()},
\code{\link{load_start_tab}()},
\code{\link{load_theta_chains}()}
}
\concept{load_data_functions}
