% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{load_antigenic_map}
\alias{load_antigenic_map}
\title{Load in antigenic map file from CSV}
\usage{
load_antigenic_map(location = getwd())
}
\arguments{
\item{location}{A character string specifying the path to the directory where the function should search for the file. Defaults to the current working directory.}
}
\value{
A data frame containing the antigenic map. Returns `NULL` if no matching file is found.
}
\description{
Reads in an antigenic map from the first CSV file found matching the pattern `*_antigenic_map.csv`.
}
\examples{
\dontrun{
  antigenic_map <- load_antigenic_map_file()
  antigenic_map <- load_antigenic_map_file("data/")
}
}
\seealso{
[load_antibody_data()], [load_start_tab()]

Other load_data_functions: 
\code{\link{load_antibody_data}()},
\code{\link{load_infection_chains}()},
\code{\link{load_mcmc_chains}()},
\code{\link{load_start_tab}()},
\code{\link{load_theta_chains}()}
}
\concept{load_data_functions}
