% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_antigenic_map_file}
\alias{load_antigenic_map_file}
\title{Read in antigenic_map}
\usage{
load_antigenic_map_file(location = getwd())
}
\arguments{
\item{location}{string giving the relative file path to be searched}
}
\value{
data frame, the first found antigenic map tables in the current working directory
}
\description{
Searches the working directory for a file with "antigenic_map.csv" in the name and reads in the first one found
}
\examples{
\dontrun{load_antigenic_map_file()}
}
\seealso{
Other load_data_functions: 
\code{\link{load_infection_chains}()},
\code{\link{load_mcmc_chains}()},
\code{\link{load_start_tab}()},
\code{\link{load_theta_chains}()},
\code{\link{load_titre_dat}()}
}
\concept{load_data_functions}
