% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{load_infection_chains}
\alias{load_infection_chains}
\title{Load MCMC chains from CSV files for infection histories}
\usage{
load_infection_chains(
  location = getwd(),
  thin = 1,
  burnin = 0,
  chain_subset = NULL,
  verbose = TRUE
)
}
\arguments{
\item{location}{Character string path to the directory to search. Defaults to `getwd()`.}

\item{thin}{Integer; thinning interval to apply to the MCMC samples. Defaults to `1` (no thinning).}

\item{burnin}{Integer; number of initial samples to discard. Defaults to `0`.}

\item{chain_subset}{if not NULL, a vector of indices to only load and store a subset of the chains detected. eg. chain_subset = 1:3 means that only the first 3 detected files will be processed.}

\item{verbose}{Logical; whether to print progress messages}
}
\value{
a list with a) a list of each chain as a data table separately; b) a combined data table, indexing each iteration by which chain it comes from
}
\description{
Searches the given working directory for MCMC outputs from \code{\link{serosolver}} ending "infection_histories.csv", loads these in, subsets for burn in and thinning, and formats as both lists and a combined data table.
}
\examples{
\dontrun{load_infection_chains(thin=10,burnin=5000,chain_subset=1:3)}
}
\seealso{
[load_mcmc_chains()], [load_start_tab()], [plot_infection_histories()]

Other load_data_functions:
\code{\link[=load_antibody_data]{load_antibody_data()}},
\code{\link[=load_antigenic_map]{load_antigenic_map()}},
\code{\link[=load_mcmc_chains]{load_mcmc_chains()}},
\code{\link[=load_start_tab]{load_start_tab()}},
\code{\link[=load_theta_chains]{load_theta_chains()}}
}
\concept{load_data_functions}
