% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_infection_chains}
\alias{load_infection_chains}
\title{Load MCMC chains for infection histories}
\usage{
load_infection_chains(location = getwd(), thin = 1, burnin = 0,
  chain_subset = NULL)
}
\arguments{
\item{location}{defaults to current working directory. Where to look for MCMC chains? These are files ending in "_infection_histories.csv"}

\item{thin}{thin the chains by every thin'th sample}

\item{burnin}{discard the first burnin samples}

\item{chain_subset}{if not NULL, a vector of indices to only load and store a subset of the chains detected. eg. chain_subset = 1:3 means that only the first 3 detected files will be processed.}
}
\value{
a list with a) a list of each chain as a data table seperately; b) a combined data table, indexing each iteration by which chain it comes from
}
\description{
Searches the given working directory for MCMC outputs from \code{\link{run_MCMC}}, loads these in, subsets for burn in and thinning, and formats as both lists and a combined data table.
}
