% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{load_mcmc_chains}
\alias{load_mcmc_chains}
\title{Load MCMC chains for the antibody kinetics parameters and infection histories from CSV outputs}
\usage{
load_mcmc_chains(
  location = getwd(),
  par_tab = NULL,
  estimated_only = FALSE,
  thin = 1,
  burnin = 0,
  convert_mcmc = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{location}{Character string path to the directory to search. Defaults to `getwd()`.}

\item{par_tab}{Data frame; optional parameter table for filtering. Defaults to `NULL`.}

\item{estimated_only}{if TRUE, only returns free model parameters (par_tab$fixed == 0) if par_tab specified}

\item{thin}{Integer; thinning interval to apply to the MCMC samples. Defaults to `1` (no thinning).}

\item{burnin}{Integer; number of initial samples to discard. Defaults to `0`.}

\item{convert_mcmc}{Logical; whether to convert output to `coda::mcmc.list`. Defaults to `TRUE`.}

\item{verbose}{Logical; whether to print progress messages}
}
\value{
a list of the concatenated and individual chains (4 elements, either data frames of coda::mcmc objects)
}
\description{
Reads in all MCMC chains for theta and infection histories from the specified directory matcing files ending "_chain.csv", adding in the total number of infections
}
\examples{
\dontrun{load_mcmc_chains(par_tab=par_tab, estimated_only=TRUE,thin=10,burnin=5000,convert_mcmc=TRUE)}
}
\seealso{
Other load_data_functions:
\code{\link[=load_antibody_data]{load_antibody_data()}},
\code{\link[=load_antigenic_map]{load_antigenic_map()}},
\code{\link[=load_infection_chains]{load_infection_chains()}},
\code{\link[=load_start_tab]{load_start_tab()}},
\code{\link[=load_theta_chains]{load_theta_chains()}}
}
\concept{load_data_functions}
