% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_mcmc_chains}
\alias{load_mcmc_chains}
\title{Combine theta and infection history chains}
\usage{
load_mcmc_chains(location = getwd(), par_tab = NULL, unfixed = TRUE,
  thin = 1, burnin = 0, convert_mcmc = FALSE)
}
\arguments{
\item{location}{defaults to current working directory. Where to look for MCMC chains? These are files ending in "_chain.csv"}

\item{par_tab}{if not NULL, can use this to only extract free model parameters}

\item{unfixed}{if TRUE, only returns free model parameters if par_tab specified}

\item{thin}{thin the chains by every thin'th sample}

\item{burnin}{discard the first burnin samples}

\item{convert_mcmc}{if TRUE, converts everything to MCMC objects (from the coda package)}
}
\value{
a list of the concatenated and individual chains
}
\description{
Reads in both the MCMC chains for theta and infection histories, adding in the total number of infections
}
