% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_mcmc_chains}
\alias{load_mcmc_chains}
\title{Combine theta and infection history chains}
\usage{
load_mcmc_chains(
  location = getwd(),
  par_tab = NULL,
  estimated_only = FALSE,
  thin = 1,
  burnin = 0,
  convert_mcmc = FALSE
)
}
\arguments{
\item{location}{defaults to current working directory. Gives relative file path to look for files ending in "_chain.csv"}

\item{par_tab}{if not NULL, can use this to only extract free model parameters}

\item{estimated_only}{if TRUE, only returns free model parameters (par_tab$fixed == 0) if par_tab specified}

\item{thin}{thin the chains by every thin'th sample}

\item{burnin}{discard the first burnin samples from the MCMC chain}

\item{convert_mcmc}{if TRUE, converts everything to MCMC objects (from the `coda` R package)}
}
\value{
a list of the concatenated and individual chains (4 elements, either data frames of coda::mcmc objects)
}
\description{
Reads in all MCMC chains for theta and infection histories from the specified directory, adding in the total number of infections
}
\examples{
\dontrun{load_mcmc_chains(par_tab=par_tab, estimated_only=TRUE,thin=10,burnin=5000,convert_mcmc=TRUE)}
}
\seealso{
Other load_data_functions: 
\code{\link{load_antibody_data}()},
\code{\link{load_antigenic_map_file}()},
\code{\link{load_infection_chains}()},
\code{\link{load_start_tab}()},
\code{\link{load_theta_chains}()}
}
\concept{load_data_functions}
