% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_start_tab}
\alias{load_start_tab}
\title{Read in start_tab}
\usage{
load_start_tab(location = getwd())
}
\value{
one of the starting parameter tables in the current working directory
}
\description{
Searches the working directory for a file called "start_tab" and reads this in
}
