% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{load_theta_chains}
\alias{load_theta_chains}
\title{Load MCMC chains from CSV files for antibody kinetics parameters (theta)}
\usage{
load_theta_chains(
  location = getwd(),
  par_tab = NULL,
  estimated_only = TRUE,
  thin = 1,
  burnin = 0,
  convert_mcmc = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{location}{Character string path to the directory containing the chain files. Defaults to `getwd()`.}

\item{par_tab}{Data frame; optional model control table used to select estimated parameters. Defaults to `NULL`.}

\item{estimated_only}{if TRUE and `par_tab` is supplied, only returns parameters with `par_tab$fixed == 0`.}

\item{thin}{Integer; keeps every `thin`th saved MCMC sample. Defaults to `1` (no thinning).}

\item{burnin}{Integer; discards samples with `samp_no <= burnin`. Defaults to `0`.}

\item{convert_mcmc}{Logical; if TRUE, also converts the chains to `coda::mcmc` or `coda::mcmc.list` objects. Defaults to `TRUE`.}

\item{verbose}{Logical; whether to print progress messages.}
}
\value{
A list with two entries: `list`, containing the chains separately, and `chain`, containing the combined chains with a `chain_no` column. If `convert_mcmc = TRUE`, these entries are converted to `coda::mcmc.list` and `coda::mcmc` objects.
}
\description{
Searches the given working directory for MCMC outputs from \code{\link{serosolver}} matching "_chain.csv", reads these in, subsets for burn in and thinning, and formats as both lists and a combined data frame.
}
\examples{
\dontrun{load_theta_chains(location="mcmc_chains", par_tab=par_tab, estimated_only=TRUE,thin=10,burnin=5000,convert_mcmc=TRUE)}
}
\seealso{
Other load_data_functions:
\code{\link[=load_antibody_data]{load_antibody_data()}},
\code{\link[=load_antigenic_map]{load_antigenic_map()}},
\code{\link[=load_infection_chains]{load_infection_chains()}},
\code{\link[=load_mcmc_chains]{load_mcmc_chains()}},
\code{\link[=load_start_tab]{load_start_tab()}}
}
\concept{load_data_functions}
