% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_theta_chains}
\alias{load_theta_chains}
\title{Load MCMC chains for theta}
\usage{
load_theta_chains(location = getwd(), par_tab = NULL, unfixed = TRUE,
  thin = 1, burnin = 0, convert_mcmc = TRUE)
}
\arguments{
\item{location}{defaults to current working directory. Where to look for MCMC chains? These are files ending in "_chain.csv"}

\item{par_tab}{if not NULL, can use this to only extract free model parameters}

\item{unfixed}{if TRUE, only returns free model parameters if par_tab specified}

\item{thin}{thin the chains by every thin'th sample}

\item{burnin}{discard the first burnin samples}

\item{convert_mcmc}{if TRUE, converts everything to MCMC objects (from the coda package)}
}
\value{
a list with a) a list of each chain seperately; b) a combined data frame, indexing each iteration by which chain it comes from
}
\description{
Searches the given working directory for MCMC outputs from \code{\link{run_MCMC}}, loads these in, subsets for burn in and thinning, and formats as both lists and a combined data frame.
}
