% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_objects.R
\name{load_theta_chains}
\alias{load_theta_chains}
\title{Load MCMC chains from CSV files for antibody kinetics parameters (theta)}
\usage{
load_theta_chains(
  location = getwd(),
  par_tab = NULL,
  estimated_only = TRUE,
  thin = 1,
  burnin = 0,
  convert_mcmc = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{location}{Character string path to the directory to search. Defaults to `getwd()`.}

\item{par_tab}{Data frame; optional parameter table for filtering. Defaults to `NULL`.}

\item{estimated_only}{if TRUE, only returns free model parameters (par_tab$fixed == 0) if par_tab specified}

\item{thin}{Integer; thinning interval to apply to the MCMC samples. Defaults to `1` (no thinning).}

\item{burnin}{Integer; number of initial samples to discard. Defaults to `0`.}

\item{convert_mcmc}{Logical; whether to convert output to `coda::mcmc.list`. Defaults to `TRUE`.}

\item{verbose}{Logical; whether to print progress messages}
}
\value{
a list with a) a list of each chain separately; b) a combined data frame, indexing each iteration by which chain it comes from
}
\description{
Searches the given working directory for MCMC outputs from \code{\link{serosolver}} matching "_chain.csv", reads these in, subsets for burn in and thinning, and formats as both lists and a combined data frame.
}
\examples{
\dontrun{load_theta_chains(location="mcmc_chains", par_tab=par_tab, estimated_only=TRUE,thin=10,burnin=5000,convert_mcmc=TRUE)}
}
\seealso{
Other load_data_functions: 
\code{\link{load_antibody_data}()},
\code{\link{load_antigenic_map}()},
\code{\link{load_infection_chains}()},
\code{\link{load_mcmc_chains}()},
\code{\link{load_start_tab}()}
}
\concept{load_data_functions}
