% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{load_titre_dat}
\alias{load_titre_dat}
\title{Read in titre_dat}
\usage{
load_titre_dat(location = getwd())
}
\arguments{
\item{location}{string giving the relative file path to be searched}
}
\value{
data frame, the first found starting parameter tables in the current working directory
}
\description{
Searches the working directory for a file with "titre_dat.csv" in the name and reads in the first one found
}
\examples{
\dontrun{load_titre_dat()}
}
\seealso{
Other load_data_functions: 
\code{\link{load_antigenic_map_file}()},
\code{\link{load_infection_chains}()},
\code{\link{load_mcmc_chains}()},
\code{\link{load_start_tab}()},
\code{\link{load_theta_chains}()}
}
\concept{load_data_functions}
