% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{logistic_transform_bounded}
\alias{logistic_transform_bounded}
\title{Transform a value to a bounded logistic scale}
\usage{
logistic_transform_bounded(x, maxX)
}
\arguments{
\item{x}{value to transform}

\item{maxX}{upper bound of the transformed scale}
}
\value{
the transformed value between 0 and `maxX`
}
\description{
Transform a value to a bounded logistic scale
}
