% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{logit_transform_bounded}
\alias{logit_transform_bounded}
\title{Inverse bounded logistic transform}
\usage{
logit_transform_bounded(p, maxX)
}
\arguments{
\item{p}{value on the bounded scale}

\item{maxX}{upper bound of the bounded scale}
}
\value{
the value on the unbounded scale
}
\description{
Inverse bounded logistic transform
}
