% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmc_settings_default}
\alias{mcmc_settings_default}
\title{Default MCMC settings}
\format{A named vector with the following variables:

  \describe{

  \item{iterations}{}

  \item{popt}{}

  \item{popt_hist}{}

  \item{opt_freq}{}

  \item{thin}{}

  \item{adaptive_period}{}

  \item{save_block}{}

  \item{thin_hist}{}

  \item{hist_sample_prob}{}

  \item{switch_sample}{}

  \item{burnin}{}

  \item{inf_propn}{}

  \item{move_size}{}

  \item{hist_opt}{}

  \item{swap_propn}{}

  \item{hist_switch_prob}{}

  \item{year_swap_propn}{}

  \item{propose_from_prior}{}

  }}
\usage{
mcmc_settings_default
}
\description{
Default MCMC settings
}
\keyword{datasets}
