% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{mvr_proposal}
\alias{mvr_proposal}
\title{Multivariate proposal function}
\usage{
mvr_proposal(values, fixed, covMat)
}
\arguments{
\item{values}{the vector of current parameter values}

\item{fixed}{set of flags corresponding to the parameter vector indicating which parameters are fixed}

\item{covMat}{the 2D covariance matrix for all of the parameters}
}
\value{
a parameter vector of a proposed move. Note that these may fall outside the allowable ranges.
}
\description{
Given the current parameters and a covariance matrix, returns a vector for a proposed jump from a multivariate normal distribution
}

