% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pad_alphas_and_betas}
\alias{pad_alphas_and_betas}
\title{Pad par_tab with alpha and betas}
\usage{
pad_alphas_and_betas(par_tab, n_times)
}
\arguments{
\item{par_tab}{as per usual}

\item{n_times}{the number of additional rows to add for each alpha and beta}
}
\description{
Pads par_tab with a new row for each infection epoch, such that each epoch has its own alpha and beta
}
\examples{
n_times <- 40
data(example_par_tab)
new_par_tab <- pad_alphas_and_betas(example_par_tab, n_times)
}
