% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pad_inf_chain}
\alias{pad_inf_chain}
\title{Pad infection history chain}
\usage{
pad_inf_chain(inf_chain, pad_by_group = FALSE)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{serosolver}}}
}
\value{
the same inf_chain that was passed in, but with 0s for missing i/j/samp_no combinations
}
\description{
Given that the MCMC sampler only stores present infections (ie. there are no entries for 0s from the infection history matrix), for some summaries we need to add these 0s back in to avoid bias.
}
