% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pad_infection_model_prior_parameters}
\alias{pad_infection_model_prior_parameters}
\title{Pad par_tab with Beta distribution shape parameters}
\usage{
pad_infection_model_prior_parameters(par_tab, n_times)
}
\arguments{
\item{par_tab}{as per usual}

\item{n_times}{the number of additional rows to add for each alpha and beta}
}
\value{
the parameter table with the additional prior-shape rows
}
\description{
Pads par_tab with a new row for each infection epoch, such that each epoch has its own shape1 and shape2
}
\examples{
\dontrun{
n_times <- 40
data(example_par_tab)
new_par_tab <- pad_infection_model_prior_parameters(example_par_tab, n_times)
}
}
\keyword{internal}
