% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{par_tab_base}
\alias{par_tab_base}
\title{Base parameter table}
\format{A data frame with the following variables:

  \describe{

  \item{name}{Parameter name.}

  \item{value}{Parameter initial value.}

  \item{fixed}{1 - fixed, 0 - needs to be estimated.}

  \item{steps}{How far to step on unit scale to begin with?}

  \item{lower_bound}{Parameters cannot step below this.}

  \item{upper_bound}{Parameters cannot step above this.}

  \item{lower_start}{No idea}

  \item{upper_start}{Haven't a clue}

  \item{type}{Don't know}

  }}
\usage{
par_tab_base
}
\description{
Base parameter table
}
\keyword{datasets}
