% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_other.R
\name{plot_antibody_data}
\alias{plot_antibody_data}
\title{Useful plot for looking at simulated data}
\usage{
plot_antibody_data(
  antibody_data,
  infection_histories,
  possible_exposure_times,
  n_indivs,
  start_inf = NULL,
  study_design = "multi-antigen"
)
}
\arguments{
\item{antibody_data}{the data frame of antibody data}

\item{infection_histories}{the infection history matrix}

\item{possible_exposure_times}{the vector of times at which individuals could be infected}

\item{n_indivs}{how many individuals to plot}

\item{start_inf}{if not NULL, plots the infection history matrix used as the starting point in the MCMC chain}

\item{study_design}{default "multi-antigen" facets by sample time. "single-antigen" gives sample time on the x-axis and colours by biomarker_id}
}
\value{
a ggplot object
}
\description{
Plots measured titres and known infection histories for all individuals, faceted by sample time (multi-antigen panel) or biomarker_id variable (longitudinal single antigen)
}
\examples{
\dontrun{
data(example_antibody_data)
data(example_inf_hist)
data(example_antigenic_map)
possible_exposure_times <- example_antigenic_map$inf_times
plot_antibody_data(example_antibody_data, example_inf_hist, possible_exposure_times, 5)
}
}
\seealso{
Other infection_history_plots: 
\code{\link{calculate_infection_history_statistics}()},
\code{\link{plot_cumulative_infection_histories}()},
\code{\link{plot_individual_number_infections}()},
\code{\link{plot_infection_history_chains_indiv}()},
\code{\link{plot_infection_history_chains_time}()},
\code{\link{plot_model_fits}()},
\code{\link{plot_posteriors_infhist}()},
\code{\link{plot_total_number_infections}()}
}
\concept{infection_history_plots}
