% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_other.R
\name{plot_antibody_data}
\alias{plot_antibody_data}
\title{Plot raw data}
\usage{
plot_antibody_data(
  antibody_data,
  possible_exposure_times,
  n_indivs,
  infection_histories = NULL,
  study_design = "cross-sectional",
  measurement_ranges = NULL
)
}
\arguments{
\item{antibody_data}{the data frame of antibody data}

\item{possible_exposure_times}{the vector of times at which individuals could be infected}

\item{n_indivs}{integer of how many individuals to plot, or vector of which individuals to plot}

\item{infection_histories}{the infection history matrix}

\item{study_design}{default "cross-sectional" facets by sample time. "longitudinal" gives sample time on the x-axis and colours by `biomarker_id`}

\item{measurement_ranges}{data frame or tibble stating for each `biomarker_group` the `min_measurement` and `max_measurement`. If NULL, this is extracted from `antibody_data`}
}
\value{
a ggplot object
}
\description{
Plots measured antibody measurements and known infection histories for all individuals, faceted by sample time (multi-antigen panel) or biomarker_id variable (longitudinal single antigen)
}
\examples{
\dontrun{
data(example_antibody_data)
data(example_inf_hist)
data(example_antigenic_map)
possible_exposure_times <- example_antigenic_map$inf_times
plot_antibody_data(example_antibody_data, possible_exposure_times, 5, example_inf_hist)
}
}
\seealso{
Other infection_history_plots: 
\code{\link{calculate_infection_history_statistics}()},
\code{\link{plot_antibody_predictions}()},
\code{\link{plot_cumulative_infection_histories}()},
\code{\link{plot_estimated_antibody_model}()},
\code{\link{plot_individual_number_infections}()},
\code{\link{plot_infection_history_chains_indiv}()},
\code{\link{plot_infection_history_chains_time}()},
\code{\link{plot_infection_history_posteriors}()},
\code{\link{plot_model_fits}()},
\code{\link{plot_total_number_infections}()}
}
\concept{infection_history_plots}
