% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_antibody_model.R
\name{plot_antibody_dependent_boosting}
\alias{plot_antibody_dependent_boosting}
\title{Antibody dependent boosting relationship}
\usage{
plot_antibody_dependent_boosting(chain, n, titres = seq(0, 8, by = 0.1))
}
\arguments{
\item{chain}{the MCMC chain}

\item{n}{number of samples to take}

\item{titres}{the vector of titres to calculate boosting values at}
}
\value{
a data frame of quantiles for the inferred boost from different titre levels
}
\description{
Calculates the inferred antibody dependent boosting relationship from the MCMC chain
}
