% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infection_histories.R
\name{plot_attack_rates}
\alias{plot_attack_rates}
\title{Plot historical attack rates}
\usage{
plot_attack_rates(
  infection_histories,
  antibody_data,
  possible_exposure_times,
  n_alive = NULL,
  ymax = 1,
  buckets = 1,
  pad_chain = TRUE,
  true_ar = NULL,
  by_group = FALSE,
  group_subset = NULL,
  cumulative = FALSE,
  add_box = FALSE
)
}
\arguments{
\item{ymax}{Numeric. the maximum y value to put on the axis. Default = 1.}

\item{cumulative}{if TRUE, plots the cumulative attack rate}

\item{resolution}{Integer. How many buckets of time is each year split into? ie. 12 for monthly data, 4 for quarterly etc. Default = 1 for annual.}
}
\value{
a ggplot2 object with the inferred attack rates for each potential epoch of circulation
}
\description{
Plots inferred historical attack rates from the MCMC output on infection histories.
}
