% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_attack_rates}
\alias{plot_attack_rates}
\title{Plot historical attack rates}
\usage{
plot_attack_rates(infectionHistories, dat, ages, yearRange,
  n_alive = NULL, pointsize = 1, fatten = 1)
}
\arguments{
\item{infectionHistories}{the MCMC chain for infection histories}

\item{dat}{the data frame of titre data}

\item{ages}{the data frame of ages for each individual, with columns for individual and DOB (date of birth)}

\item{yearRange}{vector of the first and last epoch of potential circulation}

\item{n_alive}{vector with the number of people alive in each year of circulation. Can be left as NULL, and ages will be used to infer this}

\item{pointsize}{Numeric - how big should each point be?}

\item{fatten}{Numeric - fatten parameter for ggplot pointrange}
}
\value{
a ggplot2 object with the inferred attack rates for each potential epoch of circulation
}
\description{
Plots inferred historical attack rates from the MCMC output on infection histories
}
