% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_data}
\alias{plot_data}
\title{Useful plot for looking at simulated data}
\usage{
plot_data(data, infectionHistories, strainIsolationTimes, n_samps,
  startInf = NULL)
}
\description{
Useful plot for looking at simulated data
}

