% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_antibody_model.R
\name{plot_estimated_antibody_model}
\alias{plot_estimated_antibody_model}
\title{Plots estimated antibody kinetics model}
\usage{
plot_estimated_antibody_model(
  chain,
  antibody_data = NULL,
  demographics = NULL,
  antigenic_map = NULL,
  possible_exposure_times = NULL,
  par_tab = NULL,
  nsamp = 1000,
  measurement_bias = NULL,
  solve_times = seq(1, 30, by = 1),
  data_type = 1,
  settings = NULL,
  by_group = TRUE,
  add_prediction_intervals = FALSE,
  exponential_waning = NULL
)
}
\arguments{
\item{chain}{the full MCMC chain to generate antibody level trajectories from}

\item{antibody_data}{the data frame of antibody level data}

\item{antigenic_map}{(optional) a data frame of antigenic x and y coordinates. Must have column names: x_coord; y_coord; inf_times. See \code{\link{example_antigenic_map}}}

\item{possible_exposure_times}{(optional) if no antigenic map is specified, this argument gives the vector of times at which individuals can be infected}

\item{par_tab}{the table controlling the parameters in the MCMC chain}

\item{nsamp}{number of samples to take from posterior}

\item{measurement_bias}{default NULL, optional data frame giving the index of `rho` that each biomarker_id and biomarker_group which uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters}

\item{solve_times}{vector of times to solve model over}

\item{data_type}{integer, currently accepting 1 or 2. Set to 1 for discretized, bounded data, or 2 for continuous, bounded data.}

\item{settings}{if not NULL, list of serosolver settings as returned from the main serosolver function}

\item{exponential_waning}{if TRUE, assumes exponential rather than linear waning}
}
\value{
a ggplot2 object giving model-predicted antibody level and predicted observations over time since infection
}
\description{
Plots estimated antibody kinetics model
}
\seealso{
Other infection_history_plots: 
\code{\link{calculate_infection_history_statistics}()},
\code{\link{plot_antibody_data}()},
\code{\link{plot_antibody_predictions}()},
\code{\link{plot_cumulative_infection_histories}()},
\code{\link{plot_individual_number_infections}()},
\code{\link{plot_infection_history_chains_indiv}()},
\code{\link{plot_infection_history_chains_time}()},
\code{\link{plot_model_fits}()},
\code{\link{plot_posteriors_infhist}()},
\code{\link{plot_total_number_infections}()}
}
\concept{infection_history_plots}
