% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_infection_histories}
\alias{plot_infection_histories}
\title{Plots infection histories and titre model fits}
\usage{
plot_infection_histories(chain, infectionHistories, dat, individuals,
  antigenicMap, ages, parTab, nsamp = 100, mu_indices = NULL,
  measurement_indices = NULL)
}
\arguments{
\item{chain}{the full MCMC chain to generate titre trajectories from}

\item{infectionHistories}{the MCMC chain for infection histories}

\item{dat}{the data frame of titre data}

\item{individuals}{the subset of individuals to generate credible intervals for}

\item{antigenicMap}{the unmelted antigenic map}

\item{ages}{the data frame of ages for each individual, with columns for individual and DOB (date of birth)}

\item{parTab}{the table controlling the parameters in the MCMC chain}

\item{nsamp}{number of samples to take from posterior}

\item{mu_indices}{if random effects on boosting parameter, mu, this specifies which entry in the parameter table corresponds to which year. See \code{\link{run_MCMC}}}

\item{measurement_indices}{default NULL, optional vector giving the index of `measurement_bias` that each strain uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clus}
}
\value{
a ggplot2 object
}
\description{
Given outputs from an MCMC run and the data used for fitting, generates an NxM matrix of plots where N is the number of individuals to be plotted and M is the range of sampling times. Where data are available, plots the observed titres and model predicted trajectories
}
