% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_infection_histories}
\alias{plot_infection_histories}
\title{Plots infection histories and titre model fits}
\usage{
plot_infection_histories(chain, infectionHistories, dat, individuals,
  antigenicMap, ages, parTab, nsamp = 100)
}
\arguments{
\item{chain}{the full MCMC chain to generate titre trajectories from}

\item{infectionHistories}{the MCMC chain for infection histories}

\item{dat}{the data frame of titre data}

\item{individuals}{the subset of individuals to generate credible intervals for}

\item{antigenicMap}{the unmelted antigenic map}

\item{ages}{the data frame of ages for each individual, with columns for individual and DOB (date of birth)}

\item{parTab}{the table controlling the parameters in the MCMC chain}

\item{nsamp}{number of samples to take from posterior}
}
\value{
a ggplot2 object
}
\description{
Given outputs from an MCMC run and the data used for fitting, generates an NxM matrix of plots where N is the number of individuals to be plotted and M is the range of sampling times. Where data are available, plots the observed titres and model predicted trajectories
}

