% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_infection_history_chains_indiv}
\alias{plot_infection_history_chains_indiv}
\title{Plot MCMC trace for infections per individual}
\usage{
plot_infection_history_chains_indiv(
  inf_chain,
  burnin = 0,
  indivs = NULL,
  pad_chain = TRUE
)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{run_MCMC}}}

\item{burnin}{optionally remove all sampno < burnin from the chain}

\item{indivs}{vector of integers, if not NULL, only plots a subset of individuals (where 1 is the first individual)}

\item{pad_chain}{if TRUE, pads the infection history MCMC chain to have entries for non-infection events}
}
\value{
a list of two ggplot objects - the MCMC trace and MCMC densities
}
\description{
Plot MCMC trace for infections per individual
}
\examples{
\dontrun{
data(example_inf_chain)
plot_infection_history_chains_indiv(example_inf_chain, 0, 1:10, FALSE)
}
}
\seealso{
\code{\link{plot_infection_history_chains_indiv}}

Other infection_history_plots: 
\code{\link{calculate_infection_history_statistics}()},
\code{\link{generate_cumulative_inf_plots}()},
\code{\link{plot_data}()},
\code{\link{plot_infection_histories_long}()},
\code{\link{plot_infection_histories}()},
\code{\link{plot_infection_history_chains_time}()},
\code{\link{plot_number_infections}()},
\code{\link{plot_posteriors_infhist}()},
\code{\link{plot_total_number_infections}()}
}
\concept{infection_history_plots}
