% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_infection_history_chains_indiv}
\alias{plot_infection_history_chains_indiv}
\title{Plot MCMC trace for infections per individual}
\usage{
plot_infection_history_chains_indiv(inf_chain, burnin = 0,
  indivs = NULL)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{run_MCMC}}}

\item{burnin}{optionally remove all sampno < burnin from the chain}

\item{indivs}{vector of integers, if not NULL, only plots a subset of individuals (where 1 is the first individual)}
}
\value{
a list of two ggplot objects - the MCMC trace and MCMC densities
}
\description{
Plot MCMC trace for infections per individual
}
\seealso{
\code{\link{plot_infection_history_chains_indiv}}
}
