% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_infection_history_chains_time}
\alias{plot_infection_history_chains_time}
\title{Plot MCMC trace for infections per year}
\usage{
plot_infection_history_chains_time(inf_chain, burnin = 0, years = NULL,
  n_alive = NULL, pad_chain = TRUE)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{run_MCMC}}}

\item{burnin}{optionally remove all sampno < burnin from the chain}

\item{years}{vector of integers, if not NULL, only plots a subset of years (where 1 is the first year eg. 1968)}

\item{n_alive}{if not NULL, then divides number of infections per year by number alive to give attack rates rather than total infections}

\item{pad_chain}{if TRUE, pads the infection history MCMC chain to have entries for non-infection events}
}
\value{
a list of two ggplot objects - the MCMC trace and MCMC densities
}
\description{
Plot MCMC trace for infections per year
}
\examples{
data(example_inf_chain)
data(example_titre_dat)
data(example_antigenic_map)
times <- example_antigenic_map$inf_years
n_alive_group <- get_n_alive_group(example_titre_dat, strain_isolation_times,melt_dat = TRUE)
n_alive_group$j <- strain_isolation_times[n_alive_group$j]
plot_infection_history_chains_time(example_inf_chain, 0, sample(1:length(times),10),n_alive,FALSE)
}
\seealso{
\code{\link{plot_infection_history_chains_indiv}}

Other infection_history_plots: \code{\link{calculate_infection_history_statistics}},
  \code{\link{generate_cumulative_inf_plots}},
  \code{\link{plot_data}},
  \code{\link{plot_infection_histories}},
  \code{\link{plot_infection_history_chains_indiv}},
  \code{\link{plot_number_infections}},
  \code{\link{plot_posteriors_infhist}},
  \code{\link{plot_total_number_infections}}
}
\concept{infection_history_plots}
