% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_antibody_model.R
\name{plot_model_fits}
\alias{plot_model_fits}
\title{Plots infection histories and antibody model}
\usage{
plot_model_fits(
  chain,
  infection_histories,
  antibody_data = NULL,
  demographics = NULL,
  individuals,
  par_tab = NULL,
  antigenic_map = NULL,
  possible_exposure_times = NULL,
  nsamp = 1000,
  known_infection_history = NULL,
  measurement_bias = NULL,
  p_ncol = max(1, floor(length(individuals)/2)),
  data_type = 1,
  expand_to_all_times = FALSE,
  expand_to_all_biomarker_ids = FALSE,
  orientation = "cross-sectional",
  subset_biomarker_ids = NULL,
  subset_biomarker_groups = NULL,
  start_level = "none",
  settings = NULL,
  exponential_waning = NULL,
  verbose = FALSE
)
}
\arguments{
\item{chain}{the full MCMC chain to generate antibody level trajectories from}

\item{infection_histories}{the MCMC chain for infection histories}

\item{antibody_data}{the data frame of antibody level data}

\item{individuals}{the subset of individuals to generate credible intervals for}

\item{par_tab}{the table controlling the parameters in the MCMC chain}

\item{antigenic_map}{(optional) a data frame of antigenic x and y coordinates. Must have column names: x_coord; y_coord; inf_times. See \code{\link{example_antigenic_map}}}

\item{possible_exposure_times}{(optional) if no antigenic map is specified, this argument gives the vector of times at which individuals can be infected}

\item{nsamp}{number of samples to take from posterior}

\item{known_infection_history}{nxm matrix of known infection histories}

\item{measurement_bias}{default NULL, optional data frame giving the index of `rho` that each biomarker_id and biomarker_group which uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters}

\item{p_ncol}{integer giving the number of columns of subplots to create if using orientation = "longitudinal"}

\item{data_type}{integer, currently accepting 1 or 2. Set to 1 for discretized, bounded data, or 2 for continuous, bounded data.}

\item{expand_to_all_times}{TRUE/FALSE value. If TRUE, solves antibody level predictions for all possible infection times (i.e., for the range in possible_exposure_times). If left FALSE, then only solves for the infections times at which a antibody level against the circulating biomarker_id was measured in antibody_data.}

\item{expand_to_all_biomarker_ids}{if TRUE, solves predictions for all biomarker IDs in the antigenic map while retaining the sample times in antibody_data}

\item{orientation}{either "cross-sectional" or "longitudinal"}

\item{subset_biomarker_ids}{if not NULL, then a vector giving the entries of biomarker_id to include in the longitudinal plot}

\item{start_level}{"none"}

\item{settings}{if not NULL, list of serosolver settings as returned from the main serosolver function}

\item{exponential_waning}{if TRUE, assumes exponential rather than linear waning}
}
\value{
a ggplot2 object
}
\description{
Given outputs from an MCMC run and the data used for fitting, generates an NxM matrix of plots where N is the number of individuals to be plotted and M is the range of sampling times. Where data are available, plots the observed antibody measurements and model predicted trajectories. Unlike plot_infection_histories_cross_sectional, places biomarker\_id on the x-axis and facets by sample time and individual.
}
\examples{
\dontrun{
data(example_theta_chain)
data(example_inf_chain)
data(example_antibody_data)
data(example_antigenic_map)
data(example_par_tab)

model_fit_plot <- plot_model_fits(example_theta_chain, example_inf_chain, example_antibody_data, 
                                           1:10, example_antigenic_map, example_par_tab,orientation="longitudinal")
}
}
\seealso{
Other infection_history_plots:
\code{\link[=calculate_infection_history_statistics]{calculate_infection_history_statistics()}},
\code{\link[=plot_antibody_data]{plot_antibody_data()}},
\code{\link[=plot_antibody_predictions]{plot_antibody_predictions()}},
\code{\link[=plot_cumulative_infection_histories]{plot_cumulative_infection_histories()}},
\code{\link[=plot_estimated_antibody_model]{plot_estimated_antibody_model()}},
\code{\link[=plot_individual_number_infections]{plot_individual_number_infections()}},
\code{\link[=plot_infection_history_chains_indiv]{plot_infection_history_chains_indiv()}},
\code{\link[=plot_infection_history_chains_time]{plot_infection_history_chains_time()}},
\code{\link[=plot_infection_history_posteriors]{plot_infection_history_posteriors()}},
\code{\link[=plot_total_number_infections]{plot_total_number_infections()}}
}
\concept{infection_history_plots}
