% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_antibody_model.R
\name{plot_model_fits}
\alias{plot_model_fits}
\title{Plots infection histories and antibody model}
\usage{
plot_model_fits(
  chain,
  infection_histories,
  antibody_data,
  individuals,
  antigenic_map = NULL,
  possible_exposure_times = NULL,
  par_tab,
  nsamp = 100,
  measurement_indices_by_time = NULL,
  p_ncol = length(individuals)/2,
  data_type = 1,
  orientation = "longitudinal"
)
}
\arguments{
\item{chain}{the full MCMC chain to generate antibody level trajectories from}

\item{infection_histories}{the MCMC chain for infection histories}

\item{antibody_data}{the data frame of antibody level data}

\item{individuals}{the subset of individuals to generate credible intervals for}

\item{antigenic_map}{(optional) a data frame of antigenic x and y coordinates. Must have column names: x_coord; y_coord; inf_times. See \code{\link{example_antigenic_map}}}

\item{possible_exposure_times}{(optional) if no antigenic map is specified, this argument gives the vector of times at which individuals can be infected}

\item{par_tab}{the table controlling the parameters in the MCMC chain}

\item{nsamp}{number of samples to take from posterior}

\item{measurement_indices_by_time}{default NULL, optional vector giving the index of `measurement_bias` that each strain uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters}

\item{p_ncol}{integer giving the number of columns of subplots to create if using orientation = "cross-sectional"}

\item{data_type}{integer, currently accepting 1 or 2. Set to 1 for discretized, bounded data, or 2 for continuous, bounded data.}

\item{orientation}{either "longitudinal" or "cross-sectional"}
}
\value{
a ggplot2 object
}
\description{
Given outputs from an MCMC run and the data used for fitting, generates an NxM matrix of plots where N is the number of individuals to be plotted and M is the range of sampling times. Where data are available, plots the observed antibody measurements and model predicted trajectories. Unlike plot_infection_histories_cross_sectional, places biomarker_id on the x-axis and facets by sample time and individual.
}
\examples{
\dontrun{
data(example_theta_chain)
data(example_inf_chain)
data(example_antibody_data)
data(example_antigenic_map)
data(example_par_tab)

model_fit_plot <- plot_model_fits(example_theta_chain, example_inf_chain, example_antibody_data, 
                                           1:10, example_antigenic_map, example_par_tab,orientation="longitudinal")
}
}
\seealso{
Other infection_history_plots: 
\code{\link{calculate_infection_history_statistics}()},
\code{\link{plot_cumulative_infection_histories}()},
\code{\link{plot_data}()},
\code{\link{plot_individual_number_infections}()},
\code{\link{plot_infection_history_chains_indiv}()},
\code{\link{plot_infection_history_chains_time}()},
\code{\link{plot_posteriors_infhist}()},
\code{\link{plot_total_number_infections}()}
}
\concept{infection_history_plots}
