% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_posteriors_infhist}
\alias{plot_posteriors_infhist}
\title{Plot inferred posteriors infection histories}
\usage{
plot_posteriors_infhist(inf_chain, years, n_alive, known_ar = NULL,
  known_infection_history = NULL, burnin = 0, samples = 1000)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{run_MCMC}}}

\item{years}{vector of the epochs of potential circulation}

\item{n_alive}{vector with the number of people alive in each year of circulation.}

\item{known_ar}{data frame of known attack rates, if known.}

\item{known_infection_history}{data frame of known infection histories.}

\item{burnin}{if not already discarded, discard burn in from chain (takes rows where sampno > burnin)}

\item{samples}{how many samples from the chain to take}
}
\value{
a list of ggplot objects and data frame of posterior estimates
}
\description{
Plots and calculates many summary statistics from the infection history MCMC chain
}
