% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_samples_distances}
\alias{plot_samples_distances}
\title{Plot time between serum samples}
\usage{
plot_samples_distances(titre_dat)
}
\arguments{
\item{titre_dat}{the data frame of titre data, including labels for individuals and time sample was taken}
}
\value{
a ggplot2 object
}
\description{
Plot time between serum samples
}
\examples{
data(example_titre_dat)
plot_samples_distances(example_titre_dat)
}
\seealso{
Other theta_plots: \code{\link{plot_posteriors_theta}}
}
\concept{theta_plots}
