% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_other.R
\name{plot_samples_distances}
\alias{plot_samples_distances}
\title{Plot time between serum samples}
\usage{
plot_samples_distances(antibody_data)
}
\arguments{
\item{antibody_data}{the data frame of antibody data, including labels for individuals and time sample was taken}
}
\value{
a ggplot2 object
}
\description{
Plot time between serum samples
}
\examples{
\dontrun{
data(example_antibody_data)
plot_samples_distances(example_antibody_data)
}
}
\concept{theta_plots}
