% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_total_number_infections}
\alias{plot_total_number_infections}
\title{Total number of infections}
\usage{
plot_total_number_infections(inf_chain, pad_chain = TRUE)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{run_MCMC}}}

\item{pad_chain}{if TRUE, pads the infection history MCMC chain to have entries for non-infection events}
}
\value{
two ggplot2 objects
}
\description{
Plots the total number of inferred infections in the MCMC chain as a trace plot and density plot
}
