% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_total_number_infections}
\alias{plot_total_number_infections}
\title{Total number of infections}
\usage{
plot_total_number_infections(inf_chain, pad_chain = TRUE)
}
\arguments{
\item{inf_chain}{the data table with infection history samples from \code{\link{run_MCMC}}}

\item{pad_chain}{if TRUE, pads the infection history MCMC chain to have entries for non-infection events}
}
\value{
two ggplot2 objects
}
\description{
Plots the total number of inferred infections in the MCMC chain as a trace plot and density plot
}
\examples{
data(example_inf_chain)
plot_total_number_infections(example_inf_chain)
}
\seealso{
Other infection_history_plots: \code{\link{calculate_infection_history_statistics}},
  \code{\link{generate_cumulative_inf_plots}},
  \code{\link{plot_data}},
  \code{\link{plot_infection_histories}},
  \code{\link{plot_infection_history_chains_indiv}},
  \code{\link{plot_infection_history_chains_time}},
  \code{\link{plot_number_infections}},
  \code{\link{plot_posteriors_infhist}}
}
\concept{infection_history_plots}
