% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{prob_mus}
\alias{prob_mus}
\title{Prior probability of strain specific boosting}
\usage{
prob_mus(mus, pars)
}
\arguments{
\item{mus}{the vector of boosting parameters}

\item{pars}{the vector of other model parameters including mu_mean and mu_sd}
}
\value{
a single log prior probability
}
\description{
Function find the prior probability of a set of boosting parameters assuming that boosting is drawn from a normal distribution (random effects)
}
\seealso{
Other priors: 
\code{\link{calc_phi_probs_indiv}()},
\code{\link{calc_phi_probs_spline}()},
\code{\link{calc_phi_probs}()},
\code{\link{create_prior_mu}()},
\code{\link{create_prob_shifts}()},
\code{\link{find_beta_prior_mode}()},
\code{\link{find_beta_prior_with_mean_var}()},
\code{\link{find_beta_prior_with_mean}()},
\code{\link{fit_beta_prior}()},
\code{\link{fit_normal_prior}()},
\code{\link{inf_mat_prior}()},
\code{\link{infection_history_prior}()},
\code{\link{prob_shifts}()}
}
\concept{priors}
