% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{prob_shifts}
\alias{prob_shifts}
\title{Prior on measurement shifts}
\usage{
prob_shifts(rhos, pars)
}
\arguments{
\item{rhos}{vector of measurement shifts}

\item{pars}{vector of parameters, including rho_mean and rho_sd for the normal distribution}
}
\value{
a single prior probability
}
\description{
Assumes measurement shifts are drawn from a normal distribution (random effects) with given standard deviation and mean. Code is commented out to assume log normally distributied
}
