% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{qbb}
\alias{qbb}
\title{Beta binomial quantile function}
\usage{
qbb(p, N, u, v)
}
\arguments{
\item{p}{vector of probabilities}

\item{N}{the number of trials}

\item{u}{equivalent to alpha in description}

\item{v}{equivalent to beta in description}
}
\value{
the beta binomial quantile function
}
\description{
Beta binomial quantile function
}
