% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{r_likelihood}
\alias{r_likelihood}
\title{Likelihood function given data}
\usage{
r_likelihood(expected, data, theta, expected_indices = NULL,
  measurement_shifts = NULL)
}
\arguments{
\item{expected}{vector of expected HI titres}

\item{data}{vector of observed discrete HI titres}

\item{expected_indices}{the indices of the measurement_shifts vector that each predicted titre needs adding to it}

\item{measurement_shifts}{the vector of measurement shifts for each cluster to add to the predicted titres}
}
\value{
a vector with the likelihood of making each observation given the predictions
}
\description{
Calculates the likelihood of observing a set of discrete HI titres given a corresponding set of predicted titres
}
