% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{r_likelihood_continuous}
\alias{r_likelihood_continuous}
\title{Likelihood function given continuous data (normal)}
\usage{
r_likelihood_continuous(
  expected,
  data,
  theta,
  expected_indices = NULL,
  measurement_shifts = NULL
)
}
\arguments{
\item{expected}{vector of expected titres}

\item{data}{vector of observed continuous titres}

\item{expected_indices}{the indices of the measurement_shifts vector that each predicted titre needs adding to it}

\item{measurement_shifts}{the vector of measurement shifts for each cluster to add to the predicted titres}
}
\value{
a vector with the likelihood of making each observation given the predictions
}
\description{
Calculates the likelihood of observing a set of continuous and bounded titres given a corresponding set of predicted titres
}
