% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_funcs.R
\name{rbb}
\alias{rbb}
\title{Beta binomial random generation}
\usage{
rbb(n, N, u, v)
}
\arguments{
\item{n}{number of draws}

\item{N}{the number of trials}

\item{u}{equivalent to alpha in description}

\item{v}{equivalent to beta in description}
}
\value{
the beta binomial random draw function
}
\description{
Beta binomial random generation
}
