% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{rm_scale}
\alias{rm_scale}
\title{Robins and Monro scaler, thanks to Michael White}
\usage{
rm_scale(step_scale, mc, target_acceptance_rate_theta, log_prob, N_adapt)
}
\description{
Robins and Monro scaler, thanks to Michael White
}
\seealso{
Other mcmc:
\code{\link[=generate_start_tab]{generate_start_tab()}},
\code{\link[=save_infection_history_to_disk]{save_infection_history_to_disk()}},
\code{\link[=scaletuning]{scaletuning()}},
\code{\link[=scaletuning_alt]{scaletuning_alt()}},
\code{\link[=serosolver]{serosolver()}}
}
\concept{mcmc}
