% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{rm_scale}
\alias{rm_scale}
\title{Robins and Monro scaler, thanks to Michael White}
\usage{
rm_scale(step_scale, mc, popt, log_prob, N_adapt)
}
\description{
Robins and Monro scaler, thanks to Michael White
}
