% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{row.match}
\alias{row.match}
\title{Find matching rows in two data frames}
\usage{
row.match(x, table, nomatch = NA)
}
\arguments{
\item{x}{data frame or vector of rows to match}

\item{table}{data frame or matrix in which to search}

\item{nomatch}{value returned when no match is found}
}
\value{
an integer vector of matching row numbers

This helper is based on the `row.match` function from the prodlim package.
}
\description{
Returns the row number of the first matching row in `table` for each row of `x`.
}
