% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{save_infection_history_to_disk}
\alias{save_infection_history_to_disk}
\title{Write given infection history to disk}
\usage{
save_infection_history_to_disk(infection_history, file, sampno,
  append = TRUE, col_names = FALSE)
}
\arguments{
\item{infection_history}{the infection history matrix}

\item{file}{the file location to save to}

\item{sampno}{which sample number is this matrix?}

\item{append}{if TRUE, just adds to the bottom of the file}

\item{col_names}{if TRUE, saves column names first (only set to true if append = FALSE)}
}
\value{
nothing
}
\description{
Write given infection history to disk
}
