% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{scaletuning}
\alias{scaletuning}
\title{Scale step sizes}
\usage{
scaletuning(step, popt, pcur)
}
\arguments{
\item{step}{the current step size}

\item{popt}{the desired acceptance rate}

\item{pcur}{the current acceptance rate}
}
\value{
the scaled step size
}
\description{
Scales the given step size (between 0 and 1) based on the current acceptance rate to get closed to the desired acceptance rate
}
\seealso{
Other mcmc: 
\code{\link{generate_start_tab}()},
\code{\link{rm_scale}()},
\code{\link{run_MCMC}()},
\code{\link{save_infection_history_to_disk}()}
}
\concept{mcmc}
