% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{setup_infection_histories}
\alias{setup_infection_histories}
\title{Propose initial infection histories}
\usage{
setup_infection_histories(
  titre_dat,
  strain_isolation_times,
  space = 5,
  titre_cutoff = 2,
  sample_prob = 0.9
)
}
\arguments{
\item{titre_dat}{the matrix of titres data with columns for individual, sample, and titre}

\item{strain_isolation_times}{vector of real times for all strains}

\item{space}{how many epochs must separate proposed infections}

\item{titre_cutoff}{specifies how high the titre must be to imply an infection}

\item{sample_prob}{if titre suggests an infection, then add an infection with 1 minus this probability}
}
\value{
an nxm matrix of infection histories containing 1s and 0s, where n is the number of individuals and m is the number of potential infecting strains
}
\description{
Given a matrix of titre data, proposes plausible initial infection histories from which to begin MCMC sampling.
The idea is to move along time in the context of antigenic drift and look at an individual's titre against each strain. Where titres are raised, we suggest an infection. However, to avoid suggesting multiple infections for regions of high antigenic similarity, we place a necessary gap (defined by `space`) between proposed infection times.
}
\examples{
data(example_titre_dat)
data(example_antigenic_map)
start_inf <- setup_infection_histories(example_titre_dat, example_antigenic_map$inf_times)
}
\seealso{
Other setup_infection_histories: 
\code{\link{setup_infection_histories_old}()},
\code{\link{setup_infection_histories_titre}()},
\code{\link{setup_infection_histories_total}()}
}
\concept{setup_infection_histories}
