% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{setup_infection_histories_OLD}
\alias{setup_infection_histories_OLD}
\title{Propose initial infection histories - OLD VERSION}
\usage{
setup_infection_histories_OLD(dat, strainIsolationTimes, ageMask, sample_prob,
  titre_cutoff = 3)
}
\arguments{
\item{dat}{the matrix of titres data with columns for individual, sample, and titre}

\item{strainIsolationTimes}{vector of real times for all strains}

\item{ageMask}{vector of indices for each individual corresponding to the first index of the strainIsolationTimes vector that each individual could be infected with}

\item{sample_prob}{given an infection seems likely based on titre, suggest infection with some probability}

\item{titre_cutoff}{specifies how high the titre must be to imply an infection}
}
\value{
an nxm matrix of infection histories containing 1s and 0s, where n is the number of individuals and m is the number of potential infecting strains
}
\description{
Given a matrix of titre data, proposes plausible initial infection histories from which to begin MCMC sampling.
NOTE - MIGHT NEED TO UPDATE THIS FOR GROUPS
}

