% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{setup_infection_histories_prior}
\alias{setup_infection_histories_prior}
\title{Generate random infection history matrix}
\usage{
setup_infection_histories_prior(
  antibody_data,
  possible_exposure_times,
  infection_model_prior_shape1 = 1,
  infection_model_prior_shape2 = 1
)
}
\arguments{
\item{antibody_data}{the data frame of antibody data. See \code{\link{example_antibody_data}}}

\item{possible_exposure_times}{the vector of times that individuals can be infected}

\item{infection_model_prior_shape1}{shape parameter 1 (alpha) of the Beta distribution}

\item{infection_model_prior_shape2}{shape parameter 2 (beta) of the Beta distribution}
}
\value{
an n (number of individuals) by m (number of possible exposure times) matrix containing 1s and 0s, representing infections
}
\description{
Creates a random infection history matrix by sampling from the infection history model under prior version 2
}
\examples{
\dontrun{
data(example_antibody_data)
data(example_antigenic_map)
times <- example_antigenic_map$inf_times
setup_infection_histories_prior(example_antibody_data, times, 1, 1)
}
}
\seealso{
Other setup_infection_histories:
\code{\link[=setup_infection_histories]{setup_infection_histories()}},
\code{\link[=setup_infection_histories_antibody_level]{setup_infection_histories_antibody_level()}}
}
\concept{setup_infection_histories}
\keyword{internal}
