% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setup_titredat_for_posterior_func}
\alias{setup_titredat_for_posterior_func}
\title{Setup titre data indices}
\usage{
setup_titredat_for_posterior_func(
  titre_dat,
  antigenic_map = NULL,
  strain_isolation_times = NULL,
  age_mask = NULL,
  n_alive = NULL
)
}
\arguments{
\item{titre_dat}{the data frame of data to be fitted. Must have columns: group (index of group); individual (integer ID of individual); samples (numeric time of sample taken); virus (numeric time of when the virus was circulating); obs_type (integer of the observation group type, using a unique value for each distinctive type of observation underpinned by the same generative model); titre (integer of titre value against the given virus at that sampling time). See \code{\link{example_titre_dat}}}

\item{antigenic_map}{(optional) a data frame of antigenic x and y coordinates. Must have column names: x_coord; y_coord; inf_times. See \code{\link{example_antigenic_map}}}

\item{strain_isolation_times}{(optional) if no antigenic map is specified, this argument gives the vector of times at which individuals can be infected}

\item{age_mask}{see \code{\link{create_age_mask}} - a vector with one entry for each individual specifying the first epoch of circulation in which an individual could have been exposed}

\item{n_alive}{if not NULL, uses this as the number alive in a given year rather than calculating from the ages. This is needed if the number of alive individuals is known, but individual birth dates are not}
}
\value{
a very long list. See source code directly.
}
\description{
Sets up a large list of pre-indexing and pre-processing to speed up the model solving during MCMC fitting.
Note that this should be `titre_dat` after subsetting to only `run==1`, as we will figure out elsewhere which solves to use as repeats
}
\seealso{
\code{\link{create_posterior_func}}
}
