% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_attack_rates}
\alias{simulate_attack_rates}
\title{Simulate attack rates}
\usage{
simulate_attack_rates(
  infection_years,
  mean_par = 0.15,
  sd_par = 0.5,
  large_first_year = FALSE,
  big_year_mean = 0.5,
  n_groups = 1
)
}
\arguments{
\item{infection_years}{vector of possible infection times}

\item{mean_par}{target mean attack rate, either one value or one value per population group}

\item{sd_par}{standard deviation on the log scale, either one value or one value per population group}

\item{large_first_year}{logical value indicating whether to simulate an extra-large attack rate in the first time period}

\item{big_year_mean}{target mean for the first time period when `large_first_year = TRUE`}

\item{n_groups}{number of population groups to simulate}
}
\value{
A long-format data frame with `population_group`, `time`, and `prob_infection` columns.
}
\description{
Given a vector of possible infection times, simulates attack rates from a log-normal distribution with specified mean and log-scale standard deviation.
}
\seealso{
Other simulation_functions:
\code{\link[=simulate_data]{simulate_data()}},
\code{\link[=simulate_infection_histories]{simulate_infection_histories()}}
}
\concept{simulation_functions}
