% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_attack_rates}
\alias{simulate_attack_rates}
\title{Simulate attack rates}
\usage{
simulate_attack_rates(infectionYears, meanPar = 0.15, sdPar = 0.5,
  largeFirstYear = FALSE, bigYearMean = 0.5)
}
\arguments{
\item{infectionYears}{the number of infection years}

\item{meanPar}{the mean of the log normal}

\item{sdPar}{the sd of the log normal}

\item{largeFirstYear}{simulate an extra large attach rate in the first year?}

\item{bigYearMean}{if large first year, what mean to use?}
}
\value{
a vector of attack rates
}
\description{
Given a number of possible infection years, simulates attack rates from a log normal distribution with specified mean and standard deviation.
}

