% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate full data set}
\usage{
simulate_data(par_tab, group = 1, n_indiv = 100, buckets = 12,
  strain_isolation_times, measured_strains = NULL, sampling_times,
  nsamps = 2, antigenic_map, titre_sensoring = 0, age_min = 5,
  age_max = 80, attack_rates, repeats = 1, mu_indices = NULL,
  measurement_indices = NULL, add_noise = TRUE)
}
\arguments{
\item{par_tab}{the full parameter table controlling parameter ranges and values}

\item{group}{which group index to give this simulated data}

\item{n_indiv}{number of individuals to simulate}

\item{strain_isolation_times}{vector of strain circulation times}

\item{measured_strains}{vector of strains that have titres measured}

\item{sampling_times}{possible sampling times for the individuals}

\item{nsamps}{the number of samples each individual has}

\item{antigenic_map}{the raw antigenic map with colnames x_coord, y_coord and inf_years}

\item{titre_sensoring}{what proportion of titres are randomly missing?}

\item{age_min}{minimum age to simulate}

\item{age_max}{maximum age to simulate}

\item{attack_rates}{a vector of attack_rates to be used in the simulation}

\item{repeats}{number of repeats for each year}

\item{mu_indices}{default NULL, optional vector giving the index of `mus` that each strain uses the boosting parameter from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{measurement_indices}{default NULL, optional vector giving the index of `measurement_bias` that each strain uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{add_noise}{if TRUE, adds observation noise to the simulated titres}
}
\value{
a list with: 1) the data frame of titre data as returned by \code{\link{simulate_group}}; 2) a matrix of infection histories as returned by \code{\link{simulate_infection_histories}}; 3) a vector of ages
}
\description{
Simulates a full data set for a given set of parameters etc.
}
