% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate full data set}
\usage{
simulate_data(par_tab, group = 1, n_indiv = 100, buckets = 1,
  strain_isolation_times, measured_strains = NULL, sampling_times,
  nsamps = 2, antigenic_map, titre_sensoring = 0, age_min = 5,
  age_max = 80, attack_rates, repeats = 1, mu_indices = NULL,
  measurement_indices = NULL, add_noise = TRUE,
  separate_exposures = FALSE)
}
\arguments{
\item{par_tab}{the full parameter table controlling parameter ranges and values}

\item{group}{which group index to give this simulated data}

\item{n_indiv}{number of individuals to simulate}

\item{buckets}{time resolution of the simulated data. buckets=1 indicates annual time resolution; buckets=4 indicates quarterly; buckets=12 monthly}

\item{strain_isolation_times}{vector of contiguous strain circulation times}

\item{measured_strains}{vector of strains that have titres measured matching entries in strain_isolation_times}

\item{sampling_times}{possible sampling times for the individuals, matching entries in strain_isoltion_times}

\item{nsamps}{the number of samples each individual has (eg. nsamps=2 gives each individual 2 random sampling times from sampling_times)}

\item{antigenic_map}{the raw antigenic map with colnames x_coord, y_coord and inf_years, as returned from \code{\link{generate_antigenic_map_flexible}} or \code{\link{generate_antigenic_map}}}

\item{titre_sensoring}{numeric between 0 and 1, used to censor a proportion of titre observations at random (MAR)}

\item{age_min}{simulated age minimum}

\item{age_max}{simulated age maximum}

\item{attack_rates}{a vector of attack_rates for each entry in strain_isolation_times to be used in the simulation (between 0 and 1)}

\item{repeats}{number of repeat observations for each year}

\item{mu_indices}{default NULL, optional vector giving the index of `mus` that each strain uses the boosting parameter from. eg. if there are 6 circulation years in strain_isolation_times and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{measurement_indices}{default NULL, optional vector giving the index of `measurement_bias` that each strain uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{add_noise}{if TRUE, adds observation noise to the simulated titres}

\item{separate_exposures}{if TRUE, returns the exposure and infection histories as separate matrices}
}
\value{
a list with: 1) the data frame of titre data as returned by \code{\link{simulate_group}}; 2) a matrix of infection histories as returned by \code{\link{simulate_infection_histories}}; 3) a vector of ages
}
\description{
Simulates a full data set for a given set of parameters etc.
}
