% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_group}
\alias{simulate_group}
\title{Simulate group data}
\usage{
simulate_group(n_indiv, theta, infection_histories, strain_isolation_times,
  sample_times, nsamps, antigenic_map_long, antigenic_map_short,
  repeats = 1, mus = NULL, mu_indices = NULL,
  measurement_bias = NULL, measurement_indices = NULL,
  add_noise = TRUE)
}
\arguments{
\item{n_indiv}{the number of individuals to simulate}

\item{theta}{the named parameter vector}

\item{infection_histories}{the matrix of 1s and 0s giving presence/absence of infections for each individual}

\item{strain_isolation_times}{the vector of strain circulation times (ie. possible infection times)}

\item{sample_times}{the vector of times at which samples could be taken}

\item{nsamps}{the number of samples per individual}

\item{antigenic_map_long}{the collapsed antigenic map for long term cross reactivity, after multiplying by sigma1}

\item{antigenic_map_short}{the collapsed antigenic map for short term cross reactivity, after multiplying by sigma2}

\item{repeats}{number of repeated samples for each year}

\item{mus}{default NULL, optional vector of boosting parameters for each strain}

\item{mu_indices}{default NULL, optional vector giving the index of `mus` that each strain uses the boosting parameter from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{measurement_bias}{default NULL, optional vector of measurement bias shift parameters for each strain}

\item{measurement_indices}{default NULL, optional vector giving the index of `measurement_bias` that each strain uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{add_noise}{defaults TRUE, adds observation error to simulated titres}
}
\value{
a data frame with columns individual, samples, virus and titre of simulated data
}
\description{
Simulates a full set of titre data for n_indiv individuals with known theta and infection_histories. Each individual gets nsamps random samples from sampleTimes, and infections can occur at any of strain_isolation_times
}
\seealso{
\code{\link{simulate_individual}}
}
