% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_group}
\alias{simulate_group}
\title{Simulate group data}
\usage{
simulate_group(
  n_indiv,
  theta,
  infection_histories,
  strain_isolation_times,
  measured_strains,
  sample_times,
  nsamps,
  antigenic_map,
  repeats = 1,
  measurement_bias = NULL,
  measurement_indices = NULL,
  add_noise = TRUE,
  DOBs = NULL
)
}
\arguments{
\item{n_indiv}{number of individuals to simulate}

\item{theta}{the named parameter vector}

\item{infection_histories}{the matrix of 1s and 0s giving presence/absence of infections for each individual}

\item{strain_isolation_times}{(optional) If no antigenic map is specified, this argument gives the vector of times at which individuals can be infected}

\item{measured_strains}{vector of strains that have titres measured matching entries in strain_isolation_times}

\item{nsamps}{the number of samples each individual has (eg. nsamps=2 gives each individual 2 random sampling times from sampling_times)}

\item{antigenic_map}{(optional) A data frame of antigenic x and y coordinates. Must have column names: x_coord; y_coord; inf_times. See \code{\link{example_antigenic_map}}}

\item{repeats}{number of repeat observations for each year}

\item{measurement_indices}{default NULL, optional vector giving the index of `measurement_bias` that each strain uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{add_noise}{if TRUE, adds observation noise to the simulated titres}

\item{DOBs}{vector giving the time period of birth (entries corresponding to `strain_isolation_times`)}
}
\value{
a data frame with columns individual, samples, virus and titre of simulated data
}
\description{
Simulates a full set of titre data for n_indiv individuals with known theta and infection_histories. Each individual gets nsamps random samples from sampleTimes, and infections can occur at any of strain_isolation_times
}
\seealso{
\code{\link{simulate_individual}}, \code{\link{simulate_individual_faster}}

Other simulation_functions: 
\code{\link{simulate_attack_rates}()},
\code{\link{simulate_data}()},
\code{\link{simulate_individual_faster}()},
\code{\link{simulate_individual}()},
\code{\link{simulate_infection_histories}()}
}
\concept{simulation_functions}
