% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_group}
\alias{simulate_group}
\title{Simulate group data}
\usage{
simulate_group(
  n_indiv,
  theta,
  theta_indices_unique,
  unique_biomarker_groups,
  infection_histories,
  possible_exposure_times,
  measured_biomarker_ids,
  sample_times,
  nsamps,
  antigenic_map,
  repeats = 1,
  measurement_bias = NULL,
  measurement_indices = NULL,
  data_type = NULL,
  DOBs = NULL
)
}
\arguments{
\item{n_indiv}{number of individuals to simulate}

\item{theta}{the named parameter vector}

\item{theta_indices_unique}{which index in par_tab does each element of `theta` correspond to?}

\item{unique_biomarker_groups}{vector of unique measured biomarker types}

\item{infection_histories}{the matrix of 1s and 0s giving presence/absence of infections for each individual}

\item{possible_exposure_times}{(optional) If no antigenic map is specified, this argument gives the vector of times at which individuals can be infected}

\item{measured_biomarker_ids}{vector of biomarker IDs that have titres measured matching entries in possible_exposure_times}

\item{nsamps}{the number of samples each individual has (eg. nsamps=2 gives each individual 2 random sampling times from sampling_times)}

\item{antigenic_map}{(optional) A data frame of antigenic x and y coordinates. Must have column names: x_coord; y_coord; inf_times. See \code{\link{example_antigenic_map}}.}

\item{repeats}{number of repeat observations for each year}

\item{measurement_indices}{default NULL, optional vector giving the index of `measurement_bias` that each antigen/biomarker ID uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{data_type}{if not NULL, then a vector of data types to use for each biomarker_group}

\item{DOBs}{vector giving the time period of birth (entries corresponding to `possible_exposure_times`)}
}
\value{
a data frame with columns individual, samples, virus and titre of simulated data
}
\description{
Simulates a full set of antibody data for n_indiv individuals with known theta and infection_histories. Each individual gets nsamps random samples from sample_times, and infections can occur at any of possible_exposure_times
}
\seealso{
\code{\link{simulate_individual}}, \code{\link{simulate_individual_faster}}

Other simulation_functions: 
\code{\link{simulate_attack_rates}()},
\code{\link{simulate_data}()},
\code{\link{simulate_individual_faster}()},
\code{\link{simulate_individual}()},
\code{\link{simulate_infection_histories}()}
}
\concept{simulation_functions}
