% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_individual}
\alias{simulate_individual}
\title{Simulate individual data}
\usage{
simulate_individual(theta, infection_history, antigenic_map,
  sampling_times, strain_isolation_times, measured_strains, mus = NULL,
  mu_indices = NULL, measurement_bias = NULL,
  measurement_indices = NULL, add_noise = TRUE, repeats = 1,
  DOB = NULL)
}
\arguments{
\item{theta}{the named parameter vector}

\item{infection_history}{the vector of 1s and 0s giving presence/absence of infections}

\item{antigenic_map}{the raw antigenic map with colnames x_coord, y_coord and inf_years, as returned from \code{\link{generate_antigenic_map_flexible}} or \code{\link{generate_antigenic_map}}}

\item{sampling_times}{vector of times at which blood samples were taken}

\item{strain_isolation_times}{vector of contiguous strain circulation times}

\item{measured_strains}{vector of which strains had titres measured in `strain_isolation_times`}

\item{mus}{default NULL, optional vector of boosting parameters for each strain}

\item{mu_indices}{default NULL, optional vector giving the index of `mus` that each strain uses the boosting parameter from. eg. if there are 6 circulation years in strain_isolation_times and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{measurement_indices}{default NULL, optional vector giving the index of `measurement_bias` that each strain uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{add_noise}{if TRUE, adds observation noise to the simulated titres}

\item{repeats}{number of repeat observations for each year}

\item{DOB}{date of birth of the individual}
}
\value{
a data frame with columns samples, virus and titre of simulated data
}
\description{
Simulates a full set of titre data for an individual with known theta and infection_history.
}
\examples{
data(example_par_tab)
data(example_antigenic_map)
infection_history <- sample(c(0,1),size=nrow(example_antigenic_map), replace=TRUE,prob=c(0.9,0.1))
pars <- example_par_tab$values
names(pars) <- example_par_tab$names
strain_isolation_times <- example_antigenic_map$inf_years
y <- simulate_individual(pars, infection_history, example_antigenic_map, 2009, 
                         strain_isolation_times,strain_isolation_times,add_noise=FALSE)
}
\seealso{
Other simulation_functions: \code{\link{simulate_attack_rates}},
  \code{\link{simulate_data}},
  \code{\link{simulate_group}},
  \code{\link{simulate_individual_faster}},
  \code{\link{simulate_infection_histories}}
}
\concept{simulation_functions}
