% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_individual}
\alias{simulate_individual}
\title{Simulate individual data}
\usage{
simulate_individual(theta, infectionHistory, samplingTimes, dataIndices,
  strainIsolationTimes, virusIndices, antigenicMapLong, antigenicMapShort,
  strains, repeats = 1, mus = NULL, musIndices = NULL)
}
\arguments{
\item{theta}{the named parameter vector}

\item{infectionHistory}{the vector of 1s and 0s giving presence/absence of infections}

\item{dataIndices}{see \code{\link{create_post_func}}}

\item{strainIsolationTimes}{see \code{\link{create_post_func}}}

\item{virusIndices}{see \code{\link{create_post_func}}}

\item{antigenicMapLong}{the collapsed antigenic map for long term cross reactivity, after multiplying by sigma1}

\item{antigenicMapShort}{the collapsed antigenic map for short term cross reactivity, after multiplying by sigma2}

\item{strains}{vector of all possible circulating strains}

\item{repeats}{number of repeated samples for each year}

\item{sampleTimes}{the vector of times at which samples are taken}

\item{strainIsolationTimes}{the vector of strain circulation times (ie. possible infection times)}
}
\value{
a data frame with columns samples, virus and titre of simulated data
}
\description{
Simulates a full set of titre data for an individual with known theta and infectionHistory.
}
\seealso{
\code{\link{infection_model_indiv}}
}

