% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_individual_faster}
\alias{simulate_individual_faster}
\title{Simulate individual data quickly}
\usage{
simulate_individual_faster(theta, infection_history, antigenic_map_long,
  antigenic_map_short, antigenic_distances, sampling_times,
  strain_isolation_times, measured_strains, mus = NULL,
  mu_indices = NULL, measurement_bias = NULL,
  measurement_indices = NULL, add_noise = TRUE, repeats = 1,
  DOB = NULL)
}
\arguments{
\item{theta}{the named parameter vector}

\item{infection_history}{the vector of 1s and 0s giving presence/absence of infections}

\item{antigenic_map_long}{the long term antigenic cross reactivity map generated from \code{\link{create_cross_reactivity_vector}}}

\item{antigenic_map_short}{the short term antigenic cross reactivity map generated from \code{\link{create_cross_reactivity_vector}}}

\item{antigenic_distances}{same dimensions as antigenic_map_long and antigenic_map_short, but gives the raw euclidean antigenic distances}

\item{sampling_times}{vector of times at which blood samples were taken}

\item{strain_isolation_times}{vector of contiguous strain circulation times}

\item{measured_strains}{vector of which strains had titres measured in `strain_isolation_times`}

\item{mus}{default NULL, optional vector of boosting parameters for each strain}

\item{mu_indices}{default NULL, optional vector giving the index of `mus` that each strain uses the boosting parameter from. eg. if there are 6 circulation years in strain_isolation_times and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{measurement_indices}{default NULL, optional vector giving the index of `measurement_bias` that each strain uses the measurement shift from from. eg. if there's 6 circulation years and 3 strain clusters, then this might be c(1,1,2,2,3,3)}

\item{add_noise}{if TRUE, adds observation noise to the simulated titres}

\item{repeats}{number of repeat observations for each year}

\item{DOB}{date of birth of the individual}
}
\value{
a data frame with columns samples, virus and titre of simulated data
}
\description{
FOR USERS: USE \code{\link{simulate_individual}}. This function does the same thing, but with a few short cuts for speed. Simulates a full set of titre data for an individual with known theta and infection_history.
}
\seealso{
Other simulation_functions: \code{\link{simulate_attack_rates}},
  \code{\link{simulate_data}},
  \code{\link{simulate_group}},
  \code{\link{simulate_individual}},
  \code{\link{simulate_infection_histories}}
}
\concept{simulation_functions}
