% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_infection_histories}
\alias{simulate_infection_histories}
\title{Simulate infection histories}
\usage{
simulate_infection_histories(
  p_inf,
  possible_exposure_times = 1:ncol(p_inf),
  demographics
)
}
\arguments{
\item{p_inf}{a vector or data frame of attack rates (infection probabilities) for each possible exposure time and population group}

\item{possible_exposure_times}{the vector of possible infection times}

\item{demographics}{data frame giving `individual`, `birth`, and `population_group` for each individual. `last_sample` and time-varying demographic values can also be supplied.}
}
\value{
A list containing an infection-history matrix and a data frame of empirical attack rates for each population group and exposure time.
}
\description{
Given infection probabilities and potential infection times, simulates infections for each individual in `demographics`. Infections are only added while an individual is alive and within their observation period.
}
\examples{
possible_exposure_times <- seq_len(25)
p_inf <- simulate_attack_rates(possible_exposure_times,n_groups=2)
n_indiv <- 100
demographics <- data.frame(individual = 1:n_indiv,birth=sample(1:5,n_indiv,replace=TRUE),population_group=c(rep(1,n_indiv/2),rep(2,n_indiv/2)))
inf_hist <- simulate_infection_histories(p_inf, possible_exposure_times, demographics)
}
\seealso{
Other simulation_functions:
\code{\link[=simulate_attack_rates]{simulate_attack_rates()}},
\code{\link[=simulate_data]{simulate_data()}}
}
\concept{simulation_functions}
