% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_infection_histories}
\alias{simulate_infection_histories}
\title{Simulate infection histories}
\usage{
simulate_infection_histories(pInf, infSD, strainIsolationTimes,
  samplingTimes, ages)
}
\arguments{
\item{pInf}{a vector of attack rates (infection probabilities) for each year}

\item{infSD}{the standard deviation of attack rate draws}

\item{strainIsolationTimes}{the vector of possible infection times}

\item{samplingTimes}{vector of potential sampling times}

\item{ages}{a vector of ages for each individual}
}
\value{
a list with a matrix of infection histories for each individual in ages and the true attack rate for each epoch
}
\description{
Given a vector of infection probabilities and potential infection times, simulates infections for each element of ages (ie. each element is an individual age. Only adds infections for alive individuals)
}
