% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{solve_model_individual}
\alias{solve_model_individual}
\title{Solves the re-implementation of the model for a single individual}
\usage{
solve_model_individual(parTab, infectionHistory, sampleTime, antigenicMap,
  mu_indices = NULL)
}
\description{
Solves the re-implementation of the model for a single individual
}
