% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subset_nullable_vector}
\alias{subset_nullable_vector}
\title{Takes a subset of a Nullable NumericVector, but only if it isn't NULL}
\usage{
subset_nullable_vector(x, index1, index2)
}
\description{
Takes a subset of a Nullable NumericVector, but only if it isn't NULL
}
