% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_infections_by_group}
\alias{sum_infections_by_group}
\title{Count infections by group and time}
\usage{
sum_infections_by_group(inf_hist, group_ids_vec, n_groups)
}
\description{
Count infections by group and time
}
