% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titre_data_example}
\alias{titre_data_example}
\title{Example titre data}
\format{A data frame with the following variables:

  \describe{

  \item{individual}{Subject index}

  \item{sample_epoch}{Sample epoch index}

  \item{virus_epoch}{Virus circulation epoch index}

  \item{epochs_per_year}{Number of epochs per year}

  \item{logtitre}{Log titre}

  \item{run}{Measurement index}

  \item{birth_epoch}{Epoch of birth index}

  \item{group}{Group index}

  }}
\usage{
titre_data_example
}
\description{
\code{titre_dat} from the quick start vignette
}
\keyword{datasets}
