% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{titre_data_fast}
\alias{titre_data_fast}
\title{Overall model function, fast implementation}
\usage{
titre_data_fast(theta, infection_history_mat, circulation_times,
  circulation_times_indices, sample_times, rows_per_indiv_in_samples,
  cum_nrows_per_individual_in_data, nrows_per_blood_sample,
  measurement_strain_indices, antigenic_map_long, antigenic_map_short,
  additional_arguments)
}
\value{
NumericVector of predicted titres for each entry in measurement_strain_indices
}
\description{
See documentation for \code{\link{titre_data_group}}, as the interface is almost identical
}
\seealso{
Other titre_model: \code{\link{infection_model_indiv}},
  \code{\link{titre_data_group}},
  \code{\link{titre_data_individual}}
}
\concept{titre_model}
