% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{titre_data_individual}
\alias{titre_data_individual}
\title{Model function individual}
\usage{
titre_data_individual(theta, infectionHistory, circulationTimes,
  circulationMapIndices, samplingTimes, dataIndices, measuredMapIndices,
  antigenicMapLong, antigenicMapShort, numberStrains, DOB = 0,
  additional_arguments = NULL)
}
\arguments{
\item{theta}{NumericVector, the named vector of model parameters}

\item{infectionHistory}{IntegerVector, the vector of 1s and 0s showing presence/absence of infection for each possible time.}

\item{circulationTimes}{NumericVector, the actual times of circulation that the infection history vector corresponds to}

\item{dataIndices}{IntegerVector, the indices in the overall titre data vector (of observations) that each sample corresponds to}

\item{measuredMapIndices}{IntegerVector, the indices of all measured strains in the melted antigenic map}

\item{antigenicMapLong}{NumericVector, the collapsed cross reactivity map for long term boosting, after multiplying by sigma1}

\item{antigenicMapShort}{NumericVector, the collapsed cross reactivity map for short term boosting, after multiplying by sigma2}

\item{numberStrains}{int, the maximum number of infections that an individual could experience}

\item{DOB}{double, the date of birth of this individual. Currently not used.}

\item{circualtionaMapIndices}{IntegerVector, which entry in the melted antigenic map that these infection times correspond to}

\item{samplingTime}{NumericVector, the times that each blood sample was taken}
}
\value{
NumericVector of predicted titres for each entry in measuredMapIndices
}
\description{
The main model solving function for a single individual for a vector of sampling times
}
