% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to.pdf}
\alias{to.pdf}
\title{PDF - Rich's function to print to device without potential for bad errors}
\usage{
to.pdf(expr, filename, ..., verbose = TRUE)
}
\arguments{
\item{expr}{expression to give plot}

\item{filename}{filename to print to}
}
\description{
Prints to pdf, but turns dev off if fails
}
\seealso{
Other safe_plot_saving: 
\code{\link{to.png}()},
\code{\link{to.svg}()}
}
\concept{safe_plot_saving}
