% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to.png}
\alias{to.png}
\title{PNG - Rich's function to print to device without potential for bad errors}
\usage{
to.png(expr, filename, ..., verbose = TRUE)
}
\arguments{
\item{expr}{expression to give plot}

\item{filename}{filename to print to}

\item{...}{additional arguments passed to `png()`}

\item{verbose}{if TRUE, prints the output filename}
}
\description{
Prints to png, but turns dev off if fails
}
\seealso{
Other safe_plot_saving:
\code{\link[=to.pdf]{to.pdf()}},
\code{\link[=to.svg]{to.svg()}}
}
\concept{safe_plot_saving}
