% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to.svg}
\alias{to.svg}
\title{SVG - Rich's function to print to device without potential for bad errors}
\usage{
to.svg(expr, filename, ..., verbose = TRUE)
}
\arguments{
\item{expr}{expression to give plot}

\item{filename}{filename to print to}

\item{...}{additional arguments passed to `svg()`}

\item{verbose}{if TRUE, prints the output filename}
}
\description{
Prints to SVG, but turns dev off if fails
}
\seealso{
Other safe_plot_saving:
\code{\link[=to.pdf]{to.pdf()}},
\code{\link[=to.png]{to.png()}}
}
\concept{safe_plot_saving}
\keyword{internal}
