% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_help.R
\name{univ_proposal}
\alias{univ_proposal}
\title{MCMC proposal function}
\usage{
univ_proposal(values, lower_bounds, upper_bounds, steps, index)
}
\arguments{
\item{values}{a vector of the parameters to be explored}

\item{lower_bounds}{a vector of the low allowable bounds for the proposal}

\item{upper_bounds}{a vector of the upper allowable bounds for the proposal}

\item{steps}{a vector of step sizes for the proposal}

\item{index}{numeric value for the index of the parameter to be moved from the param table and vector}
}
\value{
the parameter vector after step
}
\description{
Proposal function for MCMC random walk, taking random steps of a given size.
}
